% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-transforms.R
\name{AbsCat-type-tests}
\alias{AbsCat-type-tests}
\alias{is.abscat.subtotal}
\alias{is.abscat.heading}
\alias{is.abscat.category}
\title{Test if an abstract category object is of a conceptual type}
\usage{
is.abscat.subtotal(x)

is.abscat.heading(x)

is.abscat.category(x)
}
\arguments{
\item{x}{an AbsCat object}
}
\value{
logical if the AbsCat object is
}
\description{
Convenience functions to test if an abstract category (AbsCat) object is a
specific type. These types are defined by the properties of the abstract
category object.
}
\details{
\code{is.abscat.subtotal} is \code{x} a subtotal insertion?
\code{is.abscat.heading} is \code{x} a heading insertion?
\code{is.abscat.category} is \code{x} a category?
}
\keyword{internal}
