% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AbsCats-class}
\alias{AbsCats-class}
\alias{abscat}
\alias{AbsCats}
\alias{AbsCat-class}
\alias{AbsCat}
\title{Abstract categories}
\usage{
AbsCats(..., data = NULL)

AbsCat(..., data = NULL)
}
\arguments{
\item{...}{additional arguments to \code{[}, ignored}

\item{data}{For the constructor functions \code{AbsCat} and
\code{AbsCats}, you can either pass in attributes via \code{...} or you
can create the objects with a fully defined \code{list} representation of
the objects via the \code{data} argument. See the examples.}

\item{x}{For the attribute getters and setters, an object of class
AbsCat or AbsCats}

\item{i}{For the \code{[} methods, just as with list extract methods}

\item{j}{Invalid argument to \code{[}, but in the generic's signature}

\item{drop}{Invalid argument to \code{[}, but in the generic's signature}

\item{value}{For \code{[<-}, the replacement AbsCat to insert}
}
\description{
An abstract class that categories, elements, insertions, etc.
}
\keyword{internal}
