% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_to_crunch_useragent}
\alias{add_to_crunch_useragent}
\title{Add a string to the crunch user-agent}
\usage{
add_to_crunch_useragent(string)
}
\arguments{
\item{string}{the string to add to the end of the user-agent (typically a
package name and version number)}
}
\description{
Adds a string onto the user-agent that crunch sends to the server. Useful for
tracking requests from packages that depend on \code{\link{crunch}}
}
\keyword{internal}
