% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{addGeoMetadata}
\alias{addGeoMetadata}
\title{Add geodata metadata to a crunch variable}
\usage{
addGeoMetadata(variable)
}
\arguments{
\item{variable}{a Crunch variable to use for matching. This must be either
a text or a categorical variable.}
}
\value{
a CrunchGeography object that can be assigned into \code{geo(variable)}
}
\description{
Add geodata metadata to a crunch variable
}
\examples{
\dontrun{
geo(ds$state) <- addGeoMetadata("state", data=ds)
}
}
