% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add-subvariable.R
\name{addSubvariable}
\alias{addSubvariable}
\title{Add subvariable to an array}
\usage{
addSubvariable(variable, subvariable, ds = NULL)
}
\arguments{
\item{variable}{the array variable to modify}

\item{subvariable}{the subvariable to add}
}
\value{
a new version of \code{variable} with the indicated subvariables
}
\description{
This function conceals the dirty work in making this happen. The array
gets unbound and rebound into a new array with the new variable added.
}

