% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-variable.R
\name{addVariables}
\alias{addVariables}
\title{Add multiple variables to a dataset}
\usage{
addVariables(dataset, ...)
}
\arguments{
\item{dataset}{a CrunchDataset}

\item{...}{\code{\link{VariableDefinition}}s or a list of
VariableDefinitions.}
}
\value{
\code{dataset} with the new variables added (invisibly)
}
\description{
This function lets you add more than one variable at a time to a dataset.
If you have multiple variables to add, this function will be faster than
doing \code{ds$var <- value} assignment because it doesn't refresh the
dataset's state in between variable POST requests.
}
