% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-dataset.R
\name{appendDataset}
\alias{appendDataset}
\title{Append one Crunch dataset to another}
\usage{
appendDataset(dataset1, dataset2, cleanup = TRUE, autorollback = TRUE)
}
\arguments{
\item{dataset1}{a CrunchDataset}

\item{dataset2}{another CrunchDataset, or possibly a data.frame. If
\code{dataset2} is not a Crunch dataset, it will be uploaded as a new
dataset before appending.}

\item{cleanup}{Deprecated. See \code{\link{cleanseBatches}}.}

\item{autorollback}{Rollback to state before append in case of failure.}
}
\value{
A CrunchDataset with \code{dataset2} appended to \code{dataset1}
}
\description{
Append one Crunch dataset to another
}

