% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-stream.R
\name{appendStream}
\alias{appendStream}
\title{Manually trigger a pending append to a dataset}
\usage{
appendStream(ds)
}
\arguments{
\item{ds}{a CrunchDataset}
}
\value{
the dataset with pending stream data appended.
}
\description{
Crunch allows you to stream data to a dataset. Streaming data is useful for
datasets which have frequent updates (see the
\href{http://docs.crunch.io/#streaming-rows}{Crunch API documentation} for more
information). Crunch automatically appends streamed data periodically;
however, if you would like to trigger appending pending streamed data to a
dataset, you can call \code{appendStream()}.
}
