% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-stream.R
\name{appendStreamedRows}
\alias{appendStreamedRows}
\title{Append rows that have been streamed to a Crunch dataset}
\usage{
appendStreamedRows(ds, ...)
}
\arguments{
\item{ds}{a CrunchDataset}

\item{...}{Additional multitable attributes to set. Options include
\code{name} and \code{description}.}
}
\value{
nothing?
}
\description{
Crunch allows you to stream data to a dataset. Streaming data is useful for
datasets which have frequent updates (see
\href{http://docs.crunch.io/#streaming-rows}{the Crunch documentation} for more)
information.
}
\examples{
\dontrun{
# need examples!
}
}
