% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-transforms.R
\name{applyTransforms}
\alias{applyTransforms}
\title{From a cube, calculate the transforms and return an array}
\usage{
applyTransforms(x, ary, ...)
}
\arguments{
\item{x}{a CrunchCube}

\item{ary}{an array to use, if not using the default array from the cube
itself. (Default: not used, pulls an array from the cube directly)}

\item{...}{arguments to pass to \code{cubeToArray}}
}
\value{
an array with any transformations applied
}
\description{
\code{applyTransforms} calculates transforms (e.g. subtotals) on a CrunchCube.
Currently only the row transforms are supported. This is useful if you want
to use the values from the subtotals of the CrunchCube in an analysis.
}
\examples{
\dontrun{
transformed_array <- applyTransforms(crtabs(~opinion, ds))
}

}
