% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_crunch_array_variable.R
\name{array_apply}
\alias{array_apply}
\title{Apply function to each column of a packed data.frame array column}
\usage{
array_apply(x, func, ...)
}
\arguments{
\item{x}{A data.frame}

\item{func}{A function to apply to each column}

\item{...}{Arguments passed to \code{func}}
}
\description{
Apply function to each column of a packed data.frame array column
}
