% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{as.environment,CrunchDataset-method}
\alias{as.environment,CrunchDataset-method}
\title{as.environment method for CrunchDataset}
\usage{
\S4method{as.environment}{CrunchDataset}(x)
}
\arguments{
\item{x}{CrunchDataset}
}
\value{
an environment in which named objects are (promises that return)
CrunchVariables.
}
\description{
This method allows you to \code{eval} within a Dataset.
}
