% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.vector,CrunchVariable-method}
\alias{as.vector,CrunchVariable-method}
\title{Convert Variables to local R objects}
\usage{
\S4method{as.vector}{CrunchVariable}(x, mode = "any")
}
\arguments{
\item{x}{a CrunchVariable subclass}

\item{mode}{argument not used: part of the generic \code{as.vector}
signature}
}
\value{
an R vector of the type corresponding to the Variable. E.g.
CategoricalVariable yields type factor, NumericVariable yields numeric, etc.
}
\description{
Convert Variables to local R objects
}

