% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{axes}
\alias{axes}
\title{Get Crunch Array axes}
\usage{
axes(x, ...)
}
\arguments{
\item{x}{A crunch array variable}

\item{...}{Saved for future expansion}
}
\description{
Provide a summary of the structure of items inside a Crunch Array Variable.
Array variables ared stored in R as "packed" data.frames (eg data.frame columns
in another data.frame).
}
