% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{batches}
\alias{batches}
\title{See the appended batches of this dataset}
\usage{
batches(x)
}
\arguments{
\item{x}{a \code{CrunchDataset}}
}
\value{
a \code{BatchCatalog}
}
\description{
See the appended batches of this dataset
}

