% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bindVariables}
\alias{bindVariables}
\title{Take variables and their dataset and bind them into a new array variable}
\usage{
bindVariables(var_urls, dataset, name, ...)
}
\description{
Take variables and their dataset and bind them into a new array variable
}

