% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-array.R
\name{buildDelimRegex}
\alias{buildDelimRegex}
\title{Build Regex to find  delimited items.}
\usage{
buildDelimRegex(str, delim)
}
\arguments{
\item{str}{A string whose presence indicates a selection}

\item{delim}{The delimiter separating the responses}
}
\value{
A character string
}
\description{
A delimited item \code{maple} can appear in a list in four ways
\enumerate{
\item At the start of a list \code{maple; oak}
\item In the middle of a list \code{oak; maple; birch}
\item At the end of a list \code{oak; maple}
\item Alone with no delimiters \code{maple}
}
}
\details{
This function builds a regex expression which captures those four cases It
is mostly broken out of \code{\link[=createSubvarDeriv]{createSubvarDeriv()}} for testing purposes.
}
\keyword{internal}
