% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/categories.R
\name{c-categories}
\alias{c-categories}
\alias{c.Categories}
\alias{c.Category}
\title{S3 method to concatenate Categories and Category objects}
\usage{
\method{c}{Categories}(...)

\method{c}{Category}(...)
}
\arguments{
\item{...}{see \code{\link[base]{c}}}
}
\value{
An object of class \code{\link{Categories}}
}
\description{
S3 method to concatenate Categories and Category objects
}

