% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-calculate.R
\name{calcInsertions}
\alias{calcInsertions}
\title{Given a vector of values and elements, calculate the insertions}
\usage{
calcInsertions(vec, insert_funcs, dim_names = names(insert_funcs))
}
\arguments{
\item{vec}{values to transform (a single dimension of an array)}

\item{dim_names}{the names of the dimensions (although this is calculable at
call-time, it's much more efficient to provide this to the call)}

\item{trans_funcs}{a (named) list of functions that return the correct vector
of the array (with desired insertions and transformations included)}
}
\value{
the values given in \code{vec}, with any insertions specified in
\code{trans} calculated and inserted
}
\description{
Given a vector of values and elements, calculate the insertions
}
\keyword{internal}
