% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-transforms.R
\name{calcTransform}
\alias{calcTransform}
\title{Given an array and transforms, calculate the transformations.}
\usage{
calcTransform(ary, trans, var_cats)
}
\arguments{
\item{ary}{an array with all dimensions}

\item{trans}{a \code{Transforms} object to pull transformations from}

\item{var_cats}{the \code{Categories} object of the transform}
}
\value{
the array given in \code{ary}, with any transformations specified in
\code{trans} calculated
}
\description{
Given an array and transforms, calculate the transformations.
}
