% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-calculate.R
\name{calcTransforms}
\alias{calcTransforms}
\alias{calcInsertions}
\title{Given a vector of values and elements, calculate the insertions}
\usage{
calcTransforms(array, insert_funcs, dim_names = names(insert_funcs))

calcInsertions(vec, insert_funcs, dim_names = names(insert_funcs))
}
\arguments{
\item{insert_funcs}{a (named) list of functions that return the correct vector
of the array (with desired insertions and transformations included)}

\item{dim_names}{the names of the dimensions (although this is calculable at
call-time, it's much more efficient to provide this to the call)}

\item{vec, array}{values to transform (a single dimension of an array)}
}
\value{
the values given in \code{vec}, with any insertions specified in
\code{trans} calculated and inserted
}
\description{
Given a vector of values and elements, calculate the insertions
}
\keyword{internal}
