% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/case-variables.R
\docType{class}
\name{Case-class}
\alias{Case-class}
\alias{Case}
\alias{Case}
\alias{makeCaseVariable}
\title{Case variables}
\usage{
makeCaseVariable(cases, name, ...)
}
\arguments{
\item{cases}{A list of objects of class \code{\link{Case}}}

\item{name}{a name for the new case variable}

\item{...}{other properties to pass about the case variable}
}
\description{
Case variables
}
