% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-catalog.R, R/decks.R
\docType{methods}
\name{catalog-length}
\alias{catalog-length}
\alias{length,ShojiCatalog-method}
\alias{length,CrunchDeck-method}
\title{Length of Catalog}
\usage{
\S4method{length}{ShojiCatalog}(x)

\S4method{length}{CrunchDeck}(x)
}
\arguments{
\item{x}{a Catalog}
}
\value{
Integer: the number of elements in the index list
}
\description{
Length of Catalog
}
