% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-to-variable.R
\name{categoriesFromLevels}
\alias{categoriesFromLevels}
\title{Convert a factor's levels into Crunch categories.}
\usage{
categoriesFromLevels(level_vect, numeric_values = NULL,
  missing_levels = NULL)
}
\arguments{
\item{level_vect}{A character vector containing the levels of a factor. Usually
obtained by running \code{\link[base:levels]{base::levels()}}}

\item{numeric_values}{A numeric vector containing the numeric values to be
used. (Default: \code{NULL}, which will use the ids as default numeric values)}

\item{missing_levels}{A character vector containing any of the levels that
should be marked missing (Default: \code{NULL}, which will mark none missing)}
}
\value{
A list with each category levels id, name, numeric_value, and missingness.
}
\description{
Crunch categorical variables have slightly richer metadata than R's
factor variables. This function generates a list of category data from
a factor's levels which can then be further manipulated in R before being
imported into Crunch.
}
\examples{

categoriesFromLevels(levels(iris$Species))

}
