% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\docType{methods}
\name{category-extract}
\alias{category-extract}
\alias{$,Category-method}
\alias{$<-,Category-method}
\title{Access Category fields directly}
\usage{
\S4method{$}{Category}(x, name)

\S4method{$}{Category}(x, name) <- value
}
\arguments{
\item{x}{a Category}

\item{name}{a field within \code{x}}

\item{value}{a value for that field to update}
}
\value{
\code{$} returns the value of the desired field. Setter
returns \code{x} duly modified.
}
\description{
Don't do this. Instead, use the category setters.
}
\seealso{
\code{\link{describe-category}}
}
