% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{changeCategoryID}
\alias{changeCategoryID}
\title{Change the id of a category for a categorical variable}
\usage{
changeCategoryID(variable, from, to)
}
\arguments{
\item{variable}{the variable in a crunch dataset that will be changed}

\item{from}{the (old) id identifying the category you want to change}

\item{to}{the (new) id for the category}
}
\value{
\code{variable} with an the new category id
}
\description{
Change the id of a category for a categorical variable
}
\examples{
\dontrun{
ds$country <- changeCategoryID(ds$country, 2, 6)
ds$petloc <- changeCategoryID(ds$petloc, 2, 6)
}
}
