% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-result.R
\name{check_margins}
\alias{check_margins}
\title{Check validity of margins}
\usage{
check_margins(margin, selecteds)
}
\arguments{
\item{margin}{the margin(s) being specified}

\item{selecteds}{which dimensions are selected}
}
\value{
None
}
\description{
A helper function to check if margins supplied are compatible with the
dimensions
}
\keyword{internal}
