% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{checkForNewVersion}
\alias{checkForNewVersion}
\title{See if there's a new version of the package on GitHub}
\usage{

  checkForNewVersion(github.url = "https://api.github.com/repos/Crunch-io/rcrunch/tags",
  installed.version = as.character(packageVersion("crunch")))
}
\arguments{
\item{github.url}{character where to GET the tagged versions of the package}

\item{installed.version}{character the currently installed version string}
}
\value{
The version string if there is a new version, or NULL
}
\description{
See if there's a new version of the package on GitHub
}
\keyword{internal}
