% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{checkInstalledPackages}
\alias{checkInstalledPackages}
\title{Check if a user has packages installed}
\usage{
checkInstalledPackages(pkgs)
}
\arguments{
\item{pkgs}{a character vector of package names to check.}
}
\value{
nothing, called for side effects
}
\description{
Check if a user has packages installed
}
