% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-array.R
\name{cleanImportedArray}
\alias{cleanImportedArray}
\title{Clean up an array variable imported from a lossy file format}
\usage{
cleanImportedArray(variable, min.prefix.length = 20)
}
\arguments{
\item{variable}{An array Variable}

\item{min.prefix.length}{Integer: how many characters long does the common
string need to be in order to consider it significant enough to use? Default
is 20.}
}
\value{
\code{variable} with edits pushed to the API. A common prefix on
subvariable names is extracted and set as the variable's description.
}
\description{
Array and multiple-response variables coming in from SPSS or other file
formats generally need some work to reconstruct the "right" metadata because
they have to shove both parent and subvariable metadata into the "varlabels"
of the subvariables. This often follows a pattern of having varlabels with a
prefix containing the parent question wording (description) and a suffix that
is the actual response label.
}
\details{
This function detects this prefix and reconstructs what may have been the
original array definition.
}
