% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graft.R
\name{cloneDataset}
\alias{cloneDataset}
\title{Clone a dataset}
\usage{
cloneDataset(dataset, useAlias = default.useAlias(), ...)
}
\arguments{
\item{dataset}{CrunchDataset to clone}

\item{useAlias}{logical whether variable alias or name should be used as R
variable names when the dataset is returned. Default is TRUE, meaning alias.}

\item{...}{additional arguments, currently ignored}
}
\value{
a CrunchDataset cloned from \code{dataset}
}
\description{
This function makes a copy of a Crunch Dataset and its associated variables
and variable ordering. It does not copy filters, saved analyses, sharing
privileges, etc.
}
\keyword{internal}

