% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-collapse-dimensions.R
\name{collapse_dims}
\alias{collapse_dims}
\title{Collapse an array from a CrunchCube by summing}
\usage{
collapse_dims(array_in, margins_to_collapse, collapsed_items)
}
\arguments{
\item{margins_to_collapse}{the margins that should be collapsed}

\item{collapsed_items}{vector of logicals if any of the dimensions in
\code{margins_to_collapse} are of the \verb{_items} type}

\item{array}{the array from a CrunchCube to collapse}

\item{margins_to_keep}{the margins that should be kept}
}
\value{
a duly-collapsed array
}
\description{
This is an internal function that powers \code{dimSums()}
}
\keyword{internal}
