% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/automation.R
\name{collapseScripts}
\alias{collapseScripts}
\alias{collapseScripts,CrunchDataset-method}
\alias{collapseScripts,ScriptCatalog-method}
\title{Collapse Crunch Auatomation scripts into a single script}
\usage{
collapseScripts(x, ...)

\S4method{collapseScripts}{CrunchDataset}(x, ...)

\S4method{collapseScripts}{ScriptCatalog}(x, ...)
}
\arguments{
\item{x}{A Crunch object with automation scripts}

\item{...}{Ignored, for possible future expansion}
}
\value{
Invisibly, refreshed x
}
\description{
Collapse multiple crunch automation scripts into a single one, which
may cause improved performance in some situations.
}
\seealso{
Other crunch-automation: 
\code{\link{automation-undo}},
\code{\link{runCrunchAutomation}()},
\code{\link{scripts}()}
}
\concept{crunch-automation}
