% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine-arrays.R
\name{combineArrays}
\alias{combineArrays}
\title{Combine array variables into a new array}
\usage{
combineArrays(var1, var2, ...)
}
\arguments{
\item{variable}{1 to combine}

\item{variable}{2 to combine}

\item{other}{desired attributes such as new variable alias, name,
description as well as any other variables you would like to combine}
}
\value{
a new variable with combine values from the arrays
}
\description{
This function lets you combine array variables into a new array
Arguments include at least two array variables as well as any metadata
for the new variable that is being created. This function uses the data from
the arrays in the order that they are given. If the first variable is NA for a given
row of a subvariable, it will use the data from the second variable and so on.
}

