% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine-mrs.R
\name{combineMRs}
\alias{combineMRs}
\title{Combine multiple_response variables into a new multiple_response}
\usage{
combineMRs(var1, var2, ..., selectedFirst = TRUE)
}
\arguments{
\item{variable}{1 to combine}

\item{variable}{2 to combine}

\item{other}{desired attributes such as new variable alias, name,
description as well as any other variables you would like to combine}

\item{a}{boolean describing whether subvariables that are NA in the
original variables should be made into subvariables in the new
variable}

\item{a}{boolean describing whether 'selected' should be given priority
if false, 'not selected' will be given priority}
}
\value{
a new variable with combined responses from the given variables
}
\description{
This function lets you combine multiple_response variables into a new multiple_response
Arguments include at least two multiple_response variables as well as any metadata
for the new variable that is being created. The argument useNAsubs
is given the defult value FALSE. If it is made TRUE, any subvariables that
are NA in the categorical variables will become subvariables.
This function uses the data from the arrays in the order that they are
given. If the first variable is NA for a given row of a subvariable,
it will use the data from the second variable and so on.
}

