% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-manager.R
\name{consent}
\alias{consent}
\title{Give consent to do things that require permission}
\usage{
consent()
}
\value{
an S3 class "contextManager" object
}
\description{
Give consent to do things that require permission
}
\seealso{
\link{with-context-manager} \link{ContextManager}
}

