% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{copyFolders}
\alias{copyFolders}
\title{Copy the folder structure from one dataset to another.}
\usage{
copyFolders(source, target)
}
\arguments{
\item{source}{the dataset you want to copy the order from}

\item{target}{the dataset you want to copy the order to}
}
\value{
returns the target dataset with source's folder structure
}
\description{
Copy the folder structure from one dataset to another.
}
\examples{
\dontrun{
ds <- copyFolders(ds1, ds)
}
}
