% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{copyOrderGroup}
\alias{copyOrderGroup}
\title{Copy the order of a \code{VariableGroup} (or individual variable URL) from \code{VariableOrder}}
\usage{
copyOrderGroup(group, source_map, target_map)
}
\arguments{
\item{group}{the group or variable URL to be copied}

\item{source_map}{url to alias map for source variables}

\item{target_map}{alias to url map for target variables}
}
\value{
returns either a \code{\link{VariableGroup}} (if a group is supplied) or a URL (if just a variable URL is supplied)
}
\description{
Copy the order of a \code{VariableGroup} (or individual variable URL) from \code{VariableOrder}
}
\keyword{internal}
