% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cr_clip_to_env}
\alias{cr_clip_to_env}
\title{Get credentials from clipboard as environmental variables}
\usage{
cr_clip_to_env()
}
\description{
Some S3 buckets offer to add environmental variables to your clipboard,
this function allows directly adding those to the current R session as
environment variables.
}
