% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destinations.R
\name{cr_destination}
\alias{cr_destination}
\alias{cr_create_destination}
\title{Create or get a data destination on the crunch server}
\usage{
cr_destination(name)

cr_create_destination(name, data, ...)
}
\arguments{
\item{name}{A unique name to refer to this destination by}

\item{data}{A list, generally containing type="s3folder", bucket
with the bucket's name, and base for a prefix to use when exporting
data to.}

\item{...}{Saved for future expansion}
}
\value{
A ShojiObject data destination
}
\description{
Data destinations are S3 buckets that the crunch server can send
data to.
}
\examples{
destination <- cr_create_destination(
    "my-destination",
    data = list(type = "s3folder", bucket = "my-bucket", base = "lake")
)
destination <- cr_destination("my-destination")
}
