% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{cr_export}
\alias{cr_export}
\title{Export data & metadata from crunch lake}
\usage{
cr_export(ds, ..., path = NULL, destination = NULL)
}
\arguments{
\item{ds}{A \code{CrunchDataset} or URI for a crunch dataset}

\item{...}{Options passed to the API, see \url{https://crunch.io/api/reference/#post-/datasources/download/}
for complete list.}

\item{path}{Path to where files should be saved, defaults to a directory with dataset's
name in the current directory}

\item{destination}{Optional data destination to export data to (see \code{\link[=cr_destination]{cr_destination()}})}
}
\value{
If \code{destination} is unspecified, a string path to the folder containing
the downloaded files, or if it is a reference to the destination and path.
Both can have \code{\link[=cr_read_data]{cr_read_data()}} or \code{\link[=cr_read_meta]{cr_read_meta()}} and other reading functions called on.
}
\description{
Export data & metadata from crunch lake
}
\examples{
\dontrun{
files <- cr_export(ds, export_filter = list(variables = I(c("wave", "age"))))

md <- files |> cr_read_meta()
data <- files |> cr_read_data()
}
}
