% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lake_data.R
\name{cr_read_data_long}
\alias{cr_read_data_long}
\title{Read downloaded Crunch Lake data files}
\usage{
cr_read_data_long(x, ...)
}
\arguments{
\item{x}{A folder path or destination from \code{\link[=cr_export]{cr_export()}}}

\item{...}{Saved for future expansion}
}
\description{
Currently just a thin wrapper for reading the parquet file
as sent by the Crunch Lake in a "long" format. See \code{\link[=cr_read_data]{cr_read_data()}} to
read the data in a 'tidy'-er wide format.
}
