% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lake_meta.R
\name{cr_read_meta}
\alias{cr_read_meta}
\title{Read downloaded Crunch Lake metadata/schema files into a friendly format}
\usage{
cr_read_meta(x, ...)
}
\arguments{
\item{x}{A folder path or destination from \code{\link[=cr_export]{cr_export()}}}

\item{...}{Saved for future expansion}
}
\description{
A 'tidy' R representation of the information contained in the metadata, schema
and folders files provided by the Crunch Lake. See \code{\link[=cr_read_metadata_raw]{cr_read_metadata_raw()}},
\code{\link[=cr_read_schema]{cr_read_schema()}}, and \code{\link[=cr_read_folders]{cr_read_folders()}}  to read the raw files.
}
