% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lake_meta.R
\name{cr_read_metadata_raw}
\alias{cr_read_metadata_raw}
\alias{cr_read_schema}
\alias{cr_read_folders}
\title{Read downloaded Crunch Lake metadata/schema files}
\usage{
cr_read_metadata_raw(x, ...)

cr_read_schema(x, ...)

cr_read_folders(x, ...)
}
\arguments{
\item{x}{A folder path or destination from \code{\link[=cr_export]{cr_export()}}}

\item{...}{Saved for future expansion}
}
\description{
Currently just thin wrappers for reading the JSON/parquet files
as sent by the Crunch Lake. See \code{\link[=cr_read_meta]{cr_read_meta()}} for
a friendlier way to load this.
}
