% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-array.R
\name{createSubvarDef}
\alias{createSubvarDef}
\title{createSubvarDef}
\usage{
createSubvarDef(var, str, delim, selected, not_selected, unanswered, missing)
}
\arguments{
\item{var}{The variable containing the delimited responses}

\item{str}{A string whose presence indicates a selection}

\item{delim}{The delimiter separating the responses}

\item{selected}{A character string used to indicate a selection}

\item{not_selected}{Character string identifying non-selection}

\item{unanswered}{Character string indicating non-response}

\item{missing}{A logical vector indicating which variable entries are missing}
}
\value{
A VariableDefinition
}
\description{
This function creates a single subvariable definition based on a character string
to search for and an originating variable. It uses regex to determine whether
a string is present in a delimited list, then substitutes the user supplied values
to indicate selection, non-selection, and missingness.
}
\details{
TODO: When Crunch allows variable derivation to be created via regex, this should
create a derivation instead of a definition with values.
}
\keyword{internal}
