% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/new-dataset.R
\name{createWithMetadataAndFile}
\alias{createWithMetadataAndFile}
\title{Make a dataset with metadata and a CSV}
\usage{
createWithMetadataAndFile(metadata, file, strict = TRUE, cleanup = TRUE)
}
\arguments{
\item{metadata}{a list representation of the dataset's metadata, which
will be JSON-serialized and POSTed.}

\item{file}{a path to a CSV file, optionally zipped, that corresponds to
the above metadata.}

\item{strict}{logical: must the metadata exactly match the data? Default is
TRUE.}

\item{cleanup}{logical: if the file upload fails, delete the dataset?
Default is TRUE.}
}
\value{
On success, a new dataset.
}
\description{
This function just takes what you give it and POSTs to Crunch. No
validation, no automatic wrapping in the Shoji envelope, etc.
}
\keyword{internal}

