% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{crunch-api-key}
\alias{crunch-api-key}
\title{Crunch API Keys}
\description{
The rcrunch package recommends using API keys for authentication.
Using a username and password with the \code{\link[=login]{login()}} function is
deprecated and may be removed in a future release.
}
\details{
To get an API key for your account, (EXISTING DOCUMENTATION??).

The rcrunch package looks for the key in the environmental variable
"R_CRUNCH_API_KEY" or the option "crunch.api.key" (see \code{\link[=envOrOption]{envOrOption()}}
for details).

One way to establish your key is to add it to your ".Renviron"
file. This file is located in your home directory (you can
use \code{usethis::edit_r_environ()} to open the file if you have the
\code{usethis} package installed). The .Renviron file has the name of the
environment variable, followed by an equal sign and then the value. It
is good practice to set the API host too, (usually equal to
"https://app.crunch.io/api/").\preformatted{R_CRUNCH_API=https://app.crunch.io/api/
R_CRUNCH_API_KEY=YOUR_SECRET_KEY
}

You can either restart your session, or run \code{readRenviron("~/.Renviron")}
and then rcrunch will know to use your key going forward.
}
