% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R, R/shoji.R, R/variable.R
\name{is.dataset}
\alias{is.Array}
\alias{is.CA}
\alias{is.Categorical}
\alias{is.CategoricalArray}
\alias{is.Datetime}
\alias{is.MR}
\alias{is.Multiple}
\alias{is.MultipleResponse}
\alias{is.Numeric}
\alias{is.Text}
\alias{is.dataset}
\alias{is.shoji}
\alias{is.variable}
\title{Is it?}
\usage{
is.dataset(x)

is.shoji(x)

is.variable(x)

is.Numeric(x)

is.Categorical(x)

is.Text(x)

is.Datetime(x)

is.Multiple(x)

is.MR(x)

is.MultipleResponse(x)

is.CA(x)

is.Array(x)

is.CategoricalArray(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical
}
\description{
Is it?
}

