% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R, R/tab-book.R
\docType{methods}
\name{crunch-length}
\alias{crunch-length}
\alias{length,ShojiOrder-method}
\alias{length,OrderGroup-method}
\alias{length,TabBookResult-method}
\alias{length,MultitableResult-method}
\alias{[[,TabBookResult,numeric-method}
\alias{[[,TabBookResult,character-method}
\alias{[[,MultitableResult,ANY-method}
\title{Length of Crunch objects}
\usage{
length(x)

\S4method{length}{ShojiOrder}(x)

\S4method{length}{OrderGroup}(x)

\S4method{length}{TabBookResult}(x)

\S4method{length}{MultitableResult}(x)

\S4method{[[}{TabBookResult,numeric}(x, i, j, ...)

\S4method{[[}{TabBookResult,character}(x, i, j, ...)

\S4method{[[}{MultitableResult,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{a Crunch object (catalog, order, etc.)}
}
\value{
Integer: the number of elements in the catalog, group, folder, etc.
}
\description{
Length of Crunch objects
}
