% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{crunch-metadata}
\alias{crunch-metadata}
\alias{label}
\alias{set_label}
\alias{label<-}
\alias{description}
\alias{set_description}
\alias{description<-}
\alias{notes}
\alias{set_notes}
\alias{notes<-}
\alias{path}
\alias{set_path}
\alias{path<-}
\alias{values}
\alias{set_values}
\alias{values<-}
\title{Get & Set crunch metadata}
\usage{
label(x, ...)

set_label(x, value, ...)

label(x) <- value

description(x, ...)

set_description(x, value)

description(x) <- value

notes(x, ...)

set_notes(x, value, ...)

notes(x) <- value

path(x, ...)

set_path(x, value, ...)

path(x) <- value

values(x, ...)

set_values(x, value, ...)

values(x) <- value
}
\arguments{
\item{x}{A crunch variable}

\item{...}{Saved for future expansion}

\item{value}{Replacement value}
}
\description{
Crunch Variables have metadata associated with them beyond what is typically
used by R vectors. All variables have
\itemize{
\item label: A short description of the variable (suitable for labeling a chart).
Labels must be unique across all variables in a dataset.
\item description: A longer description of the variable
\item notes: Other information about the variable
\item path: Folder path in the CrunchDataset's folders
}
}
