% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunch.R
\docType{package}
\name{crunch-package}
\alias{crunch}
\alias{crunch-package}
\title{crunch: Crunch.io Data Tools}
\description{
The Crunch.io service \url{https://crunch.io/} provides a cloud-based data store and analytic engine, as well as an intuitive web interface. Using this package, analysts can interact with and manipulate Crunch datasets from within R. Importantly, this allows technical researchers to collaborate naturally with team members, managers, and clients who prefer a point-and-click interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://crunch.io/r/crunch/}
  \item \url{https://github.com/Crunch-io/rcrunch}
  \item Report bugs at \url{https://github.com/Crunch-io/rcrunch/issues}
}

}
\author{
\strong{Maintainer}: Greg Freedman Ellis \email{greg@crunch.io}

Authors:
\itemize{
  \item Jonathan Keane \email{jkeane@gmail.com}
  \item Mike Malecki \email{mike@crunch.io}
  \item Neal Richardson \email{neal.p.richardson@gmail.com}
  \item Gordon Shotwell \email{gordon.shotwell@gmail.com}
  \item Aljaž Sluga \email{aljaz.sluga@crunch.io}
}

}
\keyword{internal}
