% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.CategoricalVariable}
\alias{summary.CategoricalVariable}
\alias{summary.NumericVariable}
\title{#}
\usage{
\method{summary}{CategoricalVariable}(object, ...)

\method{summary}{NumericVariable}(object, ...)
}
\arguments{
\item{object}{A Variable}

\item{...}{additional arguments, ignored (they're in the summary.default)
signature}
}
\value{
a summary response. Categorical variable summaries should work like
summary.factor; Numeric variables should be like summary.numeric. Other
Variable types are not yet supported.
}
\description{
#
}

