% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_crunch_categorical_array_variable.R
\name{crunch_categorical_array_variable}
\alias{crunch_categorical_array_variable}
\title{Create Crunch Categorical Array Variable}
\usage{
crunch_categorical_array_variable(
  x,
  label,
  values = NULL,
  axes = NULL,
  description = NULL,
  notes = NULL,
  path = NULL,
  ...
)
}
\arguments{
\item{x}{A data.frame of values data}

\item{label}{The variable's label}

\item{values}{A data.frame of values metadata}

\item{axes}{A data.frame of axes metadata (\code{NULL}, the default requires that x be a
data.frame of crunch_variables, which have the necessary metadata)}

\item{description}{The variable's description}

\item{notes}{The variable's notes}

\item{path}{The variable's folder path}

\item{...}{Ignored}
}
\value{
A \code{crunch_categorical_array_variable} object
}
\description{
Crunch categorical array variables are kind of like a "packed" data.frame of factors
but with extra metadata. TODO: WRITE MORE!
}
