% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_crunch_categorical_variable.R
\name{crunch_categorical_variable}
\alias{crunch_categorical_variable}
\alias{is.crunch_categorical_variable}
\title{Create Crunch Categorical Variable}
\usage{
crunch_categorical_variable(
  x,
  label,
  values,
  description = NULL,
  notes = NULL,
  path = NULL,
  ...
)

is.crunch_categorical_variable(x)
}
\arguments{
\item{x}{A vector of values data}

\item{label}{The variable's label}

\item{values}{A data.frame of values metadata}

\item{description}{The variable's description}

\item{notes}{The variable's notes}

\item{path}{The variable's folder path}

\item{...}{Ignored}
}
\value{
A \code{crunch_categorical_variable} object
}
\description{
Crunch categorical variables are kind of like R's base factors, but contain more
metadata. TODO: WRITE MORE!
}
