% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{crunch_sitrep}
\alias{crunch_sitrep}
\title{Crunch situation report}
\usage{
crunch_sitrep(redact = TRUE, verbose = TRUE)
}
\arguments{
\item{redact}{Whether to redact the API key found (default TRUE)}

\item{verbose}{Whether to print information to the console (default TRUE)}
}
\value{
Invisibly, a list with information about the API
}
\description{
Get a  situation report on how R will connect to crunch.io
}
\examples{
\dontrun{
crunch_sitrep()
}
}
