% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_crunch_text_variable.R
\name{crunch_text_variable}
\alias{crunch_text_variable}
\alias{is.crunch_text_variable}
\title{Create Crunch Text Variable}
\usage{
crunch_text_variable(
  x,
  label,
  values,
  description = NULL,
  notes = NULL,
  path = NULL,
  ...
)

is.crunch_text_variable(x)
}
\arguments{
\item{x}{A vector of values data}

\item{label}{The variable's label}

\item{description}{The variable's description}

\item{notes}{The variable's notes}

\item{path}{The variable's folder path}

\item{...}{Ignored}
}
\value{
A \code{crunch_text_variable} object
}
\description{
Crunch text variables are very similar to regular character vectors in R,
but have a few additional metadata fields that apply to the whole variable
}
