% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{crunch_user_agent}
\alias{crunch_user_agent}
\title{Generate or extend the User-Agent string}
\usage{
crunch_user_agent(...)
}
\arguments{
\item{...}{Additional character terms to add to the User-Agent string}
}
\value{
The User-Agent string. Provide this appropriately in requests or set
globally with \code{\link[=set_crunch_config]{set_crunch_config()}}.
}
\description{
By default, the names and versions of curl, httr, and any attached Crunch
packages are included in the User-Agent request header. You can add to this
using this function.
}
\keyword{internal}
