% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchbox.R
\name{crunchBox}
\alias{crunchBox}
\title{Make a CrunchBox}
\usage{
crunchBox(dataset, filters = crunch::filters(dataset), brand_colors,
  static_colors, category_color_lookup, ...)
}
\arguments{
\item{dataset}{CrunchDataset}

\item{filters}{FilterCatalog, or \code{NULL} for no filters. Default all
filters in your catalog, \code{filters(dataset)}.}

\item{brand_colors}{an optional list with colors (in hex format) named 'primary', 'secondary', and 'message'}

\item{static_colors}{an optional list of static colors (in hex format) to use for categoricals (taken in order)}

\item{category_color_lookup}{an optional list of category names to colors (in hex format) to use for that category (for fine grained tuning of category colors.)}

\item{...}{additional metadata for the box, such as "title", "header", etc.}
}
\value{
The URL to the newly created box.
}
\description{
CrunchBoxes allow you to publish results to the world.
}
\seealso{
\code{\link{preCrunchBoxCheck}} to provide guidance on what you're including in the CrunchBox
}
