% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CrunchDataset}
\alias{CrunchDataset}
\alias{CrunchDataset-class}
\alias{CrunchDatasetEager-class}
\alias{CrunchDatasetEager}
\alias{CrunchDatasetLazy-class}
\alias{CrunchDatasetLazy}
\title{Crunch Datasets}
\usage{
CrunchDataset(..., labelSpecialVars = TRUE)
}
\arguments{
\item{...}{Passed to \code{CrunchDataset} subclass initialize functions}

\item{labelSpecialVars}{Whether to separate hidden & secure variables
from the main set of variables (setting \code{FALSE} can speed up loading
but the list of hidden & secure variables won't be available)}
}
\description{
An S4 class that contains a Crunch data. Generally you'll want to
use \code{\link[=loadDataset]{loadDataset()}} to create one.
}
