% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{CrunchVariable-class}
\alias{CrunchVariable}
\alias{CrunchVariable-class}
\title{Variables in Crunch}
\description{
Variables are S4 objects. All inherit from the base class
\code{CrunchVariable}.
}
\section{Slots}{

\describe{
\item{\code{readonly}}{logical: should changes made to this variable object locally
be persisted on the server? Default is \code{FALSE}}

\item{\code{tuple}}{An object of class VariableTuple. These contain attributes, such
as name and description, that are found in the index of the \code{\link{VariableCatalog}}}
}}

