% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cube-dims.R, R/cube-result.R,
%   R/cube-variables.R
\name{dimensions}
\alias{dimensions}
\alias{dimensions<-}
\alias{measures}
\alias{cube-methods}
\alias{dimnames,CubeDims-method}
\alias{dim,CubeDims-method}
\alias{is.na,CubeDims-method}
\alias{dimensions,CrunchCube-method}
\alias{dimensions<-,CrunchCube,CubeDims-method}
\alias{dim,CrunchCube-method}
\alias{dimnames,CrunchCube-method}
\alias{measures,CrunchCube-method}
\title{Methods on Cube objects}
\usage{
dimensions(x)

dimensions(x) <- value

measures(x)

\S4method{dimnames}{CubeDims}(x)

\S4method{dim}{CubeDims}(x)

\S4method{is.na}{CubeDims}(x)

\S4method{dimensions}{CrunchCube}(x)

\S4method{dimensions}{CrunchCube,CubeDims}(x) <- value

\S4method{dim}{CrunchCube}(x)

\S4method{dimnames}{CrunchCube}(x)

\S4method{measures}{CrunchCube}(x)
}
\arguments{
\item{x}{a CrunchCube or its CubeDims component.}

\item{value}{for \verb{dimensions<-} a \code{CubeDims} object to overwrite a CrunchCube
dimensions}
}
\value{
Generally, the same shape of result that each of these functions
return when applied to an \code{array} object.
}
\description{
These methods provide an \code{array}-like interface to the CrunchCube
object.
}
\seealso{
\code{\link{cube-computing}} \code{\link[base:array]{base::array}}
}
