% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dashboard}
\alias{dashboard}
\alias{setDashboardURL}
\alias{dashboard<-}
\title{View or set a dashboard URL}
\usage{
dashboard(x)

setDashboardURL(x, value)

dashboard(x) <- value
}
\arguments{
\item{x}{CrunchDataset}

\item{value}{For the setter, a URL (character) or \code{NULL} to unset the
dashboard.}
}
\value{
The getter returns a URL (character) or \code{NULL}. The setter
returns the dataset (\code{x}).
}
\description{
You can designate a dashboard that will show when the dataset is loaded in
the Crunch web app. This dashboard could be a Crunch Shiny ("Crunchy") app,
a CrunchBox, an RMarkdown website or something else.
}
\examples{
\dontrun{
dashboard(ds) <- "https://shiny.crunch.io/example/"
}
}
