% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataset-extract.R
\docType{methods}
\name{dataset-extract}
\alias{$,CrunchDataset-method}
\alias{[,CrunchDataset,character-method}
\alias{[,CrunchDataset,missing-method}
\alias{[[,CrunchDataset,ANY-method}
\alias{[[,CrunchDataset,character-method}
\alias{dataset-extract}
\title{Subset datasets and extract variables}
\usage{
\S4method{[}{CrunchDataset,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{CrunchDataset,character}(x, i, j, ..., drop = TRUE)

\S4method{[}{CrunchDataset,missing}(x, i, j, ..., drop = FALSE)

\S4method{[[}{CrunchDataset,ANY}(x, i, ..., drop = FALSE)

\S4method{[[}{CrunchDataset,character}(x, i, ..., drop = FALSE)

\S4method{$}{CrunchDataset}(x, name)
}
\arguments{
\item{x}{a CrunchDataset}

\item{i}{if character, identifies variables to extract based on their
aliases (by default, i.e. when x's \code{useAlias} is TRUE); if numeric or
logical, extracts variables accordingly. Note that this is the as.list
extraction, columns of the dataset rather than rows.}

\item{j}{column extraction, as described above}

\item{...}{additional arguments}

\item{drop}{logical: autmatically simplify a 1-column Dataset to a Variable?
Default is FALSE, and the TRUE option is in fact not implemented.}

\item{name}{like \code{i} but for \code{\$}}
}
\value{
\code{[} yields a Dataset; \code{[[} and \$ return a Variable
}
\description{
Subset datasets and extract variables
}

