% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{dataset-owner}
\alias{dataset-owner}
\alias{owner}
\alias{owner<-}
\alias{owner,CrunchDataset-method}
\alias{owner<-,CrunchDataset-method}
\title{Change the owner of a dataset}
\usage{
\S4method{owner}{CrunchDataset}(x)

\S4method{owner}{CrunchDataset}(x) <- value
}
\arguments{
\item{x}{CrunchDataset}

\item{value}{For the setter, either a URL (character) or a Crunch object
with a \code{self} method. Users and Projects are valid objects to assign
as dataset owners.}
}
\value{
The dataset.
}
\description{
Change the owner of a dataset
}
