% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-update.R
\docType{methods}
\name{dataset-update}
\alias{dataset-update}
\alias{[[<-,CrunchDataset,character,missing,CrunchVariable-method}
\alias{[[<-,CrunchDataset,ANY,missing,CrunchVariable-method}
\alias{[[<-,CrunchDataset,character,missing,ANY-method}
\alias{[[<-,CrunchDataset,character,missing,CrunchLogicalExpr-method}
\alias{[[<-,CrunchDataset,ANY,ANY,ANY-method}
\alias{[[<-,CrunchDataset,character,missing,NULL-method}
\alias{[[<-,CrunchDataset,ANY,missing,NULL-method}
\alias{$<-,CrunchDataset-method}
\alias{[[<-,CrunchDataset,character,missing,CrunchGeography-method}
\alias{[[<-,CrunchDataset,ANY,missing,CrunchGeography-method}
\alias{[<-,CrunchDataset,ANY,missing,list-method}
\alias{[<-,CrunchDataset,ANY,missing,CrunchDataset-method}
\alias{[<-,CrunchDataset,CrunchExpr,ANY,ANY-method}
\title{Update a variable or variables in a dataset}
\usage{
\S4method{[[}{CrunchDataset,character,missing,CrunchVariable}(x,
  i) <- value

\S4method{[[}{CrunchDataset,ANY,missing,CrunchVariable}(x, i) <- value

\S4method{[[}{CrunchDataset,character,missing,ANY}(x, i) <- value

\S4method{[[}{CrunchDataset,character,missing,CrunchLogicalExpr}(x,
  i) <- value

\S4method{[[}{CrunchDataset,ANY,ANY,ANY}(x, i) <- value

\S4method{[[}{CrunchDataset,character,missing,`NULL`}(x, i) <- value

\S4method{[[}{CrunchDataset,ANY,missing,`NULL`}(x, i) <- value

\S4method{$}{CrunchDataset}(x, name) <- value

\S4method{[[}{CrunchDataset,character,missing,CrunchGeography}(x,
  i) <- value

\S4method{[[}{CrunchDataset,ANY,missing,CrunchGeography}(x, i) <- value

\S4method{[}{CrunchDataset,ANY,missing,list}(x, i, j) <- value

\S4method{[}{CrunchDataset,ANY,missing,CrunchDataset}(x, i, j) <- value

\S4method{[}{CrunchDataset,CrunchExpr,ANY,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{a CrunchDataset}

\item{i}{For \code{[}, a \code{CrunchLogicalExpr}, numeric, or logical
vector defining a subset of the rows of \code{x}. For \code{[[}, see
\code{j} for the as.list column subsetting.}

\item{value}{replacement values to insert. These can be \code{crunchExpr}s
or R vectors of the corresponding type}

\item{name}{like \code{j} but for \code{$}}

\item{j}{if character, identifies variables to extract based on their
aliases (by default: set \code{options(crunch.namekey.dataset="name")}
to use variable names); if numeric or
logical, extracts variables accordingly. Note that this is the as.list
extraction, columns of the dataset rather than rows.}
}
\value{
\code{x}, modified.
}
\description{
Update a variable or variables in a dataset
}
