% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DatasetOrder-class}
\alias{DatasetGroup}
\alias{DatasetGroup-class}
\alias{DatasetOrder}
\alias{DatasetOrder-class}
\title{Organize Datsets}
\description{
A DatasetOrder object is a subclass of \code{list} that contains
DatasetGroups. DatasetGroup objects contain a group name and an set of
"entities", which can be dataset references or other nested DatasetGroups.
}
\section{Slots}{

\describe{
\item{\code{group}}{character, the name of the DatasetGroup. In the constructor and
more generally, this field can be referenced as "name" as well.}

\item{\code{entities}}{a character vector of dataset URLs, or a list containing a
combination of dataset URLs and DatasetGroup objects.}
}}

