% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{deck-titles}
\alias{deck-titles}
\alias{titles<-}
\alias{title}
\alias{title<-}
\alias{subtitles}
\alias{subtitles<-}
\alias{subtitle}
\alias{subtitle<-}
\title{Manipulate deck titles}
\usage{
titles(x)

titles(x) <- value

title(x)

title(x) <- value

subtitles(x, value)

subtitles(x) <- value

subtitle(x, value)

subtitle(x) <- value
}
\arguments{
\item{x}{a CrunchDeck or CrunchSlide}

\item{value}{(Character) the new title or subtitle}
}
\description{
Crunch slides have titles an subtitles. You can change these features at
either the Deck level by assigning a character vector which is
the same length as the Deck to the Crunch deck, or by assigning character strings to the
the slide.
}
