% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{decks}
\alias{decks}
\title{Get a dataset's DeckCatalog}
\usage{
decks(dataset)
}
\arguments{
\item{dataset}{a Crunch Dataset}
}
\value{
a DeckCatalog
}
\description{
Crunch decks are stored in catalogs. This function returns those catalogs so
that you can access and manipulate decks in R.
}
