% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{deleteDataset}
\alias{deleteDataset}
\title{Delete a dataset from the dataset list}
\usage{
deleteDataset(x, ...)
}
\arguments{
\item{x}{The name (character) of a dataset, a path to a dataset, or a
\code{CrunchDataset}. Unless \code{x} is a parsed folder path, it can only be of
length 1--for your protection, this function is not vectorized.}

\item{...}{additional parameters passed to \code{\link[=delete]{delete()}}}
}
\value{
(Invisibly) the API response from deleting the dataset
}
\description{
This function lets you delete a dataset without first loading it, which is
faster.
}
\details{
The function also works on \code{CrunchDataset} objects, just like
\code{\link[=delete]{delete()}}, which may be useful if you have loaded another
package that masks the \code{crunch::delete()} method.
}
\seealso{
\code{\link[=delete]{delete()}}; \code{\link[=cd]{cd()}} for details of parsing and walking dataset
folder/project paths.
}
