% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-delete.R
\name{deleteVariables}
\alias{deleteVariables}
\alias{deleteVariable}
\title{Delete Variables Within a Dataset}
\usage{
deleteVariables(dataset, variables)

deleteVariable(dataset, variables)
}
\arguments{
\item{dataset}{the Dataset to modify}

\item{variables}{aliases (following \code{crunch.namekey.dataset}) or indices
of variables to delete.}
}
\value{
(invisibly) \code{dataset} with the specified variables deleted
}
\description{
Deleting requires confirmation. In an interactive session, you will be asked
to confirm. To avoid that prompt, or to delete variables from a
non-interactive session, wrap the call in \code{\link{with_consent}} to give
your permission to delete.
}
\seealso{
\code{\link{hide}}
}
