% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-derivation.R
\docType{methods}
\name{derivations}
\alias{derivations}
\alias{derivation}
\alias{derivation<-}
\alias{derivation,CrunchVariable-method}
\alias{derivation<-,CrunchVariable,ANY-method}
\alias{derivation<-,CrunchVariable,NULL-method}
\title{Get or set a derived variable's \code{CrunchExpr}}
\usage{
\S4method{derivation}{CrunchVariable}(x)

\S4method{derivation}{CrunchVariable,ANY}(x) <- value

\S4method{derivation}{CrunchVariable,`NULL`}(x) <- value
}
\arguments{
\item{x}{a variable}

\item{value}{a \code{CrunchExpr} to be used as the derivation (for the setter only)}
}
\value{
a \code{CrunchExpr} of the derivation
}
\description{
Get a derived variable's derivation formula as a (\code{CrunchExpr})\link{expressions} with
\code{derivation([variable])}. Set (change) a derived variable's derivation with
\code{derivation([variable]) <- [expression]}
}
\examples{
\dontrun{

ds$derived_v1 <- ds$v1 + 5

derivation(ds$derived_v1)
# Crunch expression: v1 + 5

derivation(ds$derived_v1) <- ds$v1 + 10
derivation(ds$derived_v1)
# Crunch expression: v1 + 10

# to integrate or instantiate the variable in place (remove the link between
# variable v1 and the derivation) you can:
derivation(ds$derived_v1) <- NULL
}
}
