% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-catalog.R, R/dataset.R, R/filters.R, R/fork-and-merge.R, R/share.R, R/subvariables.R, R/teams.R, R/user.R, R/variable-catalog.R, R/versions.R
\docType{methods}
\name{describe-catalog}
\alias{aliases}
\alias{aliases,Subvariables-method}
\alias{aliases,VariableCatalog-method}
\alias{aliases<-}
\alias{aliases<-,Subvariables-method}
\alias{aliases<-,VariableCatalog-method}
\alias{describe-catalog}
\alias{descriptions}
\alias{descriptions,VariableCatalog-method}
\alias{descriptions,VersionCatalog-method}
\alias{descriptions<-}
\alias{descriptions<-,VariableCatalog-method}
\alias{emails}
\alias{emails,PermissionCatalog-method}
\alias{emails,UserCatalog-method}
\alias{names,CategoricalArrayVariable-method}
\alias{names,CrunchDataset-method}
\alias{names,DatasetCatalog-method}
\alias{names,FilterCatalog-method}
\alias{names,ForkCatalog-method}
\alias{names,MemberCatalog-method}
\alias{names,Subvariables-method}
\alias{names,TeamCatalog-method}
\alias{names,UserCatalog-method}
\alias{names,VariableCatalog-method}
\alias{names,VersionCatalog-method}
\alias{names<-,Subvariables-method}
\alias{names<-,VariableCatalog-method}
\alias{timestamps}
\alias{timestamps,VersionCatalog-method}
\alias{types}
\alias{types,VariableCatalog-method}
\title{Get and set names, aliases on Catalog-type objects}
\usage{
\S4method{names}{DatasetCatalog}(x)

\S4method{names}{CrunchDataset}(x)

\S4method{names}{FilterCatalog}(x)

\S4method{names}{ForkCatalog}(x)

\S4method{emails}{PermissionCatalog}(x)

\S4method{names}{Subvariables}(x)

\S4method{names}{Subvariables}(x) <- value

\S4method{aliases}{Subvariables}(x)

\S4method{aliases}{Subvariables}(x) <- value

\S4method{names}{CategoricalArrayVariable}(x)

\S4method{names}{TeamCatalog}(x)

\S4method{names}{MemberCatalog}(x)

\S4method{names}{UserCatalog}(x)

\S4method{emails}{UserCatalog}(x)

\S4method{names}{VariableCatalog}(x)

\S4method{names}{VariableCatalog}(x) <- value

\S4method{aliases}{VariableCatalog}(x)

\S4method{aliases}{VariableCatalog}(x) <- value

\S4method{descriptions}{VariableCatalog}(x)

\S4method{descriptions}{VariableCatalog}(x) <- value

\S4method{types}{VariableCatalog}(x)

\S4method{names}{VersionCatalog}(x)

\S4method{descriptions}{VersionCatalog}(x)

\S4method{timestamps}{VersionCatalog}(x)
}
\arguments{
\item{x}{a VariableCatalog, Subvariables, or similar object}

\item{value}{For the setters, an appropriate-length character vector to
assign}
}
\value{
Getters return the character object in the specified slot; setters
return \code{x} duly modified.
}
\description{
These methods let you get and set names and aliases for variables in a
Dataset's catalog, or within \code{\link{Subvariables}} in an array
variable. They work like the base R names methods.
}
\details{
Note that the \code{names} method on a Dataset returns the aliases of its
variables by default. This is controlled by
\code{getOption("crunch.namekey.dataset")}, which is "alias" by default.
Set \code{options(crunch.namekey.dataset="name")} if you wish to use
variable names. See the vignette on variables for more information.
}
\seealso{
\code{\link{Subvariables}} \code{\link{Categories}} \code{\link[base]{names}} \code{vignette("variables", package="crunch")}
}

