% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/abstract-category.R,
%   R/category.R, R/dataset.R, R/decks.R, R/geo.R, R/multitables.R,
%   R/project-folder.R, R/shoji-folder.R, R/shoji.R, R/variable.R
\docType{methods}
\name{describe-entity}
\alias{describe-entity}
\alias{describe}
\alias{name}
\alias{name<-}
\alias{description}
\alias{description<-}
\alias{alias}
\alias{alias<-}
\alias{startDate}
\alias{startDate<-}
\alias{endDate}
\alias{endDate<-}
\alias{notes}
\alias{notes<-}
\alias{digits}
\alias{digits<-}
\alias{uniformBasis}
\alias{uniformBasis<-}
\alias{id}
\alias{value}
\alias{value<-}
\alias{name,AbstractCategory-method}
\alias{name<-,AbstractCategory-method}
\alias{name<-,NULL-method}
\alias{id,AbstractCategory-method}
\alias{value,Category-method}
\alias{value<-,Category-method}
\alias{name,CrunchDataset-method}
\alias{name<-,CrunchDataset-method}
\alias{description,CrunchDataset-method}
\alias{description<-,CrunchDataset-method}
\alias{startDate,CrunchDataset-method}
\alias{startDate<-,CrunchDataset-method}
\alias{endDate,CrunchDataset-method}
\alias{endDate<-,CrunchDataset-method}
\alias{id,CrunchDataset-method}
\alias{notes,CrunchDataset-method}
\alias{notes<-,CrunchDataset-method}
\alias{name<-,CrunchDeck-method}
\alias{description,CrunchDeck-method}
\alias{description<-,CrunchDeck-method}
\alias{description,Geodata-method}
\alias{name<-,Multitable-method}
\alias{name,ProjectFolder-method}
\alias{name<-,ShojiFolder-method}
\alias{name,ShojiObject-method}
\alias{alias,VariableTuple-method}
\alias{description,VariableTuple-method}
\alias{notes,VariableTuple-method}
\alias{name,CrunchVariable-method}
\alias{name<-,CrunchVariable-method}
\alias{id,CrunchVariable-method}
\alias{description,CrunchVariable-method}
\alias{description<-,CrunchVariable-method}
\alias{alias,CrunchVariable-method}
\alias{alias<-,CrunchVariable-method}
\alias{notes,CrunchVariable-method}
\alias{notes<-,CrunchVariable-method}
\alias{digits,CrunchVariable-method}
\alias{digits<-,NumericVariable-method}
\alias{digits<-,CrunchVariable-method}
\alias{uniformBasis,MultipleResponseVariable-method}
\alias{uniformBasis<-,MultipleResponseVariable-method}
\title{Name, alias, and description for Crunch objects}
\usage{
name(x)

name(x) <- value

id(x)

value(x)

value(x) <- value

description(x)

description(x) <- value

startDate(x)

startDate(x) <- value

endDate(x)

endDate(x) <- value

alias(object, ...)

alias(x) <- value

digits(x)

digits(x) <- value

uniformBasis(x)

uniformBasis(x) <- value

notes(x)

notes(x) <- value

\S4method{name}{AbstractCategory}(x)

\S4method{name}{AbstractCategory}(x) <- value

\S4method{name}{`NULL`}(x) <- value

\S4method{id}{AbstractCategory}(x)

\S4method{value}{Category}(x)

\S4method{value}{Category}(x) <- value

\S4method{name}{CrunchDataset}(x)

\S4method{name}{CrunchDataset}(x) <- value

\S4method{description}{CrunchDataset}(x)

\S4method{description}{CrunchDataset}(x) <- value

\S4method{startDate}{CrunchDataset}(x)

\S4method{startDate}{CrunchDataset}(x) <- value

\S4method{endDate}{CrunchDataset}(x)

\S4method{endDate}{CrunchDataset}(x) <- value

\S4method{id}{CrunchDataset}(x)

\S4method{notes}{CrunchDataset}(x)

\S4method{notes}{CrunchDataset}(x) <- value

\S4method{name}{CrunchDeck}(x) <- value

\S4method{description}{CrunchDeck}(x)

\S4method{description}{CrunchDeck}(x) <- value

\S4method{description}{Geodata}(x)

\S4method{name}{Multitable}(x) <- value

\S4method{name}{ProjectFolder}(x)

\S4method{name}{ShojiFolder}(x) <- value

\S4method{name}{ShojiObject}(x)

\S4method{alias}{VariableTuple}(object)

\S4method{description}{VariableTuple}(x)

\S4method{notes}{VariableTuple}(x)

\S4method{name}{CrunchVariable}(x)

\S4method{name}{CrunchVariable}(x) <- value

\S4method{id}{CrunchVariable}(x)

\S4method{description}{CrunchVariable}(x)

\S4method{description}{CrunchVariable}(x) <- value

\S4method{alias}{CrunchVariable}(object)

\S4method{alias}{CrunchVariable}(x) <- value

\S4method{notes}{CrunchVariable}(x)

\S4method{notes}{CrunchVariable}(x) <- value

\S4method{digits}{CrunchVariable}(x)

\S4method{digits}{NumericVariable}(x) <- value

\S4method{digits}{CrunchVariable}(x) <- value

\S4method{uniformBasis}{MultipleResponseVariable}(x)

\S4method{uniformBasis}{MultipleResponseVariable}(x) <- value
}
\arguments{
\item{x}{a Dataset or Variable.}

\item{value}{For the setters, a length-1 character vector to assign}

\item{object}{Same as \code{x} but for the \code{alias} method, in order to
match the generic from another package. Note that \code{alias} and \code{digits} are
only defined for Variables.}

\item{...}{additional arguments in the \code{alias} generic, ignored.}
}
\value{
Getters return the character object in the specified slot; setters
return \code{x} duly modified.
}
\description{
Name, alias, and description for Crunch objects
}
\seealso{
\code{\link{Categories}} \code{\link{describe-catalog}}
}
