% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dichotomize,Categories,numeric-method}
\alias{dichotomize,Categories,numeric-method}
\title{Indicate how categories represent a dichotomized value}
\usage{
\S4method{dichotomize}{Categories,numeric}(x, i)
}
\description{
Indicate how categories represent a dichotomized value
}

