% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-dims.R
\docType{methods}
\name{dimType}
\alias{dimType}
\alias{dimType,CrunchCube-method}
\alias{dimType,CubeDims-method}
\title{Get dimension type}
\usage{
\S4method{dimType}{CrunchCube}(x)

\S4method{dimType}{CubeDims}(x)
}
\arguments{
\item{CrunchCube}{}
}
\value{
A character vector. This is identical to \code{types()} except that
the array variable types are more specific.

a character vector of dimension types
}
\description{
This function returns the specific type of each cube dimension. This is useful
when cubes contain categorical array or multiple response variables because it
identifies the dimensions of the cube which refer to the differ parts of
array variable:
\itemize{
\item \code{ca_items}: Categorical array items
\item \code{ca_categories}: The categories of the categorical array
\item \code{mr_items}: Multiple response options or items
\item \code{mr_selections}: The selection status for a multiple response variable
}
}
\keyword{internal}
