% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{displaySettings}
\alias{displaySettings}
\alias{displaySettings<-}
\title{Get or set a slide's display settings}
\usage{
displaySettings(x)

displaySettings(x) <- value
}
\description{
Get or set a slide's display settings
}
