% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{doc}
\alias{doc}
\title{Collapse string vector for making docs}
\usage{
doc(x)
}
\arguments{
\item{x}{character}
}
\value{
character vector of length 1
}
\description{
Empty strings will be replaced with newlines.
}

