% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{download_lake_tabulation}
\alias{download_lake_tabulation}
\title{Download a tabulation from the Crunch Lake}
\usage{
download_lake_tabulation(
  ds,
  rows,
  columns,
  file,
  ...,
  format = guess_tabulation_format(file)
)
}
\arguments{
\item{ds}{A \code{CrunchDataset} or URI for a crunch dataset}

\item{rows, columns}{A character vector of aliases, a Crunch Variables Catalog, Crunch Variable,
or list of these things for the tabulation to be calculated over.}

\item{file}{Path to a file to save the tabulation as}

\item{...}{Other options passed to the API}

\item{format}{Format option passed to API (default is to guess based on filename)}
}
\value{
Filepath to the downloaded tabulation data
}
\description{
Download a tabulation from the Crunch Lake
}
\examples{
\dontrun{
tab_file <- download_lake_tabulation(ds, "wave", c("age", "healthy_eater"), "test.csv")
tab <- readr::read_csv(tab_file)
}
}
