% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-rows.R
\name{dropRows}
\alias{dropRows}
\title{Permanently delete rows from a dataset}
\usage{
dropRows(dataset, expr)
}
\arguments{
\item{dataset}{a \code{CrunchDataset}}

\item{expr}{a \code{CrunchLogicalExpr}}
}
\value{
\code{dataset} without the rows indicated by \code{expr}
}
\description{
Permanently delete rows from a dataset
}
\examples{
\dontrun{
ds <- dropRows(ds, ds$gender == "Male")
}
}
\seealso{
\link{exclusion} for a non-destructive way to suppress rows
}
