% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/expressions.R
\name{duplicated}
\alias{duplicated}
\alias{duplicated,CrunchVariable-method}
\alias{duplicated,CrunchExpr-method}
\title{"duplicated" method for Crunch objects}
\usage{
duplicated(x, incomparables = FALSE, ...)

\S4method{duplicated}{CrunchVariable}(x, incomparables = FALSE, ...)

\S4method{duplicated}{CrunchExpr}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{\code{CrunchVariable} or \code{CrunchExpr}}

\item{incomparables}{Ignored}

\item{...}{Ignored}
}
\value{
A \code{CrunchLogicalExpr} that evaluates \code{TRUE} for all repeated
entries after the first occurrence of a value.
}
\description{
"duplicated" method for Crunch objects
}
\seealso{
\code{\link[base:duplicated]{base::duplicated()}}
}
