% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-variables.R
\name{ensureValidCase}
\alias{ensureValidCase}
\title{Validate case statements for case variables}
\usage{
ensureValidCase(case)
}
\arguments{
\item{case}{a list for one case to test

List elements:
\enumerate{
\item \code{id}: an integer to use for this category when a case variable is made
(default: none, one will automatically be assigned when the case variable
is made)
\item \code{name}: a charcater identifier for this case
\item \code{expression}: a \code{CrunchLogicalExpr} that sets the conditions for the case, or "else"
\item \code{numeric_value}: a numeric that is the value this case should take on
when analyzing the data numerically (mean, standard deviation, etc.)
\item \code{missing} a logical indicating if this case should be treated as missing
}}
}
\description{
Validate case statements for case variables
}
\keyword{internal}
