% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-catalog.R
\name{entry_to_df}
\alias{entry_to_df}
\title{entry_to_df
Turns an entry in the catalog into}
\usage{
entry_to_df(entry, list_col_names = list_columns)
}
\arguments{
\item{entry}{}

\item{list_col_names}{}
}
\value{
A one row data frame
}
\description{
entry_to_df
Turns an entry in the catalog into
}
