% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{escapeRegex}
\alias{escapeRegex}
\title{escape Regex}
\usage{
escapeRegex(string)
}
\arguments{
\item{string}{}
}
\description{
This function takes a string and escapes all of the special characters in the string.
So VB.NET becomes VB.NET. Note that R will print this as VB\.NET, but \code{cat} reveals
that there's only one \code{}.
}
\examples{
\dontrun{
escapeRegex("Tom&Jerry")
escapeRegex(".Net")
}
}
\keyword{internal}
