% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{extralogical}
\alias{extralogical}
\alias{between}
\alias{all,CrunchVarOrExpr-method}
\alias{any,CrunchVarOrExpr-method}
\alias{isNoneAbove}
\alias{textContains}
\alias{anyNA,CrunchVarOrExpr-method}
\alias{allNA}
\alias{allValid}
\title{"extra" logical method for CrunchLogicalExpr}
\usage{
between(x, lower, upper, inclusive = c(TRUE, FALSE))

\S4method{all}{CrunchVarOrExpr}(x, ..., na.rm = FALSE)

\S4method{any}{CrunchVarOrExpr}(x, ..., na.rm = FALSE)

isNoneAbove(x)

textContains(x, regex, ignore_case = FALSE)

\S4method{anyNA}{CrunchVarOrExpr}(x, recursive = FALSE)

allNA(x)

allValid(x)
}
\arguments{
\item{x}{a CrunchVariable or CrunchExpr}
}
\value{
a CrunchLogicalExpr
}
\description{
"extra" logical method for CrunchLogicalExpr
}
