% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{extraselections}
\alias{extraselections}
\alias{asSelected}
\alias{selectedDepth}
\alias{arraySelections}
\title{"extra" method for handling selected categories in crunch variables}
\usage{
asSelected(x)

selectedDepth(x)

arraySelections(x)
}
\arguments{
\item{x}{a CrunchVariable or CrunchExpr}
}
\value{
a CrunchExpr
}
\description{
"extra" method for handling selected categories in crunch variables
}
