% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/filters.R
\name{filters}
\alias{filters}
\alias{filters<-}
\alias{filter-catalog}
\alias{filters,CrunchDataset-method}
\alias{filters<-,CrunchDataset-method}
\title{Get or set a dataset's filters}
\usage{
filters(x)

filters(x) <- value

\S4method{filters}{CrunchDataset}(x)

\S4method{filters}{CrunchDataset}(x) <- value
}
\arguments{
\item{x}{a CrunchDataset}

\item{value}{for the setter, a FilterCatalog}
}
\value{
an object of class FilterCatalog containing references to Filter
entities usable in the web application. (Setter returns the Dataset.)
}
\description{
You can build and save filters in the Crunch web app, and these filters
are stored in a \code{FilterCatalog}. This function allows you to retrieve
and modify those filters.
}
