% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\docType{methods}
\name{filter-methods}
\alias{filter-methods}
\alias{is.public}
\alias{is.public,CrunchFilter-method}
\alias{is.public<-}
\alias{is.public<-,CrunchFilter-method}
\title{View and modify Filter entity attributes}
\usage{
\S4method{is.public}{CrunchFilter}(x)

\S4method{is.public}{CrunchFilter}(x) <- value
}
\arguments{
\item{x}{a CrunchFilter}

\item{value}{an attribute to set}
}
\value{
For \code{is.public}, a logical value for whether the filter is
flagged as shared with all dataset viewers. (Its setter thus takes a
logical value as well.)
}
\description{
View and modify Filter entity attributes
}

