% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\docType{methods}
\name{team}
\alias{team}
\alias{team<-}
\alias{team,CrunchFilter-method}
\alias{team<-,CrunchFilter,CrunchTeam-method}
\alias{team<-,CrunchFilter,character-method}
\title{Get the team a filter is shared with}
\usage{
team(x)

team(x) <- value

\S4method{team}{CrunchFilter}(x)

\S4method{team}{CrunchFilter,CrunchTeam}(x) <- value

\S4method{team}{CrunchFilter,character}(x) <- value
}
\arguments{
\item{x}{a \code{CrunchFilter}}

\item{value}{a \code{CrunchTeam} or url for a Crunch team}
}
\value{
a url of the team that the filter is shared with.
}
\description{
Get the team a filter is shared with
}
