% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariables.R
\name{findVariables}
\alias{findVariables}
\title{Find variables and their paths in a Crunch dataset or folder}
\usage{
findVariables(x, deep = FALSE, include.hidden = FALSE, include.private = FALSE)
}
\arguments{
\item{x}{Crunch dataset or variable folder}

\item{deep}{Defaults to \code{FALSE}, \code{TRUE} recursively examines any subfolders as well}

\item{include.hidden}{Defaults to \code{FALSE}, \code{TRUE} finds any hidden variables as well}

\item{include.private}{Defaults to \code{FALSE}, \code{TRUE} finds any private variables as well}
}
\value{
Data.frame with one row per Crunch variable and columns \code{alias} (Crunch variable alias),
\code{path} (location of the variable, with " | " indicating nesting,
e.g. "Foo | Bar" indicates that the variable can be found in the folder "Bar" and that "Bar" is located in folder "Foo"),
\code{hidden} (\code{TRUE} or \code{FALSE}), \code{private} (\code{TRUE} or \code{FALSE})
}
\description{
Returns a data.frame whose rows correspond to Crunch variables found in \code{x}.
By default, only top-level, non-hidden, non-private variables in \code{x} are returned.
}
