% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-order.R
\name{flattenOrder}
\alias{flattenOrder}
\title{Remove nesting of groups within an order/group}
\usage{
flattenOrder(x)
}
\arguments{
\item{x}{VariableOrder, DatasetOrder, VariableGroup, or DatasetGroup; or a
CrunchDataset or catalog that has an \code{ordering} property.}
}
\value{
\code{x}, or its order resource, flattened.
}
\description{
This function reduces a potentially nested order to its flattened
representation, containing no nested groups. Entities are ordered in the
result by their first appearance in the order object.
}
