% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{forceVarCat}
\alias{forceVarCat}
\title{Force variables catalog to be loaded}
\usage{
forceVarCat(x)
}
\arguments{
\item{x}{A crunch dataset}
}
\value{
A dataset with it's variable catalogs filled in
}
\description{
Variables catalogs are generally loaded lazily, but this function
allows you to force them to be loaded. Generally you shouldn't have to
care about this except when writing tests.
}
