% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-array.R
\name{formMR}
\alias{formMR}
\title{Form a Multiple Response Variable from functions of existing variables}
\usage{
formMR(name, ..., .categories = NULL, .subvariables = NULL)
}
\arguments{
\item{name}{character, the name that the new Categorical Array variable should have.}

\item{...}{named \code{SubvariableDefinnition}s}

\item{.categories}{(Optional) Categories to define, defaults to the default MR
category (1 = "Selected", 0 = "Other")}

\item{.subvariables}{}
}
\value{
A \code{VariableDefinition} that when added to a Dataset will create the multiple-response
array and subvariables.
}
\description{
Multiple Response Variables are a type of array variables (a set of "subvariables"
bound together for display in the app) that take only 2 values ("selected" and "not").
This function forms an MR variable by taking functions of existing variables and
deriving them at the same time as the new variable.
}
