% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{fortify}
\alias{fortify}
\alias{fortify.CrunchDataFrame}
\alias{fortify.CrunchDataset}
\title{Fortify crunch objects for use with ggplot}
\usage{
\method{fortify}{CrunchDataFrame}(model, data, ...)

\method{fortify}{CrunchDataset}(model, data, ...)
}
\arguments{
\item{model}{model or other R object to convert to data frame}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\description{
Fortify crunch objects for use with ggplot
}
