% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{generateCutLabels}
\alias{generateCutLabels}
\title{Generate Labels for the cut function}
\usage{
generateCutLabels(dig.lab, breaks, nb, right, include.lowest)
}
\arguments{
\item{dig.lab}{see \code{cut()}}

\item{breaks}{see \code{cut()}}

\item{nb}{The number of breaks, equal to the small of 2 or the number of breaks}

\item{right}{see \code{cut()}}

\item{include.lowest}{see\code{cut()}}
}
\value{
A character vector of labels
}
\description{
A convenience function to generate labels for the cut function. This
function is extracted from \code{\link[base:cut]{base::cut()}} and is broken out to make it easier to
test. It is not meant to be called on its own.
}
\examples{

crunch:::generateCutLabels(2, c(2, 3, 4, 5), 4, FALSE, FALSE)

}
