% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{generateOrganizeCommand}
\alias{generateOrganizeCommand}
\title{Generate Crunch Automation commands to create a dataset's current folder struct}
\usage{
generateOrganizeCommand(dataset)
}
\arguments{
\item{dataset}{A crunch Dataset}
}
\value{
A string of the commands that can be passed to \code{\link[=runCrunchAutomation]{runCrunchAutomation()}}
}
\description{
Take a datset and generate the Crunch Automation commands needed to create the
current folder structure of the dataset. Useful for saving a snapshot, or copying
one dataset's folder structure to another (though see \code{\link[=copyFolders]{copyFolders()}} for another
way of copying a dataset's folders).
}
