% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{generateWeightEntry}
\alias{generateWeightEntry}
\title{Utility function to generate entries for makeWeight}
\usage{
generateWeightEntry(expr)
}
\arguments{
\item{expr}{An expression}
}
\value{
A list of the variable id and the target weights.
}
\description{
Utility function to generate entries for makeWeight
}
\keyword{internal}
