% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/geo.R
\docType{class}
\name{CrunchGeography-class}
\alias{CrunchGeography-class}
\alias{CrunchGeodata-class}
\alias{CrunchGeodata}
\alias{geo}
\alias{geo<-}
\alias{fetchGeoFile}
\alias{CrunchGeography,}
\alias{CrunchGeodata}
\alias{geo,CrunchVariable-method}
\alias{geo<-,CrunchVariable,CrunchGeography-method}
\alias{fetchGeoFile,CrunchGeography-method}
\title{Geography properties for crunch variables}
\usage{
\S4method{geo}{CrunchVariable}(x)

\S4method{geo}{CrunchVariable,CrunchGeography}(x) <- value

\S4method{fetchGeoFile}{CrunchGeography}(x)
}
\arguments{
\item{x}{a crunch variable}

\item{value}{value of the geography property to set}
}
\description{
Geography properties for crunch variables
}
