% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getAccountUserCatalog}
\alias{getAccountUserCatalog}
\title{Find all users on your account}
\usage{
getAccountUserCatalog(x = shojiURL(getAccount(), "catalogs", "users"))
}
\arguments{
\item{x}{URL of the user catalog. The default is your own user catalog and
you should not need to change this default.}
}
\value{
a \code{UserCatalog}
}
\description{
Find all users on your account
}
\keyword{internal}
