% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunch-data-frame.R
\name{grabColNames}
\alias{grabColNames}
\title{grab column names for subsetting crdfs}
\usage{
grabColNames(x, j, allow_logical = TRUE)
}
\arguments{
\item{x}{crdf to grab columns from}

\item{j}{character, logical, or integer vector specifying the columns to grab}

\item{allow_logical}{logical should logical subsetting be allowed? (default: TRUE)}
}
\value{
character names of the columns of the crdf to select
}
\description{
Take a character, logical, or integer vector and return the appropriate names
of the crdf to subset. Disallow logicals (in the case of \code{[[} subsetting, to
conform with \code{data.frame} methods)
}
\keyword{internal}
