% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunch-data-frame.R
\name{grabRowInd}
\alias{grabRowInd}
\title{grab row indicators for subsetting crdfs}
\usage{
grabRowInd(x, i)
}
\arguments{
\item{x}{crdf to grab rows from}

\item{i}{logical or integer vector specifying the rows to grab}
}
\value{
integers of the rows of the crdf to select
}
\description{
grab row indicators for subsetting crdfs
}
\keyword{internal}
