% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graft.R
\name{graftDataset}
\alias{graftDataset}
\title{Graft data from a cloned dataset back to its source}
\usage{
graftDataset(primary, clone)
}
\arguments{
\item{primary}{the CrunchDataset onto which you wish to graft data}

\item{clone}{the CrunchDataset you wish to take data from to graft onto
\code{primary}}
}
\value{
\code{primary} with the data from \code{clone} grafted onto it
}
\description{
This lets you prepare edits for a dataset on a clone of the dataset, then
copy that data back to the original dataset once you have verified that the
edits are what you want to share with other users of the dataset.
}
\details{
As the warning message in the function states, please be absolutely sure
that you know what you're doing before you run this function. Grafting data
that was not originally cloned from the same dataset, could have
disastrous consequences for your dataset. Human sacrifice, dogs and cats
living together, mass hysteria. Proceed with caution.
}

