\name{hiddenVariables}
\alias{hiddenVariables}
\title{Show the names of hidden variables within the dataset}
\usage{
  hiddenVariables(dataset)
}
\arguments{
  \item{dataset}{the Dataset}
}
\value{
  a vector of the names of Variables marked as hidden.
}
\description{
  Show the names of hidden variables within the dataset
}

