% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide-variables.R
\name{hide}
\alias{hide}
\alias{unhide}
\alias{hide,CrunchVariable-method}
\alias{hide,VariableCatalog-method}
\alias{unhide,CrunchVariable-method}
\alias{unhide,VariableCatalog-method}
\title{Hide and Unhide Variables}
\usage{
\S4method{hide}{CrunchVariable}(x)

\S4method{hide}{VariableCatalog}(x)

\S4method{unhide}{CrunchVariable}(x)

\S4method{unhide}{VariableCatalog}(x)
}
\arguments{
\item{x}{a Variable or subset of a VariableCatalog to hide or unhide}
}
\value{
(invisibly) the Variable or VariableCatalog, hidden or unhidden
}
\description{
Hide and Unhide Variables
}
\seealso{
\code{\link{hideVariables}}
}
