% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-index-table.R
\name{index.table}
\alias{index.table}
\title{Calculate an index table for a CrunchCube}
\usage{
index.table(x, margin)
}
\arguments{
\item{x}{A CrunchCube to calculate index table for}

\item{margin}{which margin to index against (1 for rows, 2 for columns)}
}
\value{
an array of percentages indexed to the margin provided
}
\description{
Index tables are percentages of percentages. They take the percentage from
\code{prop.table()} and divide that by the proportions of the other margin.
}
\details{
\code{index.table()} is only implemented for 2 dimensional cubes.
}
