% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{interactVariables}
\alias{interactVariables}
\title{Make a variable that is the interaction of two or more other variables}
\usage{
interactVariables(..., name)
}
\arguments{
\item{...}{A sequence of categorical variables to make an interaction of}

\item{name}{a character to use as the name for the interaction variable}
}
\value{
A \code{\link{VariableDefinition}} that will create the new
interaction variable.
}
\description{
Make a variable that is the interaction of two or more other variables
}
