% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-categories.R, R/abstract-category.R,
%   R/categories.R
\docType{methods}
\name{is.na,AbsCats-method}
\alias{is.na,AbsCats-method}
\alias{is-na-categories}
\alias{is.na<-,AbsCats,character-method}
\alias{is.na<-,AbsCats,logical-method}
\alias{is.na,AbsCats-method}
\alias{is-na-categories}
\alias{is.na,AbsCat-method}
\alias{is.na<-,AbsCat,logical-method}
\alias{is-na-categories}
\title{is.na for Categories}
\usage{
\S4method{is.na}{AbsCats}(x)

\S4method{is.na}{AbsCats,character}(x) <- value

\S4method{is.na}{AbsCats,logical}(x) <- value

\S4method{is.na}{AbsCats}(x)

\S4method{is.na}{AbsCat}(x)

\S4method{is.na}{AbsCat,logical}(x) <- value
}
\arguments{
\item{x}{Categories or a single Category}

\item{value}{To change the missingness of categories, supply either:
\enumerate{
\item a logical vector of equal length of the categories (or length 1 for the
Category method); or
\item the names of the categories to mark as missing.
If supplying the latter, any categories already indicated as missing will
remain missing.
}}
}
\value{
Getters return logical, a named vector in the case of the Categories
method; setters return \code{x} duly modified.
}
\description{
Crunch categorical variables allow you to set multiple categories as missing.
For instance, you might have "not answered" and "doesn't know" both coded as
missing. This function returns a logical vector of all dataset entries that
fall into any of the missing categories. It also allows you to append
additional categories to the list of missing categories using the setter.
}
