% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R, R/multitables.R
\docType{methods}
\name{is-public}
\alias{is-public}
\alias{is.public<-}
\alias{is.public}
\alias{is.public,CrunchFilter-method}
\alias{is.public<-,CrunchFilter-method}
\alias{is.public,MultitableCatalog-method}
\alias{is.public<-,MultitableCatalog-method}
\alias{is.public,Multitable-method}
\alias{is.public<-,Multitable-method}
\title{View and modify "public" attribute}
\usage{
\S4method{is.public}{CrunchFilter}(x)

\S4method{is.public}{CrunchFilter}(x) <- value

\S4method{is.public}{MultitableCatalog}(x)

\S4method{is.public}{MultitableCatalog}(x) <- value

\S4method{is.public}{Multitable}(x)

\S4method{is.public}{Multitable}(x) <- value
}
\arguments{
\item{x}{a Crunch object}

\item{value}{an attribute to set}
}
\value{
For \code{is.public}, a logical value for whether the object is
flagged as shared with all dataset viewers. (Its setter thus takes a
logical value as well.) Catalogs of datasets return a vector of logicals
corresponding to the length of the catalog, while entities return a single value.
}
\description{
View and modify whether all dataset viewers have access to the dataset. This
will return \code{FALSE} if the dataset is in draft.
}
