% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R, R/share.R
\docType{methods}
\name{is.editor}
\alias{is.editor}
\alias{is.editor<-}
\alias{is.editor,MemberCatalog-method}
\alias{is.editor<-,MemberCatalog,logical-method}
\alias{is.editor,PermissionCatalog-method}
\alias{is.editor,PermissionTuple-method}
\title{Read and set edit privileges}
\usage{
\S4method{is.editor}{MemberCatalog}(x)

\S4method{is.editor}{MemberCatalog,logical}(x) <- value

\S4method{is.editor}{PermissionCatalog}(x)

\S4method{is.editor}{PermissionTuple}(x)
}
\arguments{
\item{x}{PermissionCatalog or MemberCatalog}

\item{value}{For the setter, logical: should the indicated users be allowed
to edit the associated object?}
}
\value{
\code{is.editor} returns a logical vector corresponding to whether
the users in the catalog can edit or not. \code{is.editor<-} returns the
catalog, modified.
}
\description{
Read and set edit privileges
}
