% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{is.TRUEorFALSE}
\alias{is.TRUEorFALSE}
\title{Check that a value is TRUE or FALSE}
\usage{
is.TRUEorFALSE(value)
}
\arguments{
\item{value}{Value to check}
}
\value{
\code{TRUE} if \code{value} is either \code{TRUE} or \code{FALSE}, \code{FALSE} otherwise
}
\description{
Check that a value is TRUE or FALSE
}
\keyword{internal}
