% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{isTRUEorFALSE}
\alias{isTRUEorFALSE}
\title{Check that a value is TRUE or FALSE}
\usage{
isTRUEorFALSE(value)
}
\arguments{
\item{value}{Value to check}
}
\value{
nothing, called for its side effects
}
\description{
Check that a value is TRUE or FALSE
}
\keyword{internal}
