% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-info.R
\name{lastAccessed}
\alias{lastAccessed}
\title{Get information about when users last accessed a dataset}
\usage{
lastAccessed(x, user_fields = c("name", "email", "access_time"))
}
\arguments{
\item{x}{a dataset object or a project object (e.g. from \code{\link[=projects]{projects()}})}

\item{user_fields}{metadata fields about users to include in the dataframe
(default: \code{c("name", "email", "access_time")})}
}
\value{
a dataframe with information about all users and when they last
accessed a particular dataset or all of the datasets in a project
}
\description{
Gets dataset-user interaction metadata about all of the users that have
access to a particular dataset or all of the datasets in a project. Returns
a data.frame to easily manipulate, subset, and transform details about users.
}
