% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\docType{methods}
\name{length}
\alias{length}
\alias{length,CrunchDeck-method}
\alias{length,ShojiCatalog-method}
\alias{length,ShojiOrder-method}
\alias{length,OrderGroup-method}
\alias{length,TabBookResult-method}
\alias{length,MultitableResult-method}
\title{Length of Crunch objects}
\usage{
\S4method{length}{CrunchDeck}(x)

\S4method{length}{ShojiCatalog}(x)

\S4method{length}{ShojiOrder}(x)

\S4method{length}{OrderGroup}(x)

\S4method{length}{TabBookResult}(x)

\S4method{length}{MultitableResult}(x)
}
\arguments{
\item{x}{a Crunch object (catalog, order, etc.)}
}
\value{
Integer: the number of elements in the catalog, group, folder, etc.
}
\description{
Length of Crunch objects
}
\keyword{internal}
