% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{listDatasetGadget}
\alias{listDatasetGadget}
\title{Open dataset selector}
\usage{
listDatasetGadget(kind = c("active", "all", "archived"), refresh = FALSE,
  autoclose = FALSE)
}
\arguments{
\item{kind}{character specifying whether to look in active, archived, or all
datasets. Default is "active", i.e. non-archived.}

\item{refresh}{logical: should the function check the Crunch API for new
datasets? Default is FALSE.}
}
\value{
A \code{loadDataset()} call is pasted into your RStudio session`
}
\description{
Open dataset selector
}
