% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-decipher.R
\name{listDecipherSurveys}
\alias{listDecipherSurveys}
\title{List Decipher surveys available}
\usage{
listDecipherSurveys()
}
\value{
ShojiCatalog of surveys available
}
\description{
List Decipher surveys available
}
