\name{loadJSONMocks}
\alias{loadJSONMocks}
\title{Load API sample objects for testing}
\usage{
  loadJSONMocks(filename)
}
\arguments{
  \item{filename}{character the name of the file. Current
  options are "dataset.json", "variables.json",
  "summaries.json".}
}
\value{
  The deserialized contents of the JSON file, typically a
  list.
}
\description{
  Load API sample objects for testing
}

