% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edit-lock.R
\name{lock}
\alias{lock}
\alias{unlock}
\title{Lock and unlock a dataset for editing}
\usage{
lock(dataset)

unlock(dataset)
}
\arguments{
\item{dataset}{a \code{CrunchDataset}}
}
\value{
\code{dataset}, invisibly, after having set the current editor.
}
\description{
Crunch allows a single active editor. If you have edit privileges but are
not currently editing the dataset, you must unlock the dataset before
making changes. You may then lock the dataset when you're done editing.
}

