% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{logout}
\alias{logout}
\alias{login}
\alias{resetPassword}
\title{DEPRECATED! Authenticate with the Crunch API}
\usage{
logout()

login(...)

resetPassword(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
A deprecated method to authenticate to the crunch.io API. See \code{\link{crunch-api-key}}
for the currently supported method, as \code{login()}, \code{logout()} and \code{resetPassword}()
no longer work.
}
