% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-variables.R
\name{makeCaseVariable}
\alias{makeCaseVariable}
\title{Make a case variable}
\usage{
makeCaseVariable(cases, ...)
}
\arguments{
\item{cases}{A list of objects of class \code{\link{Case}}}

\item{...}{other properties to pass about the case variable (e.g. name)}
}
\description{
Make a case variable
}
