% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-variable.R
\name{makeFillVariable}
\alias{makeFillVariable}
\alias{fillExpr}
\title{Fill a variable from another}
\usage{
makeFillVariable(x, fills, ..., data = NULL, name)

fillExpr(x, fills, ...)
}
\arguments{
\item{x}{a Categorical (Array) variable}

\item{fills}{a list of lists that each have a "fill" item that is a variable or expression
as well as one of an "id", "name", or "value" that will be matched to the categories of x
If x is an expression, the fills must be an id}

\item{...}{A sequence of named expressions or variables to use as fills, where the name will be
matched to the existing categories}

\item{data}{(optional) a crunch dataset to use. Specifying this means you don't have to put
\verb{dataset$} in front of each variable name}

\item{name}{A character to use as the name of the variable to create}
}
\value{
A \code{\link{VariableDefinition}} that will create the new fill variable when assigned into
the Dataset.
}
\description{
Given a categorical variable, assign one or more categories to be filled in by
another existing variable or crunch expression. (Categories not filled in will
remain unchanged).
}
\examples{
\dontrun{
makeFillVariable(
    ds$v1,
    fills = list(
        list(fill = ds$v2, name = "dog")
    ),
    name = "new fill"
)
makeFillVariable(v1, "dog" = ds$v2, name = "new fill")

# the dataset can be specified with data=
makeFillVariable(v1, "dog" = v2, data = ds, name = "new fill")
}
}
