% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{makeInteractions}
\alias{makeInteractions}
\title{Make case statments that define the interaction of two or more other variables}
\usage{
makeInteractions(..., drop = FALSE, sep = ":")
}
\arguments{
\item{...}{A sequence of categorical variables to make an interaction of}

\item{drop}{should unused levels be dropped? (currently not implemented, so \code{FALSE})}

\item{sep}{a character to separate the values of the individual variables (default: \code{:})}
}
\value{
A list of cases that can be passed to \code{\link{makeCaseVariable}} to make
the interaction variable.
}
\description{
Make case statments that define the interaction of two or more other variables
}
