% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-result.R
\name{makeMarginMap}
\alias{makeMarginMap}
\title{Make a map of margins}
\usage{
makeMarginMap(dimTypes)
}
\arguments{
\item{dimTypes}{dimension types from \code{getDimTypes()}}
}
\value{
a vector of margins, the length of which is the length of the real
cube, the values are the user cube dimensions
}
\description{
Useful when converting to and from user and real cube dimension indexes.
}
\examples{

\dontrun{
makeMarginMap(getDimTypes(cat_by_cat_cube))
# 1 2

makeMarginMap(getDimTypes(MR_by_cat_cube))
# 1 1 2

makeMarginMap(getDimTypes(cat_by_MR_cube))
# 1 2 2

makeMarginMap(getDimTypes(MR_by_MR_cube))
# 1 1 2 2
}

}
\keyword{internal}
