% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make-mr-from-cat.R
\name{makeMRfromCat}
\alias{makeMRfromCat}
\title{Combine categorical variables into an MR}
\usage{
makeMRfromCat(var1, var2, ..., useNAcats = FALSE)
}
\arguments{
\item{variable}{1 to combine}

\item{variable}{2 to combine}

\item{other}{desired attributes such as new variable alias, name,
description as well as any other variables you would like to combine}

\item{a}{boolean describing whether categories that are NA in the
original variables should be made into subvariables in the new
variable}
}
\value{
a new multiple_response variable with subvariables
that are the union of the categories of variables given
}
\description{
This function lets you combine categorical variables into an MR variable
Arguments include at least two categorical variables as well as any metadata
for the new variable that is being created. The final argument useNAcats
is given the defult value FALSE. If it is made TRUE, any categories that
are NA in the categorical variables will become subvariables.
}

