% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{matchCat2Feat}
\alias{matchCat2Feat}
\title{Match categories with features from geodata}
\usage{
matchCat2Feat(categories, all_features = availableFeatures())
}
\arguments{
\item{categories}{a vector of categories to match}

\item{all_features}{a dataframe of all available geodata features. (default:
downloaded from Crunch servers)}
}
\value{
geodatum to associate with the variable that produced categories
}
\description{
Match categories with features from geodata
}
