% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\docType{methods}
\name{members}
\alias{members}
\alias{members<-}
\alias{permissions}
\alias{members,ProjectFolder-method}
\alias{members<-,ProjectFolder,MemberCatalog-method}
\alias{members,CrunchTeam-method}
\alias{members<-,ProjectFolder,character-method}
\alias{members<-,CrunchTeam,MemberCatalog-method}
\alias{members<-,CrunchTeam,character-method}
\title{Manage access to datasets and other objects}
\usage{
members(x)

members(x) <- value

permissions(x)

\S4method{members}{ProjectFolder}(x)

\S4method{members}{ProjectFolder,MemberCatalog}(x) <- value

\S4method{members}{CrunchTeam}(x)

\S4method{members}{ProjectFolder,character}(x) <- value

\S4method{members}{CrunchTeam,MemberCatalog}(x) <- value

\S4method{members}{CrunchTeam,character}(x) <- value
}
\arguments{
\item{x}{\code{CrunchDataset}, \code{ProjectFolder}, or \code{CrunchTeam}}

\item{value}{for \code{members<-}, a character vector of emails or URLs of
users to add to the team.}
}
\value{
\code{members()} returns a
\code{MemberCatalog}, which has references to the users that are members
of the team. \code{members<-} returns \code{x} with the given users added
to the members catalog. \code{permissions()} returns a \code{PermissionCatalog} with
similar semantics.
}
\description{
These methods allow you to work with teams.
}
\seealso{
\code{\link[=users]{users()}}
}
