% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame.R
\name{merge.CrunchDataFrame}
\alias{merge.CrunchDataFrame}
\title{Merge a CrunchDataFrame}
\usage{
\method{merge}{CrunchDataFrame}(x, y, by.x, by.y, sort = "x", ...)
}
\arguments{
\item{x}{a CrunchDataFrame}

\item{y}{a standard data.frame}

\item{by.x}{name of the variable to match}

\item{by.y}{name of the variable to match}

\item{sort}{character, either "x" or "y" (default: "x"). Which of the inputs should be used for the output order. Unlike merge.data.frame, merge.CrunchDataFrame will not re-sort the order of the output. It will use the order of either x or y.}

\item{...}{ignored for now}
}
\description{
Merge a CrunchDataFrame
}
