% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{miscexpr}
\alias{miscexpr}
\alias{charLength}
\alias{unmissing}
\alias{normalize}
\title{miscellaneous expressions for crunch database}
\usage{
charLength(x)

unmissing(x)

normalize(x)
}
\arguments{
\item{x}{a CrunchVariable or CrunchExpr}
}
\value{
a CrunchExpr
}
\description{
miscellaneous expressions for crunch database
}
