% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{moveLastElement}
\alias{moveLastElement}
\title{Move and delete last element of a vector
This moves the last element of a vector to an index, then deletes the last
element, it is broken out for testing purposes}
\usage{
moveLastElement(v, idx)
}
\arguments{
\item{v}{a vector}

\item{idx}{The index to move the last element to.}
}
\value{
a vector
}
\description{
Move and delete last element of a vector
This moves the last element of a vector to an index, then deletes the last
element, it is broken out for testing purposes
}
\keyword{internal}
