% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{moveToAfter}
\alias{moveToAfter}
\title{Move a variable to after another variable}
\usage{
moveToAfter(x, value, after)
}
\arguments{
\item{x}{the variable group that \code{after} is in}

\item{value}{the variable or dataset subset you would like to move}

\item{after}{the variable you want to precede \code{value}}
}
\value{
returns a variable group
}
\description{
Move a variable to after another variable
}
\examples{
\dontrun{
ordering(ds)[['Demographics']] <- moveToAfter(ordering(ds)[['Grp A']], ds['age'], ds$educ)
}
}
