% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-array.R
\name{mrFromDelim}
\alias{mrFromDelim}
\title{Create Multiple Response Variable from Delimited}
\usage{
mrFromDelim(var, delim, name, selected = "selected",
  not_selected = "not_selected", unanswered = NA, ...)
}
\arguments{
\item{var}{The variable containing the delimited responses}

\item{delim}{The delimiter separating the responses}

\item{name}{The name of the resulting MR variable}

\item{selected}{A character string used to indicate a selection}

\item{not_selected}{Character string identifying non-selection}

\item{unanswered}{Character string indicating non-response}

\item{...}{Other arguments to be passed on to \code{\link[=makeMR]{makeMR()}}}
}
\description{
Surveys often record multiple response questions in delimited lists where
each respondent's selections are separated by a delimiter like \code{;} or \code{|}.
This function breaks the delimited responses into subvariables, uploads those
subvariables to Crunch, and finally creates a multiple response variable from
them.
}
