% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/multitables.R
\name{multitables}
\alias{multitables}
\alias{multitables<-}
\alias{multitable-catalog}
\alias{multitables,CrunchDataset-method}
\alias{multitables<-,CrunchDataset-method}
\title{Multitable entities for a dataset}
\usage{
multitables(x)

multitables(x) <- value

\S4method{multitables}{CrunchDataset}(x)

\S4method{multitables}{CrunchDataset}(x) <- value
}
\arguments{
\item{x}{a CrunchDataset}

\item{value}{for the assignment method, a MultitableCatalog}
}
\value{
an object of class MultitableCatalog containing references to
Multitable entities. (Setter returns the Dataset.)
}
\description{
Multitable entities for a dataset
}
