% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/new-dataset.R
\name{newDatasetByColumn}
\alias{newDatasetByColumn}
\title{Upload a data.frame column-by-column to make a new dataset}
\usage{
newDatasetByColumn(x, name = substitute(x), useAlias = default.useAlias(),
  ...)
}
\arguments{
\item{x}{a data.frame or other rectangular R object}

\item{name}{character, the name to give the new Crunch dataset. Default is
the name of the R object passed in \code{x}}

\item{useAlias}{logical whether variable alias or name should be used as R
variable names when the dataset is returned. Default is TRUE, meaning alias.}

\item{...}{additional arguments passed to \code{ \link{createDataset}}}
}
\value{
If successful, an object of class CrunchDataset.
}
\description{
Use this version if you have lots of variables, under 1M rows, perhaps
backed by ff or other memory-mapped files, and time to kill.
}
\seealso{
\code{\link{newDataset}} \code{\link{newDatasetByCSV}}
}

