% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{newDeck}
\alias{newDeck}
\title{Create an empty Crunch Deck}
\usage{
newDeck(dataset, name, ...)
}
\arguments{
\item{dataset}{A Crunch Dataset}

\item{name}{The name of the Deck}

\item{...}{Further attributes of the deck such as the description, see API
docs for options.}
}
\description{
Create an empty Crunch Deck
}
