% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-dataset.R
\name{newExampleDataset}
\alias{newExampleDataset}
\title{Import a fixture dataset for testing}
\usage{
newExampleDataset(name = "pets", project = defaultCrunchProject())
}
\arguments{
\item{name}{string name of the fixture dataset. Currently "pets" is the only
one available.}

\item{project}{A \code{ProjectFolder} object, string path that could be passed to \code{\link[=cd]{cd()}}
relative to the root project, or a URL for a \code{ProjectFolder}. If left empty,
rcrunch will look in \code{envOrOption('crunch.default.project')} and error if nothing
is found.}
}
\value{
A new \code{CrunchDataset} entity.
}
\description{
The \code{crunch} package includes some data for you to explore the features of
the platform. Use this function to upload one to create a demo dataset.
}
