% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-transforms.R
\name{noTransforms}
\alias{noTransforms}
\title{Remove transformations from a cube}
\usage{
noTransforms(cube)
}
\arguments{
\item{cube}{a CrunchCube}
}
\value{
the cube with no transformations
}
\description{
This is useful if you don't want to see or use any transformations like Subtotals and Headings
}
