% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{ordering}
\alias{ordering}
\alias{ordering<-}
\alias{ordering,CrunchDataset-method}
\alias{ordering<-,CrunchDataset-method}
\alias{ordering,VariableCatalog-method}
\alias{ordering<-,VariableCatalog-method}
\alias{ordering,DatasetCatalog-method}
\alias{ordering,ProjectFolder-method}
\alias{ordering<-,DatasetCatalog-method}
\alias{ordering<-,ProjectFolder-method}
\title{Get and set VariableOrder}
\usage{
ordering(x)

ordering(x) <- value

\S4method{ordering}{CrunchDataset}(x)

\S4method{ordering}{CrunchDataset}(x) <- value

\S4method{ordering}{VariableCatalog}(x)

\S4method{ordering}{VariableCatalog}(x) <- value

\S4method{ordering}{DatasetCatalog}(x)

\S4method{ordering}{ProjectFolder}(x)

\S4method{ordering}{DatasetCatalog}(x) <- value

\S4method{ordering}{ProjectFolder}(x) <- value
}
\arguments{
\item{x}{a VariableCatalog or CrunchDataset}

\item{value}{a valid VariableOrder object}
}
\value{
\code{ordering} returns a VariableOrder object, while
\verb{ordering<-} sets the VariableOrder
}
\description{
The \code{ordering} methods allow you to get a \code{\link{VariableOrder}} on a
\code{\link{CrunchDataset}} or on the \code{\link{VariableCatalog}} that the dataset contains.
}
\details{
Crunch datasets work with folders, and the ordering is deprecated. It is no
longer possible to set the ordering of a variable catalog from rcrunch.
}
\keyword{internal}
