% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-catalog.R
\name{owners}
\alias{ownerNames}
\alias{owners}
\title{See who owns these datasets}
\usage{
owners(x)

ownerNames(x)
}
\arguments{
\item{x}{DatasetCatalog}
}
\value{
For \code{owners}, the URLs of the users or projects that own
these datasets. For \code{ownerNames}, their names.
}
\description{
See who owns these datasets
}

