% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zed-scores.R
\name{p.values}
\alias{p.values}
\title{Calculate p-values for a CrunchCube}
\usage{
p.values(table, margin)
}
\arguments{
\item{table}{A CrunchCube to calculate z-scores for}

\item{margin}{which margin to test against (1 for rows, 2 for columns)}
}
\value{
an array of p-values
}
\description{
Calculate p-values for a CrunchCube
}
