% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{packageHasFunction}
\alias{packageHasFunction}
\title{Check if a package has a function installed}
\usage{
packageHasFunction(fun, pkg)
}
\arguments{
\item{fun}{Character string of the function to search for}

\item{pkg}{The package name}
}
\value{
Logical
}
\description{
This is useful for handling if the user has the correct version of a suggested
package installed.
}
\keyword{internal}
