% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/palettes.R
\name{palettes}
\alias{palettes}
\alias{defaultPalette}
\alias{palettes,CrunchDataset-method}
\alias{defaultPalette,CrunchDataset-method}
\alias{defaultPalette,AnalyticPalettes-method}
\title{Get the palettes from a crunch object}
\usage{
palettes(x)

defaultPalette(x, ...)

\S4method{palettes}{CrunchDataset}(x)

\S4method{defaultPalette}{CrunchDataset}(x, ...)

\S4method{defaultPalette}{AnalyticPalettes}(x, ...)
}
\arguments{
\item{x}{A crunch object, like a \code{CrunchDataset}}

\item{...}{ignored (reserved for future expansion)}
}
\description{
\code{CrunchDataset}s have color palettes associated with them
that can be used as default colors for dashboard tiles. One of
them can be assigned the "default".
}
