% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-stream.R
\name{pendingStream}
\alias{pendingStream}
\title{Get the pending streams for a dataset}
\usage{
pendingStream(ds)
}
\arguments{
\item{ds}{a CrunchDataset}
}
\value{
number of pending messages in the stream for the dataset
}
\description{
Crunch allows you to stream data to a dataset. Streaming data is useful for
datasets which have frequent updates (see
\href{http://docs.crunch.io/#streaming-rows}{the Crunch documentation} for more)
information. \code{pendingMessages()} retrieves the number of pending messages and \code{appendStreamedRows()} appends all pending streamed rows to the dataset.
}
