% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\docType{methods}
\name{permissions}
\alias{permissions}
\alias{permissions,CrunchDataset-method}
\title{See who has access to a dataset}
\usage{
\S4method{permissions}{CrunchDataset}(x)
}
\arguments{
\item{x}{CrunchDataset}
}
\value{
A PermissionCatalog containing information on the users and teams
that have access to this dataset.
}
\description{
See who has access to a dataset
}
