% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{pk}
\alias{pk}
\alias{pk<-}
\alias{pk<-,CrunchDataset-method}
\title{Name, alias, and description for Crunch objects}
\usage{
\S4method{pk}{CrunchDataset}(x)

\S4method{pk}{CrunchDataset}(x) <- value
}
\arguments{
\item{x}{a Dataset}

\item{value}{For the setters, a single Variable to use as the primary key.}
}
\value{
Getters return the Variable object that is used as the primary key (\code{NULL} if there is no primary key); setters
return \code{x} duly modified.
}
\description{
Name, alias, and description for Crunch objects
}
