% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pointTo}
\alias{pointTo}
\title{Change which server to point to}
\usage{
pointTo(place, api_url = "https://app.crunch.io/api/",
  use_env_api_url = FALSE)
}
\arguments{
\item{place}{a place to point to (either 'local' or 'prod')}

\item{api_url}{a url to use by default: "https://app.crunch.io/api/"}

\item{use_env_api_url}{should the api url from the environment variables be
used? (default: \code{FALSE}, this will always be \code{FALSE} when chosing \code{prod = 'local'})}
}
\value{
nothing
}
\description{
A convenience function for changing where you want the Crunch package to try
to connect to. You can select either:
}
\details{
\itemize{
\item 'local' to connect to a local (vagrant-based) backend. This will use the
\code{test.user}, \code{test.pw}, and \code{test.api} options from your \code{.Rprofile} or
corresponding environemnt variables.
\item 'prod' to connect to a hosted backend. This will use the \code{crunch.email},
\code{crunch.pw}, and \code{crunch.api} options from your \code{.Rprofile} or corresponding
environemnt variables.
}
}
\keyword{internal}
