% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{postUpload,ANY-method}
\alias{postUpload,ANY-method}
\title{Functions that modify remote Crunch variables after uploading CSV so that
all metadata R knows is translated to Crunch.}
\usage{
\S4method{postUpload}{ANY}(source.var, crunch.var)
}
\description{
Functions that modify remote Crunch variables after uploading CSV so that
all metadata R knows is translated to Crunch.
}

