\name{postUpload}
\alias{postUpload}
\title{Functions that modify remote Crunch variables after uploading CSV so that
all metadata R knows is translated to Crunch.}
\description{
  Functions that modify remote Crunch variables after
  uploading CSV so that all metadata R knows is translated
  to Crunch.
}

