% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchbox.R
\name{preCrunchBoxCheck}
\alias{preCrunchBoxCheck}
\title{Summarize any characteristics of a dataset that could make for an awkward CrunchBox}
\usage{
preCrunchBoxCheck(dataset)
}
\arguments{
\item{dataset}{CrunchDataset, potentially subsetted on variables}
}
\value{
Invisbly, the dataset. Called for side-effect of printing things.
}
\description{
Summarize any characteristics of a dataset that could make for an awkward CrunchBox
}
\seealso{
\code{\link{crunchBox}}
}

