% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{preEmbedCheck}
\alias{preEmbedCheck}
\title{Summarize any characteristics of a dataset that could make for an awkward widget}
\usage{
preEmbedCheck(dataset)
}
\arguments{
\item{dataset}{CrunchDataset, potentially subsetted on variables}
}
\value{
Invisbly, the dataset. Called for side-effect of printing things.
}
\description{
Summarize any characteristics of a dataset that could make for an awkward widget
}

