\name{prepareBindInputs}
\alias{prepareBindInputs}
\title{Given inputs to makeArray/makeMR, parse and validate}
\usage{
  prepareBindInputs(list_of_variables = NULL,
    dataset = NULL, pattern = NULL, key = namekey(dataset),
    ...)
}
\arguments{
  \item{...}{Stuff from calling function that will be
  ignored.}
}
\description{
  Given inputs to makeArray/makeMR, parse and validate
}

