% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{preUpload,data.frame-method}
\alias{preUpload,data.frame-method}
\title{Methods that prepare a data.frame to be written out as CSV and uploaded}
\usage{
\S4method{preUpload}{data.frame}(x)
}
\description{
Methods that prepare a data.frame to be written out as CSV and uploaded
}

