% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-variables.R
\name{privateVariables}
\alias{privateVariables}
\title{Show the names of a dataset's private variables}
\usage{
privateVariables(dataset, key = namekey(dataset))
}
\arguments{
\item{dataset}{the Dataset}

\item{key}{the Variable attribute to return. Default is "alias", following
\code{getOption("crunch.namekey.dataset")}.}
}
\value{
a vector of the names of Variables marked as private.
}
\description{
Show the names of a dataset's private variables
}
