% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-variables.R
\name{privatize}
\alias{privatize}
\alias{deprivatize}
\alias{privatise}
\alias{deprivatise}
\alias{privatize,CrunchVariable-method}
\alias{privatize,VariableCatalog-method}
\alias{deprivatize,CrunchVariable-method}
\alias{deprivatize,VariableCatalog-method}
\alias{privatise,CrunchVariable-method}
\alias{privatise,VariableCatalog-method}
\alias{deprivatise,CrunchVariable-method}
\alias{deprivatise,VariableCatalog-method}
\title{Make Variables Private or Not Private}
\usage{
\S4method{privatize}{CrunchVariable}(x)

\S4method{privatize}{VariableCatalog}(x)

\S4method{deprivatize}{CrunchVariable}(x)

\S4method{deprivatize}{VariableCatalog}(x)

\S4method{privatise}{CrunchVariable}(x)

\S4method{privatise}{VariableCatalog}(x)

\S4method{deprivatise}{CrunchVariable}(x)

\S4method{deprivatise}{VariableCatalog}(x)
}
\arguments{
\item{x}{a Variable or subset of a VariableCatalog to move in/out of the private folder}
}
\value{
(invisibly) the Variable or VariableCatalog with the changes
}
\description{
Make Variables Private or Not Private
}
\seealso{
\code{\link{privatizeVariables}}
}
