% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-variables.R
\name{privatizeVariables}
\alias{privatizeVariables}
\alias{privatiseVariables}
\alias{privateVariables<-}
\alias{deprivatizeVariables}
\alias{deprivatiseVariables}
\title{Make variables private (or not) within a dataset}
\usage{
privatizeVariables(dataset, variables)

privatiseVariables(dataset, variables)

privateVariables(x) <- value

deprivatizeVariables(dataset, variables)

deprivatiseVariables(dataset, variables)
}
\arguments{
\item{dataset}{the Dataset to modify}

\item{variables}{names or indices of variables to (de)privatize}

\item{x}{\code{dataset}, for \verb{privateVariables<-}}

\item{value}{\code{variables}, for \verb{privateVariables<-}}
}
\value{
(invisibly) \code{dataset} with the specified variables (de)privatized
}
\description{
Make variables private (or not) within a dataset
}
\seealso{
\code{\link{privatize}}
}
