% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{project-datasets}
\alias{datasets}
\alias{datasets<-}
\alias{project-datasets}
\title{A project's datasets}
\usage{
datasets(x)

datasets(x) <- value
}
\arguments{
\item{x}{a \code{CrunchProject}}

\item{value}{\code{CrunchDataset} for the setter}
}
\value{
An object of class \code{DatasetCatalog}. The setter returns the
project with the given dataset added to it (via changing its owner to be
the specified project, \code{x}).
}
\description{
A project's datasets
}

