% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{project-icon}
\alias{project-icon}
\alias{icon}
\alias{icon<-}
\title{Get or set a project's icon}
\usage{
icon(x)

icon(x) <- value
}
\arguments{
\item{x}{a \code{CrunchProject}}

\item{value}{charcter file path of the icon image file to set}
}
\value{
The URL of the project's icon. The setter returns the
project after having uploaded the specified file as the new icon.
}
\description{
This function allows you to change a project's icon. Note that you
must first upload the icon to a website, as this function does not
allow you to upload a local file.
}
