% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects}
\alias{projects}
\title{List project folders}
\usage{
projects(x = getAPIRoot())
}
\arguments{
\item{x}{a \code{ShojiObject} that has an associated catalog. If omitted,
the default value for \code{x} means that you will load the user's primary
folder. (Currently, there are no other folders to load.)}
}
\value{
An object of class \code{ProjectFolder}.
}
\description{
List project folders
}
\examples{
\dontrun{
myprojects <- projects()
proj <- myprojects[["Project name"]]
}
}
