% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-access.R
\name{projectUserInfo}
\alias{projectUserInfo}
\title{Get information about when users last any of the datasets in a project}
\usage{
projectUserInfo(project, ...)
}
\arguments{
\item{project}{a character of the project name to gather metadata from}

\item{...}{additional arguments passed to \code{\link[=userInfo]{userInfo()}} (e.g. user_fields)}
}
\value{
a dataframe with information about all users for a particular dataset
}
\description{
Useful to get dataset-user interaction metadata about all of the users that have
access to all of the datasets in a project.
}
