% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcrunch.R
\docType{package}
\name{rcrunch}
\alias{rcrunch}
\alias{rcrunch-package}
\title{Crunch.io: instant, visual, collaborative data analysis}
\description{
\href{http://crunch.io/}{Crunch.io} provides a cloud-based data store and
analytic engine. It has a \href{https://beta.crunch.io/}{web client} for
interactive data exploration and visualization. The rcrunch package for R
allows analysts to interact with and manipulate Crunch datasets from within
R. Importantly, this allows technical researchers to collaborate naturally
with team members, managers, and clients who prefer a point-and-click
interface: because all connect to the same dataset in the cloud, there is no
need to email files back and forth continually to share results.
}
\seealso{
To learn more about using the package, see
\code{vignette("getting-started", package="rcrunch")}. To sign up for a
Crunch.io account, visit \url{https://beta.crunch.io/}.
}

