% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-order.R
\name{removeEmptyGroups}
\alias{removeEmptyGroups}
\title{Remove OrderGroups with no entities}
\usage{
removeEmptyGroups(x)
}
\arguments{
\item{x}{VariableOrder, DatasetOrder, VariableGroup, or DatasetGroup}
}
\value{
\code{x} with empty groups removed.
}
\description{
This function recurses through a \code{ShojiOrder}/\code{OrderGroup} and
removes any groups that contain no entities.
}

