% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{reorderSlides}
\alias{reorderSlides}
\title{Reorder slides in a crunchdeck}
\usage{
reorderSlides(x, order)
}
\arguments{
\item{x}{A SlideCatalog}

\item{order}{The numeric order for slides to be reordered to.}
}
\value{
A SlideCatalog
}
\description{
Reorder slides in a crunchdeck
}
