% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-subset.R
\name{replaceCharWithNumeric}
\alias{replaceCharWithNumeric}
\title{Transform character vectors into indices}
\usage{
replaceCharWithNumeric(cat_names, idx, visible = TRUE)
}
\arguments{
\item{cat_names}{The category names for a given dimension}

\item{idx}{The index}

\item{visible}{whether or not a category is visible to the user}
}
\description{
When a user supplies character strings to subset a cube, we need to translate it
into a numeric index to subset the cube. This function does that as well as
some error checking. It is broken out for testing.
}
\keyword{internal}
