% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-result.R
\name{replaceMissingWithTRUE}
\alias{replaceMissingWithTRUE}
\title{Replace missing elements with TRUE}
\usage{
replaceMissingWithTRUE(l)
}
\arguments{
\item{l}{a list}
}
\value{
a list
}
\description{
When subsetting missingness stands in for selecting all of the elements of that
dimension. This gets tricky when you are selecting n dimensions and so need to
capture the dimensions with \code{...}.  This function checks if an element of a list
is missing and replaces that element with \code{TRUE} in order to handle this case.
}
\keyword{internal}
