% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{resetPassword}
\alias{resetPassword}
\title{Reset your password}
\usage{
resetPassword(email)
}
\arguments{
\item{email}{Your email}
}
\value{
NULL, invisibly. Called for its side effects.
}
\description{
Trigger the password reset process. Password reset
instructions will be emailed to you.
}
\examples{
\dontrun{
resetPassword("me@example.com")
}
}
