% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{retry}
\alias{retry}
\title{Retry}
\usage{
retry(
  expr,
  wait = getOption("crunch_retry_wait", default = 0.1),
  max.tries = 10
)
}
\arguments{
\item{expr}{An expression}

\item{wait}{The time in seconds to wait before retrying the expression. Defaults to 0.1.}

\item{max.tries}{The number of times to retry the expression}
}
\description{
Retry an expression. This is useful for situations where a web resource is not yet available.
You can set \code{options("crunch_retry_wait" = X)} some number larger than the default 0.1 in
your script if you are working with large exports.
}
