% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise.R
\name{rowCount}
\alias{rowCount}
\alias{rowAny}
\alias{rowAll}
\alias{rowAnyNA}
\title{Create variables based on row-wise functions for crunch Multiple Response Variables}
\usage{
rowCount(x, ...)

rowAny(x, ...)

rowAll(x, ...)

rowAnyNA(x, ...)
}
\arguments{
\item{x}{A crunch variable or expression}

\item{...}{name, description, alias, and other metadata passed to \code{\link[=VarDef]{VarDef()}}}
}
\value{
A Variable Definition
}
\description{
Quickly generate new variables that are based on row-wise summaries of Multiple Response
Variables.
}
\seealso{
\code{\link[=expresions]{expresions()}} for the more flexible expressions that power
these functions and \code{\link[=rowDistinct]{rowDistinct()}} for other row-wise functions
}
