% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-residuals.R
\docType{methods}
\name{rstandard,CrunchCube-method}
\alias{rstandard,CrunchCube-method}
\title{Calculate standardized residuals from a CrunchCube}
\usage{
\S4method{rstandard}{CrunchCube}(model)
}
\arguments{
\item{table}{A CrunchCube to calculate z-scores for}
}
\value{
an array of z-scores for each cell
}
\description{
To get a z-score for the cells of a cube by comparing to the proportions on
the row or column margins. When testing against row margins \code{z.table}
compares each cell against the proportion for all columns in that row. When
testing against column margins \code{z.table} compares each cell against the
proportion for all rows in that  column.
}
