% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{scoreCat2Feat}
\alias{scoreCat2Feat}
\title{Score similarity between a feature dataframe and categories}
\usage{
scoreCat2Feat(features, categories)
}
\arguments{
\item{features}{a vector of features to match (usually from a subset of the
output \code{[availableFeatures]}) with a single property for a single geodatum.}

\item{categories}{a vector of categories to match}
}
\value{
the Jaccard index for the values of the property given in
feat_df and the vector of categories
}
\description{
Implemented using the Jaccard index, where a number closer to 1 is
more similar.
}
