% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/automation.R
\name{scripts}
\alias{scripts}
\alias{script-catalog}
\alias{scripts,CrunchDataset-method}
\title{Crunch Automation scripts entities for a dataset}
\usage{
scripts(x)

\S4method{scripts}{CrunchDataset}(x)
}
\arguments{
\item{x}{a CrunchDataset}
}
\value{
an object of class "ScriptCatalog" containing references to
Script entities.
}
\description{
Crunch Automation scripts entities for a dataset
}
\seealso{
\code{\link[=runCrunchAutomation]{runCrunchAutomation()}} & \code{\link{automation-undo}}
}
