% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation.R
\name{scriptCheckpointVersion}
\alias{scriptCheckpointVersion}
\title{Get the version saved prior to running a crunch automation script}
\usage{
scriptCheckpointVersion(x)
}
\arguments{
\item{x}{A \code{Script} object}
}
\value{
A version list object that can be used in \code{\link[=restoreVersion]{restoreVersion()}}
}
\description{
Get the version associated with the dataset right before a crunch
automation script was run.
}
