% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{searchDatasets}
\alias{searchDatasets}
\title{Search Crunch for datasets.}
\usage{
searchDatasets(query, ...)
}
\arguments{
\item{query}{the text to search for in datasets and their variables}

\item{...}{additional options provided to the search endpoint.}
}
\value{
If successful, an object of class SearchResults
}
\description{
Search Crunch for datasets.
}
