% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure-variables.R
\name{secure}
\alias{secure}
\alias{unsecure}
\alias{secure,CrunchVariable-method}
\alias{secure,VariableCatalog-method}
\alias{unsecure,CrunchVariable-method}
\alias{unsecure,VariableCatalog-method}
\title{Secure and Unsecure Variables}
\usage{
\S4method{secure}{CrunchVariable}(x)

\S4method{secure}{VariableCatalog}(x)

\S4method{unsecure}{CrunchVariable}(x)

\S4method{unsecure}{VariableCatalog}(x)
}
\arguments{
\item{x}{a Variable or subset of a VariableCatalog to secure or move out of the secure folder}
}
\value{
(invisibly) the Variable or VariableCatalog, secured or unsecured
}
\description{
Secure and Unsecure Variables
}
\seealso{
\code{\link{secureVariables}}
}
