% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure-variables.R
\name{securedVariables}
\alias{securedVariables}
\title{Show the names of a dataset's secure variables}
\usage{
securedVariables(dataset, key = namekey(dataset))
}
\arguments{
\item{dataset}{the Dataset}

\item{key}{the Variable attribute to return. Default is "alias", following
\code{getOption("crunch.namekey.dataset")}.}
}
\value{
a vector of the names of Variables marked as secured.
}
\description{
Show the names of a dataset's secure variables
}
