% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure-variables.R
\name{secureVariables}
\alias{secureVariables}
\alias{securedVariables<-}
\alias{unsecureVariables}
\title{Secure and unsecure variables within a dataset}
\usage{
secureVariables(dataset, variables)

securedVariables(x) <- value

unsecureVariables(dataset, variables)
}
\arguments{
\item{dataset}{the Dataset to modify}

\item{variables}{names or indices of variables to (un)secure}

\item{x}{\code{dataset}, for \verb{secureVariables<-}}

\item{value}{\code{variables}, for \verb{secureVariables<-}}
}
\value{
(invisibly) \code{dataset} with the specified variables (un)secured
}
\description{
Secure and unsecure variables within a dataset
}
\seealso{
\code{\link{secure}}
}
