\name{selectHttpFunction}
\alias{selectHttpFunction}
\title{Select the right function to run that HTTP call}
\usage{
  selectHttpFunction(x)
}
\arguments{
  \item{x}{character HTTP verb name}
}
\value{
  the corresponding function from the \code{httr} package
}
\description{
  Select the right function to run that HTTP call
}

