% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{serialPaste}
\alias{serialPaste}
\title{Make a prose list}
\usage{
serialPaste(x, collapse = "and")
}
\arguments{
\item{x}{vector or list}

\item{collapse}{default="and"}
}
\description{
Function to paste together a list of items, separated by commas (if more
than 2), and with the last one having the collapse string.
}
\keyword{internal}
