% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\docType{methods}
\name{session-extract}
\alias{session-extract}
\alias{[[,Session,ANY-method}
\alias{$,Session-method}
\alias{[[<-,Session,ANY,ANY,ANY-method}
\alias{$<-,Session-method}
\title{Extract catalogs from a Session object}
\usage{
\S4method{[[}{Session,ANY}(x, i, ..., drop = FALSE)

\S4method{$}{Session}(x, name)

\S4method{[[}{Session,ANY,ANY,ANY}(x, i) <- value

\S4method{$}{Session}(x, name) <- value
}
\arguments{
\item{x}{a Session object}

\item{i}{which catalog to load. Supported values are "datasets" and "projects"}

\item{...}{additional arguments, ignored.}

\item{drop}{Invalid}

\item{name}{for \code{$}, the same as \code{i} for \code{[[}}

\item{value}{For updating, an object of the appropriate class and size to
insert. In practice \code{value} is ignored; Session objects hold no state
and it is assumed that any state modification on the server happens in other
methods.}
}
\value{
The requested catalog when extracting; a Session object if assigning.
}
\description{
Extract catalogs from a Session object
}
\seealso{
\code{\link{session}}
}
