% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{session_store}
\alias{session_store}
\title{Stay authenticated.}
\format{An environment.}
\usage{
session_store
}
\description{
The auth store keeps the session token after authentication so that all
API calls can use it. We can also store other things in it.
}
\keyword{internal}

