% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_crunch_config}
\alias{set_crunch_config}
\title{Set or modify general Crunch API request configuration}
\usage{
set_crunch_config(cfg = c(config(postredir = 3), add_headers(`user-agent`
  = crunch_user_agent())), update = FALSE)
}
\arguments{
\item{cfg}{A \code{\link[httr:config]{httr::config()}} object}
}
\value{
A list of length one containing the configuration that was set; this
function is called primarily for its side effects.
}
\description{
Set or modify general Crunch API request configuration
}
\keyword{internal}
