% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setCrunchAPI}
\alias{setCrunchAPI}
\title{Change which server to point to}
\usage{
setCrunchAPI(subdomain, port = NULL)
}
\arguments{
\item{subdomain}{the subdomain to use}

\item{port}{on optional port to use}
}
\value{
nothing
}
\description{
A convenience function for changing where you want the Crunch package to try
to connect to.
}
\examples{
setCrunchAPI("local", 8080)
setCrunchAPI("app")

}
\keyword{internal}
