% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji.R
\name{setEntitySlot}
\alias{setEntitySlot}
\title{Base setter for Crunch objects}
\usage{
setEntitySlot(x, i, value)
}
\arguments{
\item{x}{a ShojiObject or subclass thereof}

\item{i}{character the slot name to update}

\item{value}{whatever the new value of that slot should be}
}
\value{
x modified accordingly. If \code{x} isn't read-only, it will also
post the edit to the Crunch server.
}
\description{
Base setter for Crunch objects
}
\keyword{internal}
