% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji.R
\name{setEntitySlotWrapBody}
\alias{setEntitySlotWrapBody}
\title{Setter for Crunch objects that wraps in a "body"}
\usage{
setEntitySlotWrapBody(x, i, value)
}
\arguments{
\item{x}{a ShojiObject or subclass thereof}

\item{i}{character the slot name to update}

\item{value}{whatever the new value of that slot should be}
}
\value{
x modified accordingly.
}
\description{
Variable Folders require an extra list(body = ...)
compared to other places where we use \code{setEntitySlot()}
is used. This may be a bug, see pivotal:
\url{https://www.pivotaltracker.com/n/projects/2172644/stories/174429283}
}
\keyword{internal}
