% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji.R
\name{setMultiEntitySlots}
\alias{setMultiEntitySlots}
\title{setter for Crunch objects that allows a single request to update multiple parts of an entity}
\usage{
setMultiEntitySlots(x, ...)
}
\arguments{
\item{x}{a ShojiObject or subclass thereof}

\item{...}{Named arguments where the names are the slot name to update and the values
are the values that they should be updated to include.}
}
\value{
x modified accordingly.
}
\description{
setter for Crunch objects that allows a single request to update multiple parts of an entity
}
\keyword{internal}
