% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{setName}
\alias{setName}
\title{Change the name of the current folder}
\usage{
setName(object, nm)
}
\arguments{
\item{object}{A \code{VariableFolder}}

\item{nm}{A character that is the new name the folder should have}
}
\value{
\code{object}, with its name duly changed
}
\description{
If you just need to change the name of the folder you are currently in, you
can use \code{setName()} to do so. It doesn't move variables or change anything
other than the name of the current folder.
}
\examples{
\dontrun{
ds <- ds \%>\% 
    cd("Demographics") \%>\%
    setName("Key Demos.")
}
}
\seealso{
\code{\link[=cd]{cd()}} and \code{\link[=mv]{mv()}}
}
