% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-folder.R
\name{setOrder}
\alias{setOrder}
\title{Change the order of entities in folder}
\usage{
setOrder(folder, ord)
}
\arguments{
\item{folder}{A \code{VariableFolder} or other \code{*Folder} object}

\item{ord}{A vector of integer indices or character references to objects
contained in the folder}
}
\value{
\code{folder} with the order dictated by \code{ord}. The function also persists
that order on the server.
}
\description{
Change the order of entities in folder
}
