% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-decipher.R
\name{setupDecipherAuth}
\alias{setupDecipherAuth}
\title{Setup authentication with Decipher}
\usage{
setupDecipherAuth(key, company_name)
}
\arguments{
\item{key}{the API key you retrieved from Decipher}

\item{company_name}{the company name associated with your account in Decipher}
}
\value{
\code{NULL} if successful
}
\description{
You'll have to get your own API key
}
