% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/share.R
\name{share}
\alias{share}
\title{Share a dataset}
\usage{
share(dataset, email, edit = FALSE, notify = TRUE)
}
\arguments{
\item{dataset}{a CrunchDataset}

\item{email}{character: email address(es) of the users with whom to share the dataset. If there is no Crunch user associated with an email, an invitation will be sent.}

\item{edit}{logical: should the specified user be given edit privileges on the dataset? Default is \code{FALSE}. \code{edit} can be a single value or, if inviting multiple users, a vector of logical values of equal length of the number of emails given. Currently, only one user may have \code{edit=TRUE} at any given time, so there may not be more than one \code{TRUE} value in the logical vector.}

\item{notify:}{should users who are getting new privileges on this dataset be sent an email informing them of this fact? Default is \code{TRUE}.}
}
\value{
Invisibly, the dataset.
}
\description{
Share a dataset
}

