% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-dataset.R
\name{shojifyMetadata}
\alias{shojifyMetadata}
\title{Wrap variable metadata inside a dataset entity}
\usage{
shojifyMetadata(metadata, order = I(names(metadata)), ...)
}
\arguments{
\item{metadata}{list of variable metadata}

\item{order}{a valid "order" payload: list containing either aliases or
list(group, entities)}

\item{...}{dataset entity metadata. "name" is required.}

\item{return}{list suitiable for JSONing and POSTing to create a dataset}
}
\description{
Wrap variable metadata inside a dataset entity
}
\keyword{internal}
