% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-order.R, R/variable-order.R
\docType{methods}
\name{ShojiOrder-extract}
\alias{ShojiOrder-extract}
\alias{[,ShojiOrder,ANY,ANY-method}
\alias{[,ShojiOrder,character,ANY-method}
\alias{[[,ShojiOrder,ANY-method}
\alias{[[,ShojiOrder,character-method}
\alias{$,ShojiOrder-method}
\alias{[<-,ShojiOrder,character,missing,ShojiOrder-method}
\alias{[<-,ShojiOrder,ANY,missing,ShojiOrder-method}
\alias{[[<-,ShojiOrder,character,missing,list-method}
\alias{[[<-,ShojiOrder,character,missing,character-method}
\alias{[[<-,ShojiOrder,character,missing,OrderGroup-method}
\alias{[[<-,ShojiOrder,ANY,missing,OrderGroup-method}
\alias{[[<-,ShojiOrder,ANY,missing,ANY-method}
\alias{[[<-,ShojiOrder,ANY,missing,NULL-method}
\alias{[[<-,ShojiOrder,character,missing,NULL-method}
\alias{[[<-,ShojiOrder,character,missing,ShojiOrder-method}
\alias{$<-,ShojiOrder-method}
\alias{[,OrderGroup,ANY,ANY-method}
\alias{[,OrderGroup,character,ANY-method}
\alias{[[,OrderGroup,character-method}
\alias{[[,OrderGroup,ANY-method}
\alias{$,OrderGroup-method}
\alias{[[<-,OrderGroup,character,missing,list-method}
\alias{[[<-,OrderGroup,character,missing,character-method}
\alias{[[<-,OrderGroup,character,missing,ShojiOrder-method}
\alias{[[<-,OrderGroup,character,missing,OrderGroup-method}
\alias{[[<-,OrderGroup,ANY,missing,OrderGroup-method}
\alias{[[<-,OrderGroup,numeric,missing,NULL-method}
\alias{[[<-,OrderGroup,character,missing,NULL-method}
\alias{$<-,OrderGroup-method}
\alias{[[<-,VariableOrder,character,missing,CrunchDataset-method}
\alias{[[<-,VariableGroup,character,missing,CrunchDataset-method}
\title{Extract and update in VariableOrders and VariableGroups}
\usage{
\S4method{[}{ShojiOrder,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ShojiOrder,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{ShojiOrder,ANY}(x, i, j, ...)

\S4method{[[}{ShojiOrder,character}(x, i, j, ...)

\S4method{$}{ShojiOrder}(x, name)

\S4method{[}{ShojiOrder,character,missing,ShojiOrder}(x, i, j) <- value

\S4method{[}{ShojiOrder,ANY,missing,ShojiOrder}(x, i, j) <- value

\S4method{[[}{ShojiOrder,character,missing,list}(x, i, j) <- value

\S4method{[[}{ShojiOrder,character,missing,character}(x, i, j) <- value

\S4method{[[}{ShojiOrder,character,missing,OrderGroup}(x, i, j) <- value

\S4method{[[}{ShojiOrder,ANY,missing,OrderGroup}(x, i, j) <- value

\S4method{[[}{ShojiOrder,ANY,missing,ANY}(x, i, j) <- value

\S4method{[[}{ShojiOrder,ANY,missing,`NULL`}(x, i, j) <- value

\S4method{[[}{ShojiOrder,character,missing,`NULL`}(x, i, j) <- value

\S4method{[[}{ShojiOrder,character,missing,ShojiOrder}(x, i, j) <- value

\S4method{$}{ShojiOrder}(x, name) <- value

\S4method{[}{OrderGroup,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{OrderGroup,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{OrderGroup,character}(x, i, j, ...)

\S4method{[[}{OrderGroup,ANY}(x, i, j, ...)

\S4method{$}{OrderGroup}(x, name)

\S4method{[[}{OrderGroup,character,missing,list}(x, i, j) <- value

\S4method{[[}{OrderGroup,character,missing,character}(x, i, j) <- value

\S4method{[[}{OrderGroup,character,missing,ShojiOrder}(x, i, j) <- value

\S4method{[[}{OrderGroup,character,missing,OrderGroup}(x, i, j) <- value

\S4method{[[}{OrderGroup,ANY,missing,OrderGroup}(x, i, j) <- value

\S4method{[[}{OrderGroup,numeric,missing,`NULL`}(x, i, j) <- value

\S4method{[[}{OrderGroup,character,missing,`NULL`}(x, i, j) <- value

\S4method{$}{OrderGroup}(x, name) <- value

\S4method{[[}{VariableOrder,character,missing,CrunchDataset}(x, i,
  j) <- value

\S4method{[[}{VariableGroup,character,missing,CrunchDataset}(x, i,
  j) <- value
}
\arguments{
\item{x}{a VariableOrder or VariableGroup}

\item{i}{an index. Numeric and logical indexing supported for both classes;
character indexing supported for VariableOrder, matching on VariableGroup
names}

\item{j}{Invalid}

\item{...}{additional arguments}

\item{drop}{Ignored}

\item{name}{Same as i but for \code{$}}

\item{value}{For update methods, an object equivalent in class to what is
being updated}
}
\value{
\code{[[} and \code{$} on a VariableOrder return the VariableGroup.
\code{[[} on VariableGroup returns the entity within, either a character
(URL) or nested VariableGroup. \code{[} and assignment methods return
objects of the same class as \code{x}
}
\description{
Extract and update in VariableOrders and VariableGroups
}
