% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-order.R
\docType{methods}
\name{ShojiOrder-length}
\alias{ShojiOrder-length}
\alias{length,OrderGroup-method}
\alias{length,ShojiOrder-method}
\title{Length of an Order}
\usage{
\S4method{length}{ShojiOrder}(x)

\S4method{length}{OrderGroup}(x)
}
\arguments{
\item{x}{a ShojiOrder}
}
\value{
Integer: the number of elements in the Order
}
\description{
Length of an Order
}

