% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji.R
\name{shojiURL}
\alias{shojiURL}
\title{Get a resource URL from a Shoji Object}
\usage{
shojiURL(x, collection = c("catalogs", "views", "fragments"), key)
}
\arguments{
\item{x}{a shojiObject}

\item{collection}{one of c("catalogs", "views", "fragments")}

\item{key}{character name of the URL to get from \code{collection}}
}
\value{
character URL
}
\description{
Get a resource URL from a Shoji Object
}
\keyword{internal}

