% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-transforms.R
\docType{methods}
\name{showTransforms,CategoricalVariable-method}
\alias{showTransforms,CategoricalVariable-method}
\alias{showTransforms}
\title{Show the variable transformations on a Categorical variable}
\usage{
\S4method{showTransforms}{CategoricalVariable}(x)
}
\arguments{
\item{x}{a Categorical variable}
}
\value{
summary of the variable, with transforms applied
}
\description{
Show the variable transformations on a Categorical variable
}
