% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding-categories.R
\name{slideCategories}
\alias{slideCategories}
\title{Create sliding subvariable definitions}
\usage{
slideCategories(variable, step, width, ..., complete = TRUE, useNA = FALSE)
}
\arguments{
\item{variable}{A categorical crunch variable}

\item{step}{number of categories between starting points of groups}

\item{width}{number of categories wide the grouping should be}

\item{...}{additional attributes to be included in the \code{SubvariableDefinition},
can be either functions that take the category names to be included in the
sliding group and returns a single string, or a character vector the same length
as the number of subvariables that will be created.}

\item{complete}{whether to only include category groupings that are as wide as width
(defaults to \code{TRUE})}

\item{useNA}{whether to use missing categories from the original variable (defaults
to \code{FALSE})}
}
\value{
A list of \code{SubvariableDefinition}s appropriate for use in \code{deriveArray()}
}
\description{
Create
}
