% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{slides}
\alias{slides}
\title{Access the slides of a CrunchDeck}
\usage{
slides(x)
}
\arguments{
\item{x}{a CrunchDeck}
}
\description{
Return a SlideCatalog from a CrunchDeck. All slide catalog methods should be
available for CrunchDecks, but this function is used internally to model the
API.
}
