% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide-transforms.R
\name{slideTransform}
\alias{slideTransform}
\alias{transforms,CrunchSlide-method}
\alias{transforms,AnalysisCatalog-method}
\alias{transforms,Analysis-method}
\alias{transforms<-,CrunchSlide,ANY-method}
\alias{transforms<-,AnalysisCatalog,ANY-method}
\alias{transforms<-,Analysis,ANY-method}
\title{Helper for creating slide dimension transformations for dashboards and exports}
\usage{
slideTransform(
  palette = NULL,
  hide = NULL,
  renames = NULL,
  order = NULL,
  name = NULL,
  description = NULL,
  ...
)

\S4method{transforms}{CrunchSlide}(x)

\S4method{transforms}{AnalysisCatalog}(x)

\S4method{transforms}{Analysis}(x)

\S4method{transforms}{CrunchSlide,ANY}(x) <- value

\S4method{transforms}{AnalysisCatalog,ANY}(x) <- value

\S4method{transforms}{Analysis,ANY}(x) <- value
}
\arguments{
\item{palette}{A vector of color RGB hex color codes that will be used for the color of
graphs in the dashboard.}

\item{hide}{A vector of category names/ids or subvariable names/aliases to hide from display}

\item{renames}{A named vector of category names/ids or subvariable names/aliases to override
their default values}

\item{order}{A vector of category names/ids or subvariable names/aliases to override the
default ordering of the dimension.}

\item{name}{A name for the dimension, overrides the variable's name}

\item{description}{A description for the dimension, overrides the variable's description}

\item{...}{Other arguments, passed directly to the API for future expansion}
}
\description{
When displayed in a Crunch Dashboard or exported, crunch slides can have transformations
that customize their display. This is a helper to form the correct data structure for
the functions \code{\link[=newSlide]{newSlide()}} pr setting the transformation directly. For more details see the
\href{https://crunch.io/api/reference/#post-/datasets/-dataset_id-/decks/-deck_id-/slides/}{API documentation}
}
