% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO-survey.R
\docType{data}
\name{SO_schema}
\alias{SO_schema}
\title{Schema for the 2017 Stack Overflow developer survey}
\format{An object of class \code{data.frame} with 23 rows and 2 columns.}
\usage{
SO_schema
}
\keyword{datasets}
