% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-stream.R
\docType{methods}
\name{streaming-flag}
\alias{streaming-flag}
\alias{streaming<-,CrunchDataset-method}
\title{Set the streaming property of a dataset}
\usage{
\S4method{streaming}{CrunchDataset}(x)

\S4method{streaming}{CrunchDataset}(x) <- value
}
\arguments{
\item{value}{for setting only (values can be: \code{"no"}, \code{"streaming"}, or
\code{"finished"})}

\item{ds}{a CrunchDataset}
}
\value{
the streaming status
}
\description{
Only datasets that have their streamining property set to "streaming" can
have rows streamed to them. Before attempting to streaming rows (with
\link{streamRows} for example), the dataset has to be set up to stream rows. Use
\code{streaming(ds)} to get the streaming status, and \code{streaming(ds) <- "streaming"} to set the streaming status.
}
