% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-stream.R
\name{streamRows}
\alias{streamRows}
\title{Stream data to a Crunch dataset}
\usage{
streamRows(ds, data)
}
\arguments{
\item{ds}{a CrunchDataset}

\item{data}{a dataframe with data to send as a stream, The given data values
must be in the Crunch I/O format (for example, category ids instead of
names or numeric_values)}
}
\description{
Stream data to a Crunch dataset
}
\keyword{internal}
