% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/subtotals-and-headings.R
\docType{class}
\name{Subtotal-class}
\alias{Subtotal-class}
\alias{subtotals}
\alias{subtotals<-}
\alias{makeInsertion}
\alias{Subtotal}
\alias{Heading-class}
\alias{Heading}
\alias{subtotals,CrunchVariable-method}
\alias{subtotals,VariableTuple-method}
\alias{subtotals<-,CrunchVariable,ANY-method}
\alias{subtotals<-,CrunchVariable,NULL-method}
\alias{makeInsertion,Subtotal-method}
\alias{makeInsertion,Heading-method}
\alias{makeInsertion,ANY-method}
\title{Subtotals and headings}
\usage{
Subtotal(..., data = NULL)

Heading(..., data = NULL)

\S4method{subtotals}{CrunchVariable}(x)

\S4method{subtotals}{VariableTuple}(x)

\S4method{subtotals}{CrunchVariable,ANY}(x) <- value

\S4method{subtotals}{CrunchVariable,`NULL`}(x) <- value

\S4method{makeInsertion}{Subtotal}(x, var)

\S4method{makeInsertion}{Heading}(x, var)

\S4method{makeInsertion}{ANY}(x, var)
}
\arguments{
\item{...}{additional arguments to \code{[}, ignored}

\item{data}{For the constructor functions \code{Subtotal} and
\code{Subtotal}, you can either pass in attributes via \code{...} or you
can create the objects with a fully defined \code{list} representation of
the objects via the \code{data} argument. See the examples.}

\item{x}{For the attribute getters and setters, an object of class
Subtotal}

\item{value}{For \code{[<-}, the replacement Subtotal to insert}

\item{var}{the variable to use to make \code{Insertions}}
}
\description{
Subtotals and headings allow you to add subtotals and headings to variables
or CrunchCubes. These are especially useful for making aggregates across
multiple categories (sometimes referred to as \emph{nets}, \emph{top box}, or \emph{top 2 box}).
}
