% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtotals-and-headings.R
\name{subtypeInsertion}
\alias{subtypeInsertion}
\alias{subtypeInsertions}
\alias{subtypeInsertion}
\title{Convert from Insertion to Insertion subtypes}
\usage{
subtypeInsertions(inserts)

subtypeInsertion(insert)
}
\arguments{
\item{inserts}{an \code{Insertions} object for \code{subtypeInsertions}}

\item{insert}{a single \code{Insertion} object for \code{subtypeInsertion}}
}
\value{
an \code{Insertions} object with members subtyped for \code{subtypeInsertions}
or a \code{Subtotal}, \code{Heading}, or \code{Insertion} for \code{subtypeInsertion}
}
\description{
Convert from Insertions with only Insertion objects (the only kind that are
available from the Crunch API) to the more user-friendly and user-facing
Subtotal and Heading object classes.
}
\keyword{internal}
