% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-definition.R
\name{SubvariableDefinition}
\alias{SubvariableDefinition}
\title{Construct a Subvariable Definition}
\usage{
SubvariableDefinition(x, ...)
}
\arguments{
\item{x}{A \code{CrunchExpr}}

\item{...}{additional attributes to be included in the references section (eg name, alias, etc.)}
}
\value{
a SubvariableDefinition object, ready to be included in a POST to Crunch.
}
\description{
Some specifications for creating Crunch Variables allow for subvariables to
be posted as well. This function converts a \code{CrunchExpr} to a format that
can be posted to the server.
}
