% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subvariables.R
\docType{methods}
\name{subvars-extract}
\alias{$,CategoricalArrayVariable-method}
\alias{$,Subvariables-method}
\alias{$<-,CategoricalArrayVariable-method}
\alias{$<-,Subvariables-method}
\alias{[,CategoricalArrayVariable,character,ANY-method}
\alias{[,Subvariables,character,ANY-method}
\alias{[<-,Subvariables,ANY,missing,ANY-method}
\alias{[<-,Subvariables,ANY,missing,Subvariables-method}
\alias{[<-,Subvariables,character,missing,Subvariables-method}
\alias{[[,CategoricalArrayVariable,ANY-method}
\alias{[[,CategoricalArrayVariable,character-method}
\alias{[[,Subvariables,ANY-method}
\alias{[[,Subvariables,character-method}
\alias{[[<-,CategoricalArrayVariable,ANY,missing,ANY-method}
\alias{[[<-,CategoricalArrayVariable,character,missing,ANY-method}
\alias{[[<-,Subvariables,ANY,missing,ANY-method}
\alias{[[<-,Subvariables,ANY,missing,CrunchVariable-method}
\alias{[[<-,Subvariables,ANY,missing,NULL-method}
\alias{[[<-,Subvariables,character,missing,CrunchVariable-method}
\alias{subvars-extract}
\title{Extract and modify subsets of subvariables}
\usage{
\S4method{[[}{Subvariables,character}(x, i, j, ...)

\S4method{[[}{Subvariables,ANY}(x, i, j, ...)

\S4method{$}{Subvariables}(x, name)

\S4method{[}{Subvariables,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{Subvariables,character,missing,CrunchVariable}(x, i) <- value

\S4method{[[}{Subvariables,ANY,missing,CrunchVariable}(x, i) <- value

\S4method{[[}{Subvariables,ANY,missing,`NULL`}(x, i) <- value

\S4method{[[}{Subvariables,ANY,missing,ANY}(x, i) <- value

\S4method{$}{Subvariables}(x, name) <- value

\S4method{[}{Subvariables,character,missing,Subvariables}(x, i) <- value

\S4method{[}{Subvariables,ANY,missing,Subvariables}(x, i) <- value

\S4method{[}{Subvariables,ANY,missing,ANY}(x, i) <- value

\S4method{[}{CategoricalArrayVariable,character,ANY}(x, i, j, ...,
  drop = TRUE)

\S4method{[[}{CategoricalArrayVariable,ANY}(x, i, j, ...)

\S4method{[[}{CategoricalArrayVariable,character}(x, i, j, ...)

\S4method{$}{CategoricalArrayVariable}(x, name)

\S4method{[[}{CategoricalArrayVariable,ANY,missing,ANY}(x, i) <- value

\S4method{[[}{CategoricalArrayVariable,character,missing,ANY}(x, i) <- value

\S4method{$}{CategoricalArrayVariable}(x, name) <- value
}
\arguments{
\item{x}{Subvariables or an array Variable (which contains subvariables)}

\item{i}{which subvariables to extract}

\item{j}{Invalid}

\item{...}{additional arguments}

\item{name}{For \code{$}, the name (not alias) of the subvariable to
extract}

\item{drop}{Invalid}

\item{value}{For updating, a CrunchExpr}
}
\value{
A subset of \code{x} if extracting, otherwise \code{x} duly modified
}
\description{
Extract and modify subsets of subvariables
}

