% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-book.R
\docType{methods}
\name{tabbook-dim}
\alias{tabbook-dim}
\alias{dim,TabBookResult-method}
\title{TabBookResult and MultitableResult dimension}
\usage{
\S4method{dim}{TabBookResult}(x)
}
\arguments{
\item{x}{a TabBookResult or MultitableResult}
}
\value{
Returns what you'd expect.
}
\description{
TabBookResult and MultitableResult dimension
}
