% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-book.R
\docType{methods}
\name{tabbook-methods}
\alias{tabbook-methods}
\alias{length,TabBookResult-method}
\alias{[[,TabBookResult,numeric-method}
\alias{[[,TabBookResult,character-method}
\alias{dim,TabBookResult-method}
\alias{names,TabBookResult-method}
\alias{length,MultitableResult-method}
\alias{[[,MultitableResult,ANY-method}
\title{TabBookResult and MultitableResult methods}
\usage{
\S4method{length}{TabBookResult}(x)

\S4method{[[}{TabBookResult,numeric}(x, i, j, ...)

\S4method{[[}{TabBookResult,character}(x, i, j, ...)

\S4method{dim}{TabBookResult}(x)

\S4method{names}{TabBookResult}(x)

\S4method{length}{MultitableResult}(x)

\S4method{[[}{MultitableResult,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{a TabBookResult or MultitableResult}

\item{i}{an index into \code{x}}

\item{j}{an index into \code{x}, ignored}

\item{...}{also ignored}
}
\value{
Returns what you'd expect.
}
\description{
TabBookResult and MultitableResult methods
}
