% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-book.R
\name{tabFramePrepare}
\alias{tabFramePrepare}
\title{A utility function that prepares}
\usage{
tabFramePrepare(dataset, weight, include_original_weighted)
}
\arguments{
\item{dataset}{CrunchDataset, which may be subset with a filter expression
on the rows, and a selection of variables to use on the columns.}

\item{weight}{a CrunchVariable that has been designated as a potential
weight variable for \code{dataset}, or \code{NULL} for unweighted results. Alternatively
a named list where the name is the alias of the weight and the contents of
the list component are a character vector of aliases to which that weight
should apply.
Default is the currently applied \code{\link{weight}}.}

\item{include_original_weighted}{Logical, if you have specified complex weights
should the original weighted variable be included or only the custom weighted version?}
}
\description{
A utility function that prepares
}
