% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R, R/project-folder.R, R/projects.R,
%   R/teams.R
\docType{methods}
\name{members<-,CrunchTeam,MemberCatalog-method}
\alias{members<-,CrunchTeam,MemberCatalog-method}
\alias{members<-,CrunchTeam,character-method}
\alias{members,ProjectFolder-method}
\alias{members<-,ProjectFolder,MemberCatalog-method}
\alias{members<-,ProjectFolder,character-method}
\alias{members,CrunchProject-method}
\alias{members<-,CrunchProject,MemberCatalog-method}
\alias{members<-,CrunchProject,character-method}
\alias{teams}
\alias{members}
\alias{members<-}
\alias{members,CrunchTeam-method}
\title{Teams}
\usage{
\S4method{members}{CrunchTeam,MemberCatalog}(x) <- value

\S4method{members}{CrunchTeam,character}(x) <- value

\S4method{members}{ProjectFolder}(x)

\S4method{members}{ProjectFolder,MemberCatalog}(x) <- value

\S4method{members}{ProjectFolder,character}(x) <- value

\S4method{members}{CrunchProject}(x)

\S4method{members}{CrunchProject,MemberCatalog}(x) <- value

\S4method{members}{CrunchProject,character}(x) <- value

\S4method{members}{CrunchTeam}(x)
}
\arguments{
\item{x}{a \code{CrunchTeam}}

\item{value}{for \code{members<-}, a character vector of emails or URLs of
users to add to the team.}
}
\value{
\code{members} returns a
\code{MemberCatalog}, which has references to the users that are members
of the team. \code{members<-} returns \code{x} with the given users added
to the members catalog.
}
\description{
Teams contain users and datasets. You can share a dataset with a group of
users by sharing the dataset with a team. You can also share a set of
datasets with a user all at once by adding the user to a team that contains those
datasets.
}
\details{
These methods allow you to work with teams. Find your teams with the
\code{\link[=getTeams]{getTeams()}} function, which returns your \code{TeamCatalog}. You can extract an individual team by name,
or create a team by assigning into the function. To create a team by assignment, assign a list
to \code{teams("myteam") <- value_list}, the \code{value_list} can either empty (to just create a team
with that name), or can contain a "members" element with the emails or URLs of
users to add to the team. Users can be also be added later with the \code{members<-}
method.
}
\seealso{
\code{\link{getTeams}}
}
