% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-manager.R
\name{temp.options}
\alias{temp.options}
\alias{temp.option}
\title{Set some global options temporarily}
\usage{
temp.options(..., crunch = list())

temp.option(..., crunch = list())
}
\arguments{
\item{...}{named options to set using \code{options()}}

\item{crunch}{named list of options to set in crunch's higher priority
options environment}
}
\value{
an S3 class "contextManager" object
}
\description{
Set some global options temporarily
}
\seealso{
\code{\link{with-context-manager}} \code{\link{ContextManager}}
}
