% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise.R
\name{tieredVar}
\alias{tieredVar}
\title{Create a categorical variable using tiers of a Categorical Array}
\usage{
tieredVar(x, tiers, name, ...)
}
\arguments{
\item{x}{A Categorical Array variable}

\item{tiers}{A vector of ids, category names, or values. Can pass a named
vector, where the names indicate the type, or if there are no names, then
numeric values are assumed to be ids and characters are assumed to be names.}

\item{name}{a character to use as the name of the case variable to create}

\item{...}{Metadata like name or description that is passed to \code{\link[=VarDef]{VarDef()}}}
}
\value{
A Variable Definition
}
\description{
Collapses a categorical array to the first value of tiers that is found.
}
