% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/categories.R, R/variable-order.R
\docType{methods}
\name{toJSON,Categories-method}
\alias{toJSON,Categories-method}
\alias{toJSON,VariableGroup-method}
\alias{toJSON,VariableOrder-method}
\title{toJSON methods for Crunch objects}
\usage{
\S4method{toJSON}{Categories}(x, container = isContainer(x, asIs, .level),
  collapse = "\\n", ..., .level = 1L, .withNames = length(x) > 0 &&
  length(names(x)) > 0, .na = "null", .escapeEscapes = TRUE,
  pretty = FALSE, asIs = NA, .inf = " Infinity")

\S4method{toJSON}{VariableOrder}(x, container = isContainer(x, asIs, .level),
  collapse = "\\n", ..., .level = 1L, .withNames = length(x) > 0 &&
  length(names(x)) > 0, .na = "null", .escapeEscapes = TRUE,
  pretty = FALSE, asIs = NA, .inf = " Infinity")

\S4method{toJSON}{VariableGroup}(x, container = isContainer(x, asIs, .level),
  collapse = "\\n", ..., .level = 1L, .withNames = length(x) > 0 &&
  length(names(x)) > 0, .na = "null", .escapeEscapes = TRUE,
  pretty = FALSE, asIs = NA, .inf = " Infinity")
}
\arguments{
\item{x}{the object}

\item{container}{argument to the generic, not used}

\item{collapse}{argument to the generic, not used}

\item{...}{additional arguments argument to the generic, not used}

\item{.level}{argument to the generic, not used}

\item{.withNames}{argument to the generic, not used}

\item{.na}{argument to the generic, not used}

\item{.escapeEscapes}{argument to the generic, not used}

\item{pretty}{argument to the generic, not used}

\item{asIs}{argument to the generic, not used}

\item{.inf}{argument to the generic, not used}
}
\value{
JSON-serialized character object
}
\description{
toJSON methods for Crunch objects
}

