% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/R-to-variable.R,
%   R/expressions.R
\docType{methods}
\name{toVariable}
\alias{toVariable}
\alias{toVariable,character-method}
\alias{toVariable,numeric-method}
\alias{toVariable,factor-method}
\alias{toVariable,Date-method}
\alias{toVariable,POSIXt-method}
\alias{toVariable,AsIs-method}
\alias{toVariable,VariableDefinition-method}
\alias{toVariable,logical-method}
\alias{toVariable,CrunchExpr-method}
\title{Generic method for converting objects to Crunch representations}
\usage{
toVariable(x, ...)

\S4method{toVariable}{character}(x, ...)

\S4method{toVariable}{numeric}(x, ...)

\S4method{toVariable}{factor}(x, ...)

\S4method{toVariable}{Date}(x, ...)

\S4method{toVariable}{POSIXt}(x, ...)

\S4method{toVariable}{AsIs}(x, ...)

\S4method{toVariable}{VariableDefinition}(x, ...)

\S4method{toVariable}{logical}(x, ...)

\S4method{toVariable}{CrunchExpr}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional arguments}
}
\value{
a list object suitable for POSTing to the Crunch API. See the API
documentation for specifications.
}
\description{
R objects are converted to crucnh objects using the following rules:
}
\details{
\itemize{
\item Character vectors are converted into Crunch text variables
\item Numeric vectors are converted into Crunch numeric variables
\item Factors are converted to categorical variables
\item Date and POSIXt vectors are converted into Crunch datetime variables
\item Logical vectors are converted to Crunch categorical variables
\item \link{VariableDefinition}s are not converted, but the function can still append
additional metadata
}

If you have other object types you wish to convert to Crunch variables,
you can declare methods for \code{toVariable}
}
