% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variable-type.R
\docType{methods}
\name{type}
\alias{type}
\alias{type<-}
\alias{type<-,CrunchVariable-method}
\title{Change the type of Crunch variables}
\usage{
\S4method{type}{CrunchVariable}(x)

\S4method{type}{CrunchVariable}(x) <- value
}
\arguments{
\item{x}{a Variable}

\item{value}{For the setter, a character value in c("numeric", "text",
"categorical")}
}
\value{
Getter returns character; setter returns \code{x} duly modified.
}
\description{
Numeric, text, and categorical variables can be cast to one another by
assigning them a new "type". This modifes the storage of the data on the
server and should only be done in narrow circumstances, as in when importing
data from a different file format has resulted in incorrect types being
specified.
}

