% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unbind}
\alias{unbind}
\title{Split an array or multiple-response variable into its CategoricalVariables}
\usage{
unbind(x)
}
\arguments{
\item{x}{a CategoricalArrayVariable or MultipleResponseVariable}
}
\value{
invisibly, the API response from DELETEing the array variable
definition. If you \code{\link{refresh}} the corresponding dataset after
unbinding, you should see the array variable removed and its subvariables
promoted to regular variables.
}
\description{
Split an array or multiple-response variable into its CategoricalVariables
}

