% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{unshare}
\alias{unshare}
\title{Revoke a user's access to a dataset}
\usage{
unshare(dataset, users)
}
\arguments{
\item{dataset}{a CrunchDataset}

\item{users}{character: email address(es) or URLs of the users or teams to
unshare with.}
}
\value{
Invisibly, the dataset.
}
\description{
Revoke a user's access to a dataset
}
\seealso{
\code{\link{share}}
}
