% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-datasets.R
\name{updateDatasetList}
\alias{updateDatasetList}
\title{Refresh the local list of Crunch datasets}
\usage{
updateDatasetList()
}
\value{
Nothing. Called for its side effects of setting local environment
variables.
}
\description{
Refresh the local list of Crunch datasets
}

