% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shoji-catalog.R, R/variable-order-slots.R
\docType{methods}
\name{urls,ShojiCatalog-method}
\alias{urls}
\alias{urls,ShojiCatalog-method}
\alias{urls,VariableGroup-method}
\alias{urls,VariableOrder-method}
\title{Get the URLs contained in a Catalog or Order object}
\usage{
\S4method{urls}{ShojiCatalog}(x)

\S4method{urls}{VariableOrder}(x)

\S4method{urls}{VariableGroup}(x)
}
\arguments{
\item{x}{a Catalog, Order, or Group object}
}
\value{
A character vector of URLs
}
\description{
Sometimes it is useful to extract flattened vector of URLs from more
complex objects for purposes like subsetting or doing set comparisons.
}
\keyword{internal}

