% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/user.R
\docType{methods}
\name{email}
\alias{email}
\alias{email,UserEntity-method}
\title{Extract the email from a User Entity}
\usage{
email(x)

\S4method{email}{UserEntity}(x)
}
\arguments{
\item{x}{a UserEntity returned from \code{me()}}
}
\value{
a character string of the user's email
}
\description{
Extract the email from a User Entity
}
