% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-result.R
\name{user2real}
\alias{user2real}
\title{Convert from user margins to real cube margins}
\usage{
user2real(margin, dimTypes = getDimTypes(cube), cube)
}
\arguments{
\item{margin}{the margin or margins for the user cube to be translated}

\item{dimTypes}{dimension types from \code{getDimTypes()} (by default:
\code{getDimTypes(cube)})}

\item{cube}{the cube to translate the margin for (optional if \code{dimTypes} is
explicitly supplied)}
}
\value{
margin or margins in the higher-dimension real cube
}
\description{
It is helpful to programmatically move from user-specified margins to real
cube margins that apply to the higher-dimensional real cube.
}
\keyword{internal}
