% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-access.R
\name{userInfo}
\alias{userInfo}
\title{Get information about when users last accessed a dataset}
\usage{
userInfo(dataset_entity, user_fields = c("name", "email", "access_time"))
}
\arguments{
\item{dataset_entity}{a dataset object or a dataset entity (e.g. a single
entry from \code{selectDatasetCatalog()})}

\item{user_fields}{metadata fields about users to include in the dataframe
(default: c("name", "email", "access_time"))}
}
\value{
a dataframe with information about all users for a particular dataset
}
\description{
Useful to get dataset-user interaction metadata about all of the users that
have access to a particular dataset.
}
