% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-info.R
\docType{methods}
\name{users}
\alias{users}
\alias{users,CrunchDataset-method}
\alias{users,DatasetTuple-method}
\alias{users,ProjectFolder-method}
\title{Get information about users who have access to a dataset}
\usage{
\S4method{users}{CrunchDataset}(x)

\S4method{users}{DatasetTuple}(x)

\S4method{users}{ProjectFolder}(x)
}
\arguments{
\item{x}{a \code{CrunchDataset}, \code{DatasetTuple}, or \code{ProjectFolder} object to get
users from}
}
\value{
a \code{UserCatalog} with information about users who have access to the
dataset
}
\description{
Get user metadata about all of the users that have access to a particular
Crunch object like a dataset or project. Returns a \code{UserCatalog} object which
can be translated into a data.frame with \code{\link[=catalogToDataFrame]{catalogToDataFrame()}} if information
needs to be extracted, queried, transformed, etc.
}
