% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-transforms.R
\name{validateNamesInDims}
\alias{validateNamesInDims}
\title{error iff the names are not a dimension in the cube provided}
\usage{
validateNamesInDims(names, cube, what = "transforms")
}
\arguments{
\item{names}{the names to check for in the cube}

\item{cube}{a CrunchCube object to check}

\item{what}{a character describing what is being checked (default: transforms
) to include in the error to make it easier for users to see what is failing.}
}
\description{
error iff the names are not a dimension in the cube provided
}
\keyword{internal}
