% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{validateWeightVariableValue}
\alias{validateWeightVariableValue}
\title{Validate that a value is a logical vector of length 1}
\usage{
validateWeightVariableValue(value)
}
\arguments{
\item{value}{The right hand side of an assignment}
}
\value{
nothing, called for its side effects
}
\description{
Validate that a value is a logical vector of length 1
}
\keyword{internal}
