% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{validatWeightExpression}
\alias{validatWeightExpression}
\title{Validate an expression passed to makeWeight}
\usage{
validatWeightExpression(expr)
}
\arguments{
\item{expr}{}
}
\description{
Convenience function
}
