% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R, R/variable.R
\docType{methods}
\name{[,CrunchExpr,CrunchLogicalExpr,ANY-method}
\alias{[,CrunchExpr,CrunchLogicalExpr,ANY-method}
\alias{[,CrunchExpr,logical,ANY-method}
\alias{[,CrunchExpr,numeric,ANY-method}
\alias{[,CrunchVariable,CrunchExpr,ANY-method}
\alias{[,CrunchVariable,logical,ANY-method}
\alias{[,CrunchVariable,numeric,ANY-method}
\alias{variable-extract}
\title{"Subset" a Variable}
\usage{
\S4method{[}{CrunchExpr,CrunchLogicalExpr,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{CrunchExpr,logical,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{CrunchExpr,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{CrunchVariable,CrunchExpr,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{CrunchVariable,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{CrunchVariable,logical,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a Variable}

\item{i}{a CrunchExpr, logical, or numeric}

\item{j}{Invalid}

\item{...}{additional arguments, ignored}

\item{drop}{Invalid}
}
\value{
a CrunchExpr containing references to the variable \code{x} and the
filter logic contained in \code{i}
}
\description{
These methods subset variables by creating Expressions, which can be
composed and evaluated as needed.
}

