% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame.R
\name{as.data.frame.VariableCatalog}
\alias{as.data.frame.VariableCatalog}
\title{as.data.frame method for VariableCatalog}
\usage{
\method{as.data.frame}{VariableCatalog}(x, row.names = NULL,
  optional = FALSE, keys = c("alias", "name", "type"), ...)
}
\arguments{
\item{x}{A VariableCatalog produced by \code{variables(ds)}.}

\item{row.names}{part of as.data.frame signature. Ignored.}

\item{optional}{part of as.data.frame signature. Ignored.}

\item{keys}{A character vector of the variable catalog attributes which you would like
included in the datafame. To include all attributes or see which ones are available
set keys to "all".}

\item{...}{Additional arguments passed to \code{data.frame}}
}
\value{
A data frame which includes information about each variable stored in the variable catelog.
The fields in the dataframe match the keys argument provided to the function, and each row represents a variable.
}
\description{
This method allows you to import the VariableCatalog into your R sessions in
order to facilliate interactive use. Modifying the dataframe produced by this
method will not update the Crunch web app, for information about modifying
variables in the app see: \code{vingette("variables",
package = "crunch")}
}
\examples{
\dontrun{
ds <- loadDataset("iris")
vars <- variables(ds)
var_df <- as.data.frame(vars, keys = "all")
} 

}
