% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{VariableCatalog-class}
\alias{VariableCatalog-class}
\alias{VariableCatalog}
\title{Collection of Variables within a Dataset}
\description{
A VariableCatalog contains references to all variables in a dataset, plus
some descriptive metadata about each. Each VariableCatalog also contains a
\code{\link{VariableOrder}} that governs how variables within it are
organized.
}
