% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{VariableOrder-class}
\alias{VariableOrder}
\alias{VariableOrder-class}
\title{Organize Variables within a Dataset}
\description{
Variables in the Crunch web application can be viewed in an ordered,
hierarchical list. These objects and methods allow you to modify that order
from R.
}
\details{
A VariableOrder object is a subclass of \code{list} that contains
VariableGroups. VariableGroup objects contain a group name and an set of
"entities", which can be variable references or other nested VariableGroups.

The \code{ordering} methods allow you to get and set a VariableOrder on
a \code{\link{CrunchDataset}} or on the \code{\link{VariableCatalog}} that
the dataset contains.
}
\section{Slots}{

\describe{
\item{\code{group}}{character, the name of the VariableGroup. In the constructor and
more generally, this field can be referenced as "name" as well.}

\item{\code{entities}}{a character vector of variable URLs, or a list containing a
combination of variable URLs and VariableGroup objects.}
}}

