% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variable-order.R
\docType{methods}
\name{VariableOrder-extract}
\alias{$,VariableGroup-method}
\alias{$,VariableOrder-method}
\alias{$<-,VariableGroup-method}
\alias{$<-,VariableOrder-method}
\alias{VariableOrder-extract}
\alias{[,VariableGroup,ANY,ANY-method}
\alias{[,VariableGroup,character,ANY-method}
\alias{[,VariableOrder,ANY,ANY-method}
\alias{[,VariableOrder,character,ANY-method}
\alias{[<-,VariableOrder,ANY,missing,VariableOrder-method}
\alias{[<-,VariableOrder,character,missing,VariableOrder-method}
\alias{[[,VariableGroup,ANY-method}
\alias{[[,VariableGroup,character-method}
\alias{[[,VariableOrder,ANY-method}
\alias{[[,VariableOrder,character-method}
\alias{[[<-,VariableGroup,ANY,missing,VariableGroup-method}
\alias{[[<-,VariableGroup,character,missing,CrunchDataset-method}
\alias{[[<-,VariableGroup,character,missing,VariableGroup-method}
\alias{[[<-,VariableGroup,character,missing,VariableOrder-method}
\alias{[[<-,VariableGroup,character,missing,character-method}
\alias{[[<-,VariableGroup,character,missing,list-method}
\alias{[[<-,VariableOrder,ANY,missing,ANY-method}
\alias{[[<-,VariableOrder,ANY,missing,NULL-method}
\alias{[[<-,VariableOrder,ANY,missing,VariableGroup-method}
\alias{[[<-,VariableOrder,character,missing,CrunchDataset-method}
\alias{[[<-,VariableOrder,character,missing,NULL-method}
\alias{[[<-,VariableOrder,character,missing,VariableGroup-method}
\alias{[[<-,VariableOrder,character,missing,VariableOrder-method}
\title{Extract and update in VariableOrder and VariableGroup}
\usage{
\S4method{[}{VariableOrder,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{VariableOrder,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{VariableOrder,ANY}(x, i, j, ...)

\S4method{[[}{VariableOrder,character}(x, i, j, ...)

\S4method{$}{VariableOrder}(x, name)

\S4method{[}{VariableOrder,character,missing,VariableOrder}(x, i, j) <- value

\S4method{[}{VariableOrder,ANY,missing,VariableOrder}(x, i, j) <- value

\S4method{[[}{VariableOrder,character,missing,VariableGroup}(x, i, j) <- value

\S4method{[[}{VariableOrder,character,missing,CrunchDataset}(x, i, j) <- value

\S4method{[[}{VariableOrder,character,missing,VariableOrder}(x, i, j) <- value

\S4method{[[}{VariableOrder,ANY,missing,VariableGroup}(x, i, j) <- value

\S4method{[[}{VariableOrder,ANY,missing,ANY}(x, i, j) <- value

\S4method{[[}{VariableOrder,ANY,missing,`NULL`}(x, i, j) <- value

\S4method{[[}{VariableOrder,character,missing,`NULL`}(x, i, j) <- value

\S4method{$}{VariableOrder}(x, name) <- value

\S4method{[}{VariableGroup,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{VariableGroup,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{VariableGroup,character}(x, i, j, ...)

\S4method{[[}{VariableGroup,ANY}(x, i, j, ...)

\S4method{$}{VariableGroup}(x, name)

\S4method{[[}{VariableGroup,character,missing,CrunchDataset}(x, i, j) <- value

\S4method{[[}{VariableGroup,character,missing,list}(x, i, j) <- value

\S4method{[[}{VariableGroup,character,missing,character}(x, i, j) <- value

\S4method{[[}{VariableGroup,character,missing,VariableOrder}(x, i, j) <- value

\S4method{[[}{VariableGroup,character,missing,VariableGroup}(x, i, j) <- value

\S4method{[[}{VariableGroup,ANY,missing,VariableGroup}(x, i, j) <- value

\S4method{$}{VariableGroup}(x, name) <- value
}
\arguments{
\item{x}{a VariableOrder or VariableGroup}

\item{i}{an index. Numeric and logical indexing supported for both classes;
character indexing supported for VariableOrder, matching on VariableGroup
names}

\item{j}{Invalid}

\item{...}{additional arguments}

\item{drop}{Ignored}

\item{name}{Same as i but for \code{$}}

\item{value}{For update methods, an object equivalent in class to what is
being updated}
}
\value{
\code{[[} and \code{$} on a VariableOrder return the VariableGroup.
\code{[[} on VariableGroup returns the entity within, either a character
(URL) or nested VariableGroup. \code{[} and assignment methods return
objects of the same class as \code{x}
}
\description{
Extract and update in VariableOrder and VariableGroup
}

