% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-order.R
\docType{methods}
\name{VariableOrder-length}
\alias{VariableOrder-length}
\alias{length,VariableOrder-method}
\title{Length of VariableOrder}
\usage{
\S4method{length}{VariableOrder}(x)
}
\arguments{
\item{x}{a VariableOrder}
}
\value{
Integer: the number of VariableGroups in the VariableOrder
}
\description{
Length of VariableOrder
}

