% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-order-slots.R
\docType{methods}
\name{VariableOrder-slots}
\alias{VariableOrder-slots}
\alias{duplicates}
\alias{duplicates,VariableCatalog-method}
\alias{duplicates,VariableGroup-method}
\alias{duplicates,VariableOrder-method}
\alias{duplicates<-}
\alias{duplicates<-,VariableCatalog,logical-method}
\alias{duplicates<-,VariableGroup,logical-method}
\alias{duplicates<-,VariableOrder,logical-method}
\alias{entities}
\alias{entities,VariableGroup-method}
\alias{entities,VariableOrder-method}
\alias{entities,list-method}
\alias{entities<-}
\alias{entities<-,VariableGroup-method}
\alias{entities<-,VariableOrder-method}
\alias{name,VariableGroup-method}
\alias{name<-,VariableGroup,character-method}
\alias{names,VariableGroup-method}
\alias{names,VariableOrder-method}
\alias{names<-,VariableOrder-method}
\title{Manipulate VariableGroup and VariableOrder}
\usage{
\S4method{entities}{VariableGroup}(x, simplify = FALSE)

\S4method{entities}{VariableOrder}(x, simplify = FALSE)

\S4method{entities}{list}(x, simplify = FALSE)

\S4method{entities}{VariableGroup}(x) <- value

\S4method{entities}{VariableOrder}(x) <- value

\S4method{name}{VariableGroup}(x)

\S4method{name}{VariableGroup,character}(x) <- value

\S4method{names}{VariableOrder}(x)

\S4method{names}{VariableGroup}(x)

\S4method{names}{VariableOrder}(x) <- value

\S4method{duplicates}{VariableOrder}(x)

\S4method{duplicates}{VariableGroup}(x)

\S4method{duplicates}{VariableCatalog}(x)

\S4method{duplicates}{VariableOrder,logical}(x) <- value

\S4method{duplicates}{VariableGroup,logical}(x) <- value

\S4method{duplicates}{VariableCatalog,logical}(x) <- value
}
\arguments{
\item{x}{a VariableGroup or VariableOrder}

\item{simplify}{logical: should variable URLs inside of groups be flattened
or preserved in their nested lists? Default is \code{FALSE}.}

\item{value}{(1) For name, a character (length-1 vector); for names, a character
vector of equal length to the number of VariableGroups beind modified; for
entities, either a character vector of variable URLs or a list containing a
combination of variable URLs and VariableGroups. Note that group names must
be unique, should be greater than 0 characters long, and "ungrouped" is a
reserved group name. (2) For duplicates, logical for whether duplicate
variable entries should be allowed in the VariableOrder.}
}
\value{
\code{entities} returns Variable references and VariableGroups; \code{names} returns group names; \code{duplicates} returns logical for whether duplicate variable entries should be allowed
}
\description{
Manipulate VariableGroup and VariableOrder
}
\seealso{
\code{\link{VariableOrder}}

\code{\link{grouped}}
}

