% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_crunch_numeric_variable.R
\name{vec_arith.crunch_numeric_variable}
\alias{vec_arith.crunch_numeric_variable}
\title{Internal vctrs methods}
\usage{
\method{vec_arith}{crunch_numeric_variable}(op, x, y, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
