% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{versions}
\alias{versions}
\title{Access the saved versions of a dataset}
\usage{
versions(x)
}
\arguments{
\item{x}{a \code{CrunchDataset}}
}
\value{
an object of class \code{VersionCatalog}. Supported methods on the
catalog include "names" and "timestamps".
}
\description{
Access the saved versions of a dataset
}
\seealso{
\code{\link{saveVersion}} \code{\link{restoreVersion}}
}

