% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{webApp}
\alias{webApp}
\title{View a Dataset in the Web Application}
\usage{
webApp(dataset)
}
\arguments{
\item{dataset}{a CrunchDataset}
}
\value{
Nothing; called for side effect of opening your web browser.
}
\description{
Convenience function that will use your system's "open" command to open
a dataset in our web application in your default browser.
}
\details{
Note that this function does not do anything on Windows.
}
