\name{weight<-}
\alias{weight<-}
\title{Set the dataset's weight}
\usage{
  weight(x, value) <- value
}
\arguments{
  \item{x}{a Dataset}

  \item{value}{a Variable to set as weight, or NULL to
  remove the existing weight}
}
\value{
  x, modified accordingly
}
\description{
  Set the dataset's weight
}

