% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{weight}
\alias{weight}
\alias{weight<-}
\title{Dataset weights}
\usage{
weight(x)

weight(x) <- value
}
\arguments{
\item{x}{a Dataset}

\item{value}{a Variable to set as weight, or NULL to remove the existing
weight}
}
\value{
For the getter, a Variable if there is a weight, else NULL. For the
setter, x, modified accordingly
}
\description{
Dataset weights
}

