% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-comparisons.R
\name{wishartMatrix}
\alias{wishartMatrix}
\title{Matrix of Chi-Squared Statistics for all rows or columns}
\usage{
wishartMatrix(cube, dim = c("cols", "rows"))
}
\arguments{
\item{cube}{a CrunchCube of counts: a crosstab of two categorical variables}

\item{dim}{dimension along which to compare proportions (default \code{cols} or \code{rows})}
}
\value{
A symmetric square matrix of Chi-squared statistics for a crosstab
}
\description{
Use the alternative Wishart method of forming the matrix of column- or row-wise
comparison Chi-squared test statistics for a categorical-by-categorical
contingency table.
}
\examples{
\dontrun{
# TODO: include Hirotsu occupation-illness example cube and output
illness_occupation <- crunch_example_data(hirotsu)
# Chi-squared statistic for joint null hypothesis that each
# column is equal to each other column.

wishartMatrix(illness_occupation)
}
}
