% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-comparisons.R
\name{wishartPvalues}
\alias{wishartPvalues}
\title{P-values for pairwise comparison on a dimension of a categorical crosstab}
\usage{
wishartPvalues(cube, dim = c("cols", "rows"))
}
\arguments{
\item{cube}{a CrunchCube}

\item{dim}{dimension which jointly forms the null hypothesis that each
vector is equal to all of the others.}
}
\value{
A symmetric square matrix of the size of \code{dim}. Typical element
\eqn{i,j} is the P-value associated with the hypothesis that proportions in
row or column \eqn{i} are equal to those in row or column \eqn{j}.
}
\description{
P-values for pairwise comparison on a dimension of a categorical crosstab
}
\examples{
\dontrun{
# TODO: include Hirotsu occupation-illness example cube and output
illness_occupation <- crunch_example_data(hirotsu)
# P-value for joint null hypothesis that each
# column is equal to each other column.

wishartPvalues(illness_occupation)
}
}
