% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{with_api_fixture}
\alias{with_api_fixture}
\title{Test helper}
\usage{
with_api_fixture(fixture_path, expr)
}
\arguments{
\item{fixture_path}{Path to fixtures that have been stored by
\code{httptest:::capture_requests()}}

\item{expr}{An expression (or multiple if wrapped in brackets \verb{\{}, \verb{\}})}
}
\description{
Helpers for using \code{httptest} & crunch to write mock tests for
your package.
}
\keyword{internal}
