% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-dataset.R
\name{write.csv.gz}
\alias{write.csv.gz}
\title{Write CSV to a compressed file}
\usage{
write.csv.gz(x, file, na = "", row.names = FALSE, ...)
}
\arguments{
\item{x}{A data.frame or similar CSV-writeable object}

\item{file}{character destination to write the gzipped CSV to}

\item{na}{See \code{\link[utils:write.csv]{utils::write.csv}}. This just changes the default
to a Crunch-friendly empty string.}

\item{row.names}{logical: write out row names? See \code{\link[utils:write.csv]{utils::write.csv}}.}

\item{...}{Additional arguments passed to \code{write.csv}.}
}
\value{
A csv file written to dist
}
\description{
Write CSV to a compressed file
}
