% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zed-scores.R
\name{z.scores}
\alias{z.scores}
\title{Calculate z-scores for a CrunchCube}
\usage{
z.scores(cube, margin)
}
\arguments{
\item{cube}{A CrunchCube to calculate z-scores for}

\item{margin}{which margin to test against (1 for rows, 2 for columns)}
}
\value{
an array of z-scores
}
\description{
Calculate z-scores for each cell of a cube by comparing to the proportions on
the row or column margins. When testing against row margins \code{z.table}
compares each cell against the proportion for all columns in that row. When
testing against column margins \code{z.table} compares each cell against the
proportion for all rows in that  column.
}
