% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zed-scores.R
\name{z.table}
\alias{z.table}
\title{Calculate z-scores for a CrunchCube}
\usage{
z.table(table, margin)
}
\arguments{
\item{table}{A CrunchCube to calculate z-scores for}

\item{margin}{which margin to test against (1 for rows, 2 for columns)}
}
\value{
an array of z-scores
}
\description{
To get a z-score for the cells of a cube by comparing to the proportions on
the row or column margin. When testing against row marigns compare the cell
against the proprotion for all columns in that row. When testing against
column marigns compare the cell against the proprotion for all rows in that
column.
}
