% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{addGeoMetadata}
\alias{addGeoMetadata}
\title{Add geodata metadata to a crunch variable}
\usage{
addGeoMetadata(variable, data, match_field = "name")
}
\arguments{
\item{variable}{a Crunch variable to use for matching. This must be either
a text or a categorical variable.}

\item{data}{a Crunch dataset to use}

\item{match_field}{the field in the variable to use (default: "name")}
}
\value{
a CrunchGeography object that can be assigned into \code{geo(variable)}
}
\description{
Add geodata metadata to a crunch variable
}
\examples{
\dontrun{
geo(ds$state) <- addGeoMetadata("state", data=ds)
}
}
