% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R, R/slides.R
\docType{methods}
\name{analyses}
\alias{analyses}
\alias{analysis}
\alias{analysis<-}
\alias{query<-}
\alias{cube}
\alias{cubes}
\alias{analyses,CrunchSlide-method}
\alias{analysis,CrunchSlide-method}
\alias{analysis<-,CrunchSlide,formula-method}
\alias{query<-,CrunchSlide,ANY-method}
\alias{cubes,CrunchSlide-method}
\alias{cube,CrunchSlide-method}
\alias{cubes,AnalysisCatalog-method}
\alias{query<-,Analysis,formula-method}
\alias{cube,Analysis-method}
\title{Get and set slide analyses}
\usage{
analyses(x)

analysis(x)

analysis(x) <- value

query(x) <- value

cube(x)

cubes(x)

\S4method{analyses}{CrunchSlide}(x)

\S4method{analysis}{CrunchSlide}(x)

\S4method{analysis}{CrunchSlide,formula}(x) <- value

\S4method{query}{CrunchSlide,ANY}(x) <- value

\S4method{cubes}{CrunchSlide}(x)

\S4method{cube}{CrunchSlide}(x)

\S4method{cubes}{AnalysisCatalog}(x)

\S4method{query}{Analysis,formula}(x) <- value

\S4method{cube}{Analysis}(x)
}
\arguments{
\item{x}{A CrunchSlide, AnalysisCatalog, or Analysis}

\item{value}{for the setter, a query}
}
\description{
Slides are composed of analyses, which are effectively \code{CrunchCubes} with some
additional metadata. You can get and set a slide's Analysis Catalog with the
\code{analyses} method, and access an individual analysis with \code{analysis}.
}
\details{
You can get the CrunchCube from a slide or analysis with the \code{cube} method and
from a CrunchDeck with \code{cubes}. Analyses can be changed by assigning a formula
into the \code{query} function.
}
\examples{
\dontrun{
analysis(slide)
cube(slide)
cubes(deck)
query(slide) <- ~ cyl + wt
}
}
