% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-dataset.R
\name{appendDataset}
\alias{appendDataset}
\title{Append one Crunch dataset to another}
\usage{
appendDataset(dataset1, dataset2, autorollback = TRUE)
}
\arguments{
\item{dataset1}{a CrunchDataset}

\item{dataset2}{another CrunchDataset, or possibly a data.frame. If
\code{dataset2} is not a Crunch dataset, it will be uploaded as a new
dataset before appending. If it is a CrunchDataset, it may be subsetted with
a filter expression on the rows and a selection of variables on the columns.}

\item{autorollback}{logical: If the append fails, revert the dataset back
to its state before attempting to append? Default is \code{TRUE}, and you
probably won't want to change that.}
}
\value{
\code{dataset1}, updated with \code{dataset2}, potentially filtered on rows and
variables, appended to it.
}
\description{
Append one Crunch dataset to another
}
