% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-dataset.R
\name{appendDataset}
\alias{appendDataset}
\title{Append one Crunch dataset to another}
\usage{
appendDataset(dataset1, dataset2, cleanup = TRUE)
}
\arguments{
\item{dataset1}{a CrunchDataset}

\item{dataset2}{another CrunchDataset, or possibly a data.frame. If
\code{dataset2} is not a Crunch dataset, it will be uploaded as a new
dataset before appending.}

\item{cleanup}{logical: if the append operation fails or is aborted, should
the intermediate batch created on \code{dataset1} be deleted? Default is
\code{TRUE}; \code{FALSE} may be useful if you want to review the append
conflicts in the web application.}
}
\value{
A CrunchDataset with \code{dataset2} appended to \code{dataset1}
}
\description{
Append one Crunch dataset to another
}

