% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-dataset.R
\name{appendDataset}
\alias{appendDataset}
\title{Append one Crunch dataset to another}
\usage{
appendDataset(dataset1, dataset2, autorollback)
}
\arguments{
\item{dataset1}{a CrunchDataset}

\item{dataset2}{another CrunchDataset, or possibly a data.frame. If
\code{dataset2} is not a Crunch dataset, it will be uploaded as a new
dataset before appending. If it is a CrunchDataset, it may be subsetted with
a filter expression on the rows and a selection of variables on the columns.}

\item{autorollback}{Deprecated. This option no longer does anything.}
}
\value{
\code{dataset1}, updated with \code{dataset2}, potentially filtered on rows and
variables, appended to it.
}
\description{
Crunch allows you to add additional rows to a dataset by appending a second
dataset to the bottom of the original dataset. The app attempts to make intelligent
guesses about the two datasets to do things like match the variables in the two
datasets or harmonize the categories of categorical or array variables. However you
will get the best results if you ensure that the two datasets have the same variable names and types,
and that their categorical and array variables have consistent categories.
}
\details{
We recommend using the fork-edit-merge workflow to append datasets. To do this
fork the dataset with \code{forkDataset()} append the new data to the forked dataset,
ensure that the append worked as expected, and merge it back to the original dataset
with \code{mergeFork()}. For more see \code{vignette("fork-and-merge", package
= "crunch")}.
}
