% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{availableGeodataFeatures}
\alias{availableGeodataFeatures}
\title{Get the property features for available geographies}
\usage{
availableGeodataFeatures(
  x = getAPIRoot(),
  geodatum_fields = c("name", "description", "location")
)
}
\arguments{
\item{x}{an API root address (default: the R-session default)}

\item{geodatum_fields}{character, what pieces of information about each
geodatum should be retained? (default: `c("name", "description", "location")``)}
}
\value{
a dataframe with all of the available features and
geographies for matching
}
\description{
Get the property features for available geographies
}
