% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{c-categories}
\alias{c-categories}
\alias{c.Categories}
\alias{c.Category}
\title{S3 method to concatenate Categories and Category objects}
\usage{
\method{c}{Categories}(...)

\method{c}{Category}(...)
}
\arguments{
\item{...}{see \code{\link[base]{c}}}
}
\value{
An object of class \code{\link{Categories}}
}
\description{
S3 method to concatenate Categories and Category objects
}
\examples{
cat.a <- Category(name="First", id=1, numeric_value=1, missing=FALSE)
cat.b <- Category(name="Second", id=2)
cat.c <- Category(name="Third", id=3, missing=TRUE)
cats.1 <- Categories(cat.a, cat.b)
identical(cats.1, c(cat.a, cat.b))
identical(c(cats.1, cat.c), Categories(cat.a, cat.b, cat.c))
}
