% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-calculate.R
\name{calcInsertions}
\alias{calcInsertions}
\title{Given values from an array and transforms, calculate the insertions}
\usage{
calcInsertions(vec, elements, var_cats)
}
\arguments{
\item{vec}{values to transform (a single dimension of an array)}

\item{elements}{AbsCats of both categories and insertions to calculate.
Generally derived from \code{mapInsertions()}}

\item{var_cats}{the \code{Categories} object of the transform}
}
\value{
the values given in \code{vec}, with any insertions specified in
\code{trans} calculated
}
\description{
Given values from an array and transforms, calculate the insertions
}
\keyword{internal}
