% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-transforms.R
\name{calcInsertions}
\alias{calcInsertions}
\title{Given values from an array and transforms, calculate the insertions}
\usage{
calcInsertions(vec, inserts, var_cats)
}
\arguments{
\item{vec}{values to transform (typically a single dimension of an array)}

\item{var_cats}{the \code{Categories} object of the transform}

\item{inserst}{an \code{Insertions} object to pull transformations from}
}
\value{
the values given in \code{vec}, with any insertions specified in
\code{trans} calculated
}
\description{
Given values from an array and transforms, calculate the insertions
}
\keyword{internal}
