% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-catalog.R, R/filters.R, R/members.R, R/projects.R, R/shoji-catalog.R, R/teams.R, R/variable-catalog.R
\docType{methods}
\name{catalog-extract}
\alias{$,ShojiCatalog-method}
\alias{[,ShojiCatalog,ANY,ANY-method}
\alias{[,ShojiCatalog,character,ANY-method}
\alias{[,ShojiCatalog,logical,ANY-method}
\alias{[,ShojiCatalog,numeric,ANY-method}
\alias{[,VariableCatalog,VariableGroup,ANY-method}
\alias{[,VariableCatalog,VariableOrder,ANY-method}
\alias{[<-,VariableCatalog,VariableGroup,missing,VariableCatalog-method}
\alias{[<-,VariableCatalog,VariableOrder,missing,VariableCatalog-method}
\alias{[<-,VariableCatalog,character,missing,VariableCatalog-method}
\alias{[[,DatasetCatalog,ANY-method}
\alias{[[,DatasetCatalog,character-method}
\alias{[[,FilterCatalog,character-method}
\alias{[[,FilterCatalog,numeric-method}
\alias{[[,ProjectCatalog,ANY-method}
\alias{[[,ProjectCatalog,character-method}
\alias{[[,ShojiCatalog,ANY-method}
\alias{[[,ShojiCatalog,character-method}
\alias{[[,TeamCatalog,character-method}
\alias{[[,TeamCatalog,numeric-method}
\alias{[[,VariableCatalog,ANY-method}
\alias{[[,VariableCatalog,character-method}
\alias{[[<-,FilterCatalog,character,missing,CrunchFilter-method}
\alias{[[<-,FilterCatalog,character,missing,CrunchLogicalExpr-method}
\alias{[[<-,FilterCatalog,numeric,missing,CrunchFilter-method}
\alias{[[<-,FilterCatalog,numeric,missing,CrunchLogicalExpr-method}
\alias{[[<-,MemberCatalog,ANY,missing,ANY-method}
\alias{[[<-,MemberCatalog,character,missing,NULL-method}
\alias{[[<-,ProjectCatalog,character,missing,CrunchProject-method}
\alias{[[<-,ProjectCatalog,character,missing,list-method}
\alias{[[<-,TeamCatalog,character,missing,CrunchTeam-method}
\alias{[[<-,TeamCatalog,character,missing,list-method}
\alias{[[<-,VariableCatalog,character,missing,CrunchVariable-method}
\alias{[[<-,VariableCatalog,character,missing,VariableTuple-method}
\alias{catalog-extract}
\title{Extract and modify subsets of Catalog-type objects}
\usage{
\S4method{[[}{DatasetCatalog,character}(x, i, j, ...)

\S4method{[[}{DatasetCatalog,ANY}(x, i, j, ...)

\S4method{[[}{FilterCatalog,character}(x, i, j, ...)

\S4method{[[}{FilterCatalog,numeric}(x, i, j, ...)

\S4method{[[}{FilterCatalog,character,missing,CrunchLogicalExpr}(x, i, j) <-
  value

\S4method{[[}{FilterCatalog,numeric,missing,CrunchLogicalExpr}(x, i, j) <- value

\S4method{[[}{FilterCatalog,character,missing,CrunchFilter}(x, i, j) <- value

\S4method{[[}{FilterCatalog,numeric,missing,CrunchFilter}(x, i, j) <- value

\S4method{[[}{MemberCatalog,ANY,missing,ANY}(x, i, j) <- value

\S4method{[[}{MemberCatalog,character,missing,`NULL`}(x, i, j) <- value

\S4method{[[}{ProjectCatalog,character}(x, i, j, ...)

\S4method{[[}{ProjectCatalog,ANY}(x, i, j, ...)

\S4method{[[}{ProjectCatalog,character,missing,list}(x, i, j) <- value

\S4method{[[}{ProjectCatalog,character,missing,CrunchProject}(x, i, j) <- value

\S4method{[}{ShojiCatalog,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ShojiCatalog,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ShojiCatalog,logical,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ShojiCatalog,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{ShojiCatalog,ANY}(x, i, j, ...)

\S4method{[[}{ShojiCatalog,character}(x, i, j, ...)

\S4method{$}{ShojiCatalog}(x, name)

\S4method{[[}{TeamCatalog,character}(x, i, j, ...)

\S4method{[[}{TeamCatalog,numeric}(x, i, j, ...)

\S4method{[[}{TeamCatalog,character,missing,list}(x, i, j) <- value

\S4method{[[}{TeamCatalog,character,missing,CrunchTeam}(x, i, j) <- value

\S4method{[[}{VariableCatalog,character}(x, i, j, ...)

\S4method{[[}{VariableCatalog,ANY}(x, i, j, ...)

\S4method{[[}{VariableCatalog,character,missing,VariableTuple}(x, i, j) <- value

\S4method{[[}{VariableCatalog,character,missing,CrunchVariable}(x, i, j) <-
  value

\S4method{[}{VariableCatalog,VariableOrder,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{VariableCatalog,VariableGroup,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{VariableCatalog,character,missing,VariableCatalog}(x, i, j) <-
  value

\S4method{[}{VariableCatalog,VariableOrder,missing,VariableCatalog}(x, i, j) <-
  value

\S4method{[}{VariableCatalog,VariableGroup,missing,VariableCatalog}(x, i, j) <-
  value
}
\arguments{
\item{x}{a Catalog object}

\item{i}{which catalog elements to extract}

\item{j}{Invalid}

\item{...}{additional arguments}

\item{value}{For updating, an object of the appropriate class and size to
insert}

\item{drop}{Invalid}

\item{name}{for \code{$}, the same as \code{i} for \code{[[}}
}
\value{
A subset of \code{x} if extracting, otherwise \code{x} duly modified
}
\description{
Extract and modify subsets of Catalog-type objects
}

