% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shoji-catalog.R
\docType{methods}
\name{catalog-length}
\alias{catalog-length}
\title{Length of Catalog}
\usage{
\S4method{length}{ShojiCatalog}(x)
}
\arguments{
\item{x}{a Catalog}
}
\value{
Integer: the number of elements in the index list
}
\description{
Length of Catalog
}

