% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-catalog.R
\name{catalogToDataFrame}
\alias{catalogToDataFrame}
\title{Utility to get a more human-readable view of a Shoji Catalog}
\usage{
catalogToDataFrame(x, keys = TRUE, rownames, ...)
}
\arguments{
\item{x}{ShojiCatalog or subclass}

\item{keys}{character vector of attribute names from each catalog tuple to
include in the result. Default is TRUE, which means all.}

\item{rownames}{See \code{\link[base:data.frame]{base::data.frame}}, the \code{row.names}
argument, to which this is passed in \code{data.frame}. The difference here
is that if \code{rownames} is explicitly set as \code{NULL}, the resulting
object will not have row names set. By default, row names will be the URLs
of the catalog tuples.}

\item{...}{additional arguments passed to \code{data.frame}}
}
\value{
a \code{data.frame} view of the catalog
}
\description{
Utility to get a more human-readable view of a Shoji Catalog
}
