% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-catalog.R
\name{catalogToDataFrame}
\alias{catalogToDataFrame}
\title{A utility to retun a dataframe from a ShojiCatalog.}
\usage{
catalogToDataFrame(x, keys = TRUE, rownames = "default",
  list_columns = c("subvariables", "subvariables_catalog"), ...)
}
\arguments{
\item{x}{ShojiCatalog or subclass}

\item{keys}{character vector of attribute names from each catalog tuple to
include in the result. Default is TRUE, which means all.}

\item{rownames}{The rownames of the resulting dataframe. If set to NULL rownames
will default to the resulting dataframe will not have row names. By default the row names
will be the URLs of the catalog tuples.}

\item{list_columns}{A character vector of the names of the attributes which should be stored
in a list-column.}

\item{...}{additional arguments passed to \code{data.frame}}
}
\value{
a \code{data.frame} view of the catalog
}
\description{
Some of the attributes of a ShojiCatalog will not naturally fit in
a conventional dataframe. For instance a single variable might have multiple subvariables,
and these subvariables will not fit in a single row of a dataframe. In this case the
list of subvariables are stored in a list-column in the resulting dataframe.
}
