\name{Categories}
\alias{Categories}
\alias{Category}
\alias{ids}
\alias{values}
\title{Categories in CategoricalVariables}
\usage{
  Categories(...)

  Category(...)
}
\description{
  CategoricalVariables, as well as the array types composed
  from Categoricals, contain Categories. Categories are a
  subclass of list that contains only Category objects.
  Category objects themselves subclass list and contain the
  following fields: "name", "id", "numeric_value",
  "missing", and optionally "selected".
}

