% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R, R/categories.R
\docType{class}
\name{Categories-class}
\alias{Categories}
\alias{Categories-class}
\alias{Category}
\alias{Category-class}
\alias{[,Categories,ANY,ANY-method}
\alias{[<-,Categories,ANY,ANY,ANY-method}
\alias{ids}
\alias{ids,Categories-method}
\alias{ids<-}
\alias{ids<-,Categories-method}
\alias{names,Categories-method}
\alias{names<-,Categories-method}
\alias{values}
\alias{values,Categories-method}
\alias{values<-}
\alias{values<-,Categories-method}
\title{Categories in CategoricalVariables}
\usage{
\S4method{[}{Categories,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{Categories,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{names}{Categories}(x)

\S4method{values}{Categories}(x)

\S4method{ids}{Categories}(x)

\S4method{names}{Categories}(x) <- value

\S4method{values}{Categories}(x) <- value

\S4method{ids}{Categories}(x) <- value
}
\arguments{
\item{x}{For the attribute getters and setters, an object of class
Category or Categories}

\item{i}{For the [ methods, just as with list extract methods}

\item{j}{Invalid argument to [, but in the generic's signature}

\item{...}{additional arguments to [, ignored}

\item{drop}{Invalid argument to [, but in the generic's signature}

\item{value}{For [<-, the replacement Category to insert}
}
\description{
CategoricalVariables, as well as the array types composed from
Categoricals, contain Categories. Categories are a subclass of list that
contains only Category objects. Category objects themselves subclass list
and contain the following fields: "name", "id", "numeric_value", "missing",
and optionally "selected".
}

