% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Categories-class}
\alias{Categories}
\alias{Categories-class}
\alias{Category}
\alias{Category-class}
\alias{ids,Categories-method}
\alias{values,Categories-method}
\title{Categories in CategoricalVariables}
\usage{
\S4method{values}{Categories}(x)

\S4method{ids}{Categories}(x)
}
\description{
CategoricalVariables, as well as the array types composed from
Categoricals, contain Categories. Categories are a subclass of list that
contains only Category objects. Category objects themselves subclass list
and contain the following fields: "name", "id", "numeric_value", "missing", and optionally "selected".
}

