% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-categories.R
\name{combine}
\alias{combine}
\title{Combine categories or responses}
\usage{
combine(variable, combinations = list(), ...)
}
\arguments{
\item{variable}{Categorical, Categorical Array, or Multiple Response
variable}

\item{combinations}{list of named lists containing (1) "categories":
category ids or names for categorical types, or for multiple response,
"responses": subvariable names, aliases, or positional indices; (2) a
"name" for the new category or response; and (3) optionally, other category
("missing", "numeric_value") or subvariable ("alias", "description")
attributes. If \code{combinations} is omitted, the resulting variable will
essentially be a copy (but see \code{link{copy}} for a more natural way to
do that, if desired).}

\item{...}{Additional variable metadata for the new derived variable}
}
\value{
A \code{\link{VariableDefinition}} that will create the new
comined-category or -response derived variable. Categories/responses not
referenced in \code{combinations} will be appended to the end of the
combinations.
}
\description{
Combine categories or responses
}

