% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional-transform.R
\name{conditionalTransform}
\alias{conditionalTransform}
\title{Conditional transformation}
\usage{
conditionalTransform(cases, values)
}
\arguments{
\item{cases}{a list of cases to evaluate}

\item{values}{a list of values to substitute when the corresponding case is true}
}
\description{
Conditional transformation
}
