% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{copyOrder}
\alias{copyOrder}
\title{Copy the variable order from one dataset to another.}
\usage{
copyOrder(source, target)
}
\arguments{
\item{source}{the dataset you want to copy the order from}

\item{target}{the dataset you want to copy the order to}
}
\value{
returns an object of class \code{\link{VariableOrder}} (which can be assigned
to a dataset with \code{\link{ordering}})
}
\description{
Copy the variable order from one dataset to another.
}
\examples{
\dontrun{
ordering(ds) <- copyOrder(ds1, ds)
}
}
