% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{copyOrderGroup}
\alias{copyOrderGroup}
\title{Copy the order of a \code{VariableGroup} (or individual variable url) from \code{VariableOrder}}
\usage{
copyOrderGroup(gr, source, target)
}
\arguments{
\item{gr}{the group or variable url to be copied}

\item{source}{the dataset you want to copy the order from}

\item{target}{the dataset you want to copy the order to}
}
\value{
returns either a \code{\link{VariableGroup}} (if a group is supplied) or a url (if just a variable url is supplied)
}
\description{
Copy the order of a \code{VariableGroup} (or individual variable url) from \code{VariableOrder}
}
\examples{
\dontrun{
ordering(ds) <- copyOrder(ds1, ds)
}
}
\keyword{internal}
