% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lake_data.R
\name{cr_read_data}
\alias{cr_read_data}
\title{Read downloaded Crunch Lake data files into a 'tidy' format}
\usage{
cr_read_data(
  x,
  as_crunch_variables = TRUE,
  array_strategy = c("packed", "qualified", "unqualified"),
  name_repair = "check_unique",
  ...
)
}
\arguments{
\item{x}{A folder path or destination from \code{\link[=cr_export]{cr_export()}}}

\item{as_crunch_variables}{Whether to load the data as special data structures with
crunch metadata (defaults to \code{TRUE})}

\item{array_strategy}{One of "packed" for packed data.frame columns to represent
array variables, "qualified" to have flattened representations of the array variables
with the parent array's alias before the axis item names to distinguish items with the
same axis value but from another array, or "unqualified" which only uses the array
axis item.}

\item{name_repair}{Passed to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} to fix column name collisions
that can arise due to array variables. Defaults to "check_unique", which gives an error
when finding duplicate names.}

\item{...}{Saved for future expansion}
}
\description{
Work-in-progress R representation of the lake data. See \code{\link[=cr_read_data_long]{cr_read_data_long()}} to
read the data in the format it is sent.
}
