% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-dataset.R
\name{createDataset}
\alias{createDataset}
\title{Create an empty dataset}
\usage{
createDataset(name, body, ...)
}
\arguments{
\item{name}{character, the name to give the new Crunch dataset. This is
required.}

\item{body}{list correctly formatted metadata definition for a dataset. See
docs.crunch.io.}

\item{...}{additional arguments for the POST to create the dataset, such as
"description".}
}
\value{
An object of class CrunchDataset.
}
\description{
Use only if you're writing a function to create a Crunch dataset from a
custom data structure. If you have a data.frame, just call
\code{\link{newDataset}} on it.
}
\seealso{
\code{\link{newDataset}}
}
\keyword{internal}

