% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/new-dataset.R
\name{createDataset}
\alias{createDataset}
\title{Create an empty dataset}
\usage{
createDataset(name, useAlias = default.useAlias(), ...)
}
\arguments{
\item{name}{character, the name to give the new Crunch dataset. This is
required.}

\item{useAlias}{logical whether variable alias or name should be used as R
variable names when the dataset is returned. Default is TRUE, meaning alias.
This argument is only relevant for the dataset object that is returned;
it has no effect on the contents of the object created on the server.}

\item{...}{additional arguments for the POST to create the dataset, such as
"description".}
}
\value{
An object of class CrunchDataset.
}
\description{
Use only if you're writing a function to create a Crunch dataset from a
custom data structure. If you have a data.frame, just call
\code{\link{newDataset}} on it.
}

