\name{is.dataset}
\alias{is.categoricalVariable}
\alias{is.dataset}
\alias{is.numericVariable}
\alias{is.textVariable}
\alias{is.variable}
\title{Is it?}
\usage{
  is.dataset(x)

  is.variable(x)

  is.numericVariable(x)

  is.categoricalVariable(x)

  is.textVariable(x)
}
\arguments{
  \item{x}{an object}
}
\value{
  logical
}
\description{
  Is it?
}

