% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/expressions.R, R/geo.R, R/shoji.R,
%   R/variable-derivation.R, R/variable-type.R
\docType{methods}
\name{is.dataset}
\alias{is.dataset}
\alias{is.CrunchExpr}
\alias{is.Expr}
\alias{is.Geodata}
\alias{is.shoji}
\alias{is.derived,CrunchVariable-method}
\alias{is.derived}
\alias{is.derived<-,CrunchVariable,logical-method}
\alias{is.derived<-}
\alias{is.variable}
\alias{is.Numeric}
\alias{is.Categorical}
\alias{is.Text}
\alias{is.Datetime}
\alias{is.Multiple}
\alias{is.MR}
\alias{is.MultipleResponse}
\alias{is.CA}
\alias{is.Array}
\alias{is.CategoricalArray}
\title{Is it?}
\usage{
is.dataset(x)

is.CrunchExpr(x)

is.Expr(x)

is.Geodata(x)

is.shoji(x)

\S4method{is.derived}{CrunchVariable}(x)

\S4method{is.derived}{CrunchVariable,logical}(x) <- value

is.variable(x)

is.Numeric(x)

is.Categorical(x)

is.Text(x)

is.Datetime(x)

is.Multiple(x)

is.MR(x)

is.MultipleResponse(x)

is.CA(x)

is.Array(x)

is.CategoricalArray(x)
}
\arguments{
\item{x}{an object}

\item{value}{a value to assign}
}
\value{
logical
}
\description{
Is it?
}
