\name{is.dataset}
\alias{is.Categorical}
\alias{is.Numeric}
\alias{is.Text}
\alias{is.dataset}
\alias{is.variable}
\title{Is it?}
\usage{
  is.dataset(x)

  is.variable(x)

  is.Numeric(x)

  is.Categorical(x)

  is.Text(x)
}
\arguments{
  \item{x}{an object}
}
\value{
  logical
}
\description{
  Is it?
}

