% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-definition.R, R/automation.R,
%   R/dataset.R, R/expressions.R, R/geo.R, R/shoji.R, R/variable-type.R
\name{is.VariableDefinition}
\alias{is.VariableDefinition}
\alias{is.VarDef}
\alias{is.script}
\alias{is.dataset}
\alias{is.CrunchExpr}
\alias{is.Expr}
\alias{is.Geodata}
\alias{is.shoji}
\alias{is.variable}
\alias{is.Numeric}
\alias{is.Categorical}
\alias{is.Text}
\alias{is.Datetime}
\alias{is.Multiple}
\alias{is.MR}
\alias{is.MultipleResponse}
\alias{is.CA}
\alias{is.Array}
\alias{is.CategoricalArray}
\title{Test whether a Crunch object belongs to a class}
\usage{
is.VariableDefinition(x)

is.VarDef(x)

is.script(x)

is.dataset(x)

is.CrunchExpr(x)

is.Expr(x)

is.Geodata(x)

is.shoji(x)

is.variable(x)

is.Numeric(x)

is.Categorical(x)

is.Text(x)

is.Datetime(x)

is.Multiple(x)

is.MR(x)

is.MultipleResponse(x)

is.CA(x)

is.Array(x)

is.CategoricalArray(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical
}
\description{
Test whether a Crunch object belongs to a class
}
