% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/univariate.R
\docType{methods}
\name{mean,CrunchVariable-method}
\alias{max}
\alias{max,CrunchVariable-method}
\alias{max,DatetimeVariable-method}
\alias{max,NumericVariable-method}
\alias{mean}
\alias{mean,CrunchVariable-method}
\alias{mean,NumericVariable-method}
\alias{median}
\alias{median,CrunchVariable-method}
\alias{median,NumericVariable-method}
\alias{min}
\alias{min,CrunchVariable-method}
\alias{min,DatetimeVariable-method}
\alias{min,NumericVariable-method}
\alias{sd}
\alias{sd,CrunchVariable-method}
\alias{sd,NumericVariable-method}
\title{Univariate statistics on Crunch objects}
\usage{
\S4method{mean}{CrunchVariable}(x, ...)

\S4method{sd}{CrunchVariable}(x, na.rm = FALSE)

\S4method{median}{CrunchVariable}(x, na.rm = FALSE)

\S4method{min}{CrunchVariable}(x, na.rm)

\S4method{max}{CrunchVariable}(x, na.rm)

\S4method{mean}{NumericVariable}(x, ...)

\S4method{median}{NumericVariable}(x, na.rm = FALSE)

\S4method{sd}{NumericVariable}(x, na.rm = FALSE)

\S4method{min}{NumericVariable}(x, na.rm = FALSE)

\S4method{max}{NumericVariable}(x, na.rm = FALSE)

\S4method{min}{DatetimeVariable}(x, na.rm = FALSE)

\S4method{max}{DatetimeVariable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a NumericVariable, or for \code{min} and \code{max}, possibly a
DatetimeVariable}

\item{...}{additional arguments to \code{mean}}

\item{na.rm}{logical: exclude missings?}
}
\description{
Univariate statistics on Crunch objects
}
\seealso{
\code{\link[base]{mean}} \code{\link[stats]{sd}} \code{\link[stats]{median}} \code{\link[base]{min}} \code{\link[base]{max}}
}

