\name{crunchAuth}
\alias{crunchAuth}
\title{Validate authentication inputs and then POST to the API}
\usage{
  crunchAuth(email, password = NULL, ...)
}
\arguments{
  \item{email}{character, see \code{\link{login}}}

  \item{password}{character, see \code{\link{login}}}

  \item{...}{see \code{\link{login}}}
}
\description{
  Validate authentication inputs and then POST to the API
}

