% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CrunchVariable-class}
\alias{CategoricalArrayVariable}
\alias{CategoricalArrayVariable-class}
\alias{CategoricalVariable}
\alias{CategoricalVariable-class}
\alias{CrunchVariable}
\alias{CrunchVariable-class}
\alias{DatetimeVariable}
\alias{DatetimeVariable-class}
\alias{MultipleResponseVariable}
\alias{MultipleResponseVariable-class}
\alias{NumericVariable}
\alias{NumericVariable-class}
\alias{TextVariable}
\alias{TextVariable-class}
\title{Variables in Crunch}
\description{
Variables are S4 objects. All inherit from the base class
\code{CrunchVariable}.
}
\section{Slots}{

\describe{
\item{\code{readonly}}{logical: should changes made to this variable object locally
be persisted on the server? Default is \code{FALSE}}

\item{\code{tuple}}{An object of class VariableTuple. These contain attributes, such
as name and description, that are found in the index of the \code{\link{VariableCatalog}}}
}}

