% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CrunchVariable-class}
\alias{CrunchVariable-class}
\alias{NumericVariable-class}
\alias{NumericVariable}
\alias{CategoricalVariable-class}
\alias{CategoricalVariable}
\alias{TextVariable-class}
\alias{TextVariable}
\alias{DatetimeVariable-class}
\alias{DatetimeVariable}
\alias{CategoricalArrayVariable-class}
\alias{CategoricalArrayVariable}
\alias{MultipleResponseVariable-class}
\alias{MultipleResponseVariable}
\title{Variables in Crunch}
\description{
Variables are S4 objects. All inherit from the base class
\code{CrunchVariable}.
be persisted on the server? Default is \code{FALSE}
}
\section{Slots}{

\describe{
\item{\code{filter}}{either \code{NULL} or \code{CrunchLogicalExpr}}
}}

