\name{CrunchVariable}
\alias{CategoricalArrayVariable}
\alias{CategoricalVariable}
\alias{CrunchVariable}
\alias{DatetimeVariable}
\alias{MultipleResponseVariable}
\alias{NumericVariable}
\alias{TextVariable}
\title{Variables in Crunch}
\usage{
  CrunchVariable(...)

  NumericVariable(...)

  CategoricalVariable(...)

  TextVariable(...)

  DatetimeVariable(...)

  CategoricalArrayVariable(...)

  MultipleResponseVariable(...)
}
\description{
  Variables are S4 objects. All inherit from the base class
  \code{CrunchVariable}.
}

