% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cube-result.R
\docType{methods}
\name{cube-computing}
\alias{cube-computing}
\alias{margin.table}
\alias{margin.table,CrunchCube-method}
\alias{prop.table}
\alias{prop.table,CrunchCube-method}
\alias{round,CrunchCube-method}
\title{Work with CrunchCubes}
\usage{
\S4method{margin.table}{CrunchCube}(x, margin = NULL)

\S4method{prop.table}{CrunchCube}(x, margin = NULL)

\S4method{round}{CrunchCube}(x, digits = 0)
}
\arguments{
\item{x}{a CrunchCube}

\item{margin}{index, or vector of indices to generate margin for. See
\code{\link[base]{prop.table}}}

\item{digits}{see \code{\link[base]{round}}}
}
\value{
The appropriate margin.table or prop.table.
}
\description{
Crunch.io supports more complex data types than base R does, such as
multiple response and array types. If you want to compute margin or
proportion tables on an aggregation of these variable types, special methods
are required. These functions provide an interface like
\code{\link[base]{margin.table}} and \code{\link[base]{prop.table}} for
the CrunchCube object, handling those special data types.
}
\seealso{
\code{\link[base]{margin.table}} \code{\link[base]{prop.table}}
}

