% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cube-dims.R
\docType{methods}
\name{cube-methods}
\alias{cube-methods}
\alias{dim,CubeDims-method}
\alias{dimnames,CubeDims-method}
\alias{is.na,CubeDims-method}
\title{Methods on Cube objects}
\usage{
\S4method{dimnames}{CubeDims}(x)

\S4method{dim}{CubeDims}(x)

\S4method{is.na}{CubeDims}(x)
}
\arguments{
\item{x}{a CrunchCube or its CubeDims component.}
}
\value{
Generally, the same shape of result that each of these functions
return when applied to an \code{array} object.
}
\description{
These methods provide an \code{array}-like interface to the CrunchCube
object.
}
\seealso{
\code{\link{cube-computing}}
}

