% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\docType{methods}
\name{cut,NumericVariable-method}
\alias{cut,NumericVariable-method}
\title{Cut a numeric Crunch variable}
\usage{
\S4method{cut}{NumericVariable}(x, breaks, variableName, labels = NULL,
  include.lowest = FALSE, right = TRUE, dig.lab = 3,
  ordered_result = FALSE, ...)
}
\arguments{
\item{x}{A crunch variable of the class NumericVariable}

\item{breaks}{Either a numeric vector of two or more unique cut points
or a single number (greater than or equal to 2) giving the number of intervals
into which x is to be cut.}

\item{variableName}{The name of the resulting case variable as a character string.}

\item{labels}{labels for the levels of the resulting category.
By default, labels are constructed using interval notation.
If labels = FALSE, simple integer codes are returned instead of a factor.}

\item{include.lowest}{logical, indicating if an \code{x[i]} equal to the lowest
(or highest, for right = FALSE) \code{breaks} value should be included.}

\item{right}{logical, indicating if the intervals should be closed on the right
(and open on the left) or vice versa.}

\item{dig.lab}{integer which is used when labels are not given.
It determines the number of digits used in formatting the break numbers.}

\item{ordered_result}{Ignored.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a Crunch VariableDefinition
}
\description{
crunch::cut() is equivalent to base::cut() except that it operates on
crunch variables instead of in-memory variables. The function divides the range of
x into intervals and codes the values in x according to which interval they fall.
The leftmost interval corresponds to level one, the next leftmost to level two and
so on.
}
\examples{
\dontrun{
ds <- loadDataset("mtcars")
ds$cat_var <- cut(ds$mpg, 3, variableName = "new_var")
}
 
}
