% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dataset.R, R/expressions.R,
%   R/variable.R
\docType{methods}
\name{datasetReference}
\alias{datasetReference}
\alias{datasetReference,CrunchDataset-method}
\alias{datasetReference,CrunchExpr-method}
\alias{datasetReference,CrunchVariable-method}
\alias{datasetReference,character-method}
\alias{datasetReference,ANY-method}
\title{Get a Crunch object's dataset}
\usage{
datasetReference(x)

\S4method{datasetReference}{CrunchDataset}(x)

\S4method{datasetReference}{CrunchExpr}(x)

\S4method{datasetReference}{CrunchVariable}(x)

\S4method{datasetReference}{character}(x)

\S4method{datasetReference}{ANY}(x)
}
\arguments{
\item{x}{a Crunch object}
}
\value{
The url of the dataset which contains that object
}
\description{
Get a Crunch object's dataset
}
\keyword{internal}
