% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.data.frame.CrunchDataset}
\alias{as.data.frame.CrunchDataset}
\title{as.data.frame method for CrunchDataset}
\usage{
\method{as.data.frame}{CrunchDataset}(x, row.names = NULL, optional = FALSE,
  ...)
}
\description{
as.data.frame method for CrunchDataset
}

