% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variable-to-R.R
\name{as.data.frame.CrunchDataset}
\alias{as.data.frame.CrunchDataset}
\title{as.data.frame method for CrunchDataset}
\usage{
\method{as.data.frame}{CrunchDataset}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{a CrunchDataset}

\item{row.names}{part of as.data.frame signature. Ignored.}

\item{optional}{part of as.data.frame signature. Ignored.}

\item{...}{additional arguments passed to as.data.frame.default}
}
\value{
a data.frame
}
\description{
as.data.frame method for CrunchDataset
}

