% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-datasets.R, R/projects.R
\name{datasets}
\alias{datasets}
\alias{datasets<-}
\title{Get the dataset catalog}
\usage{
datasets(x = getAPIRoot())

datasets(x) <- value
}
\arguments{
\item{x}{a \code{ShojiObject}, such as a \code{CrunchProject}. If omitted,
the function will load the user's primary dataset catalog. #'}

\item{value}{\code{CrunchDataset} for the setter}
}
\value{
An object of class \code{DatasetCatalog}. The setter returns the
project (or other object that contains a dataset catalog with the given
dataset added to it (via changing its owner to be
the specified object, \code{x}).
}
\description{
Crunch datasets are associated with catalogs. For instance a project catalog will
have a set of datasets associated with it, as will a user or team. This function
allows you to get or set the datasets associated with a catalog.
}
\examples{
\dontrun{
# Get the primary dataset catalog
mydatasets <- datasets()
# Can load a dataset from that
ds <- loadDataset(mydatasets[["Dataset name"]])
# Can use the same function to get the dataset catalog for a project
proj <- projects()[["Project name"]]
projdatasets <- datasets(proj)
# The assignment method lets you move a dataset to a project
datasets(proj) <- ds
}
}
