% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R, R/slides.R
\name{titles}
\alias{titles}
\alias{titles<-}
\alias{title}
\alias{title<-}
\alias{subtitles}
\alias{subtitles<-}
\alias{subtitle}
\alias{subtitle<-}
\alias{titles,CrunchDeck-method}
\alias{titles<-,CrunchDeck-method}
\alias{subtitles,CrunchDeck-method}
\alias{subtitles<-,CrunchDeck-method}
\alias{titles,SlideCatalog-method}
\alias{titles<-,SlideCatalog-method}
\alias{subtitles,SlideCatalog-method}
\alias{subtitles<-,SlideCatalog-method}
\alias{title,CrunchSlide-method}
\alias{title<-,CrunchSlide-method}
\alias{subtitle,CrunchSlide-method}
\alias{subtitle<-,CrunchSlide-method}
\title{Manipulate deck titles}
\usage{
titles(x)

titles(x) <- value

title(x)

title(x) <- value

subtitles(x, value)

subtitles(x) <- value

subtitle(x, value)

subtitle(x) <- value

\S4method{titles}{CrunchDeck}(x)

\S4method{titles}{CrunchDeck}(x) <- value

\S4method{subtitles}{CrunchDeck}(x)

\S4method{subtitles}{CrunchDeck}(x) <- value

\S4method{titles}{SlideCatalog}(x)

\S4method{titles}{SlideCatalog}(x) <- value

\S4method{subtitles}{SlideCatalog}(x)

\S4method{subtitles}{SlideCatalog}(x) <- value

\S4method{title}{CrunchSlide}(x)

\S4method{title}{CrunchSlide}(x) <- value

\S4method{subtitle}{CrunchSlide}(x)

\S4method{subtitle}{CrunchSlide}(x) <- value
}
\arguments{
\item{x}{a \code{CrunchDeck} or \code{CrunchSlide}}

\item{value}{character, the new title or subtitle}
}
\value{
\code{x}, modified
}
\description{
Crunch slides have titles and subtitles. You can change these features at
either the deck level by assigning a character vector which is
the same length as the deck to the CrunchDeck, or by assigning character strings to the
the slide.
}
\examples{
\dontrun{
titles(deck)
titles(deck) <- c(new_title1, new_title2)
slide <- deck[[1]]
title(slide) <- "new title"
subtitle(slide) <- "new subtitle"
subtitles(deck)
}
}
