% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-order.R
\name{dedupeOrder}
\alias{dedupeOrder}
\title{Remove duplicated entities from an order/group}
\usage{
dedupeOrder(x)
}
\arguments{
\item{x}{VariableOrder, DatasetOrder, VariableGroup, or DatasetGroup}
}
\value{
\code{x}
}
\description{
This function no longer does anything because variables can no longer appear
in more than one folder. It is deprecated and scheduled for removal.
}
\seealso{
\code{\link{duplicates}}, which when set to \code{FALSE} also calls this function.
}
