% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{delete,CrunchDataset-method}
\alias{delete}
\alias{delete,ANY-method}
\alias{delete,CrunchDataset-method}
\alias{delete,ShojiObject-method}
\title{Delete a Crunch object from the server}
\usage{
\S4method{delete}{CrunchDataset}(x, confirm = interactive() | is.readonly(x),
  ...)

\S4method{delete}{ShojiObject}(x, ...)

\S4method{delete}{ANY}(x, ...)
}
\arguments{
\item{x}{a Crunch object}

\item{confirm}{logical: should the user be asked to confirm deletion.
Option available for datasets only. Default is \code{TRUE} if in an
interactive session.}

\item{...}{additional arguments, in the generic}
}
\description{
These methods delete entites, notably Datasets and Variables within them,
from the server. This action is permanent and cannot be undone, so it
should not be done lightly. Consider instead using \code{\link{archive}}
for datasets and \code{\link{hide}} for variables
}
\seealso{
archive hide deleteDataset
}

