% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-datasets.R
\name{deleteDataset}
\alias{deleteDataset}
\title{Delete a dataset from the dataset list}
\usage{
deleteDataset(x, ...)
}
\arguments{
\item{x}{The name (character) of a dataset, its (numeric) position in the
return of \code{\link{listDatasets}}, or an object of class
\code{CrunchDataset}. x can only be of length 1--this function is not 
vectorized (for your protection).}

\item{...}{additional parameters (such as \code{confirm}) passed to 
\code{delete}}
}
\value{
(Invisibly) the API response from deleting the dataset
}
\description{
This function lets you delete a dataset without first loading it. If you
have a dataset that somehow is corrupted and won't load, you can delete it
this way.
}
\details{
The function also works on CrunchDataset objects, just like
\code{\link{delete}}, which may be useful if you have loaded another 
package that masks the \code{delete} method.
}
\seealso{
\code{\link{delete}}
}

