% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subvariables.R
\name{deleteSubvariables}
\alias{deleteSubvariable}
\alias{deleteSubvariables}
\title{Delete subvariables from an array}
\usage{
deleteSubvariables(variable, to.delete)

deleteSubvariable(variable, to.delete)
}
\arguments{
\item{variable}{the array variable}

\item{to.delete}{character alias(es) or name(s), depending on the value of
\code{getOption("crunch.namekey.array")}, of the subvariables to delete.}
}
\value{
a new version of variable without the indicated subvariables
}
\description{
This function conceals the dirty work in making this happen. The array
gets unbound, the subvariables deleted, and then the remaining subvariable
are rebound into a new array.
}

