% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-delete.R
\name{deleteSubvariables}
\alias{deleteSubvariables}
\alias{deleteSubvariable}
\title{Delete subvariables from an array}
\usage{
deleteSubvariables(variable, to.delete)

deleteSubvariable(variable, to.delete)
}
\arguments{
\item{variable}{the array variable}

\item{to.delete}{aliases (following \code{crunch.namekey.dataset}) or indices
of variables to delete.}
}
\value{
a new version of variable without the indicated subvariables
}
\description{
Deleting variables requires confirmation. In an interactive session, you will be asked
to confirm. To avoid that prompt, or to delete subvariables from a
non-interactive session, wrap the call in \code{\link[=with_consent]{with_consent()}} to give
your permission to delete.
}
\details{
To delete the subvariables the function unbinds the array, deletes the subvariable, and
then binds the remaining subvariables into a new array.
}
