% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-variable.R
\name{deleteVariables}
\alias{deleteVariable}
\alias{deleteVariables}
\title{Delete Variables Within a Dataset}
\usage{
deleteVariables(dataset, variables, confirm = requireConsent())

deleteVariable(dataset, variables, confirm = requireConsent())
}
\arguments{
\item{dataset}{the Dataset to modify}

\item{variables}{aliases (following \code{crunch.namekey.dataset}) or indices
of variables to delete.}

\item{confirm}{logical: should the user be asked to confirm deletion.
Default is \code{TRUE} if in
an interactive session. You can avoid the confirmation prompt if you delete
\code{with(\link{consent})}.}
}
\value{
(invisibly) \code{dataset} with the specified variables deleted
}
\description{
Delete Variables Within a Dataset
}
\seealso{
\code{\link{hide}}
}

