% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{deleteVariables}
\alias{deleteVariables}
\alias{deleteVariable}
\title{Delete Variables Within a Dataset}
\usage{
deleteVariables(dataset, variables)

deleteVariable(dataset, variables)
}
\arguments{
\item{dataset}{the Dataset to modify}

\item{variables}{aliases (following \code{crunch.namekey.dataset}) or indices
of variables to delete.}
}
\value{
(invisibly) \code{dataset} with the specified variables deleted
}
\description{
This function permanently deletes a variable from a dataset. For a non-destructive
alternative see \code{\link[=hide]{hide()}}.
}
\details{
In an interactive session, you will be prompted to confirm that you
wish to delete the variable. To avoid that prompt, or to delete variables from a
non-interactive session, wrap the call in \code{\link[=with_consent]{with_consent()}} to give
your permission to delete.
}
\seealso{
\code{\link[=hide]{hide()}} \code{\link[=delete]{delete()}} \code{\link[=deleteSubvariable]{deleteSubvariable()}}
}
