% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide-variables.R
\name{deleteVariables}
\alias{deleteVariable}
\alias{deleteVariables}
\title{Delete Variables Within a Dataset}
\usage{
deleteVariables(dataset, variables = NULL, pattern = NULL,
  key = namekey(dataset), confirm = requireConsent(), ...)

deleteVariable(dataset, variables = NULL, pattern = NULL,
  key = namekey(dataset), confirm = requireConsent(), ...)
}
\arguments{
\item{dataset}{the Dataset to modify}

\item{variables}{names or indices of variables to delete}

\item{pattern}{optional regular expression to identify Variables to delete}

\item{key}{the Variable attribute to \code{\link{grep}} with the
\code{pattern}. Default is "alias"}

\item{confirm}{logical: should the user be asked to confirm deletion.
Default is \code{TRUE} if in
an interactive session. You can avoid the confirmation prompt if you delete
\code{with(\link{consent})}.}

\item{...}{optional additional arguments to \code{grep}}
}
\value{
(invisibly) \code{dataset} with the specified variables deleted
}
\description{
Delete Variables Within a Dataset
}
\seealso{
\code{\link{hide}}
}

