% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataset-catalog.R, R/dataset.R, R/subvariables.R, R/variable-catalog.R
\docType{methods}
\name{names,DatasetCatalog-method}
\alias{aliases}
\alias{aliases,Subvariables-method}
\alias{aliases,VariableCatalog-method}
\alias{aliases<-}
\alias{aliases<-,Subvariables-method}
\alias{aliases<-,VariableCatalog-method}
\alias{describe-catalog}
\alias{names,CrunchDataset-method}
\alias{names,DatasetCatalog-method}
\alias{names,Subvariables-method}
\alias{names,VariableCatalog-method}
\alias{names<-,Subvariables-method}
\alias{names<-,VariableCatalog-method}
\title{Get and set names, aliases on Catalog-type objects}
\usage{
\S4method{names}{DatasetCatalog}(x)

\S4method{names}{CrunchDataset}(x)

\S4method{names}{Subvariables}(x)

\S4method{names}{Subvariables}(x) <- value

\S4method{aliases}{Subvariables}(x)

\S4method{aliases}{Subvariables}(x) <- value

\S4method{names}{VariableCatalog}(x)

\S4method{names}{VariableCatalog}(x) <- value

\S4method{aliases}{VariableCatalog}(x)

\S4method{aliases}{VariableCatalog}(x) <- value
}
\arguments{
\item{x}{a VariableCatalog, Subvariables, or similar object}

\item{value}{For the setters, an appropriate-length character vector to
assign}
}
\value{
Getters return the character object in the specified slot; setters
return \code{x} duly modified.
}
\description{
These methods let you get and set names and aliases for variables in a
Dataset's catalog, or within \code{\link{Subvariables}} in an array
variable. They work like the base R names methods.
}
\details{
Note that the \code{names} method on a Dataset returns the aliases of its
variables by default. See the vignette on variables for more information.
}
\seealso{
Subvariables Categories base::names
}

