% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batches.R, R/dataset-catalog.R, R/dataset.R,
%   R/multitables.R, R/shoji-catalog.R, R/subvariables.R, R/variable-catalog.R,
%   R/versions.R
\docType{methods}
\name{names,BatchCatalog-method}
\alias{names,BatchCatalog-method}
\alias{describe-catalog}
\alias{aliases}
\alias{aliases<-}
\alias{descriptions}
\alias{descriptions<-}
\alias{types}
\alias{emails}
\alias{timestamps}
\alias{names,CrunchDataset-method}
\alias{names<-,MultitableCatalog-method}
\alias{names,ShojiCatalog-method}
\alias{names<-,ShojiCatalog-method}
\alias{emails,ShojiCatalog-method}
\alias{aliases,Subvariables-method}
\alias{aliases<-,Subvariables-method}
\alias{names,CategoricalArrayVariable-method}
\alias{aliases,VariableCatalog-method}
\alias{aliases<-,VariableCatalog-method}
\alias{notes,VariableCatalog-method}
\alias{notes<-,VariableCatalog-method}
\alias{descriptions,VariableCatalog-method}
\alias{descriptions<-,VariableCatalog-method}
\alias{types,VariableCatalog-method}
\alias{ids,VariableCatalog-method}
\alias{names,VersionCatalog-method}
\alias{descriptions,VersionCatalog-method}
\alias{timestamps,VersionCatalog-method}
\title{Get and set names, aliases on Catalog-type objects}
\usage{
\S4method{names}{BatchCatalog}(x)

\S4method{names}{CrunchDataset}(x)

\S4method{names}{MultitableCatalog}(x) <- value

\S4method{names}{ShojiCatalog}(x)

\S4method{names}{ShojiCatalog}(x) <- value

\S4method{emails}{ShojiCatalog}(x)

\S4method{aliases}{Subvariables}(x)

\S4method{aliases}{Subvariables}(x) <- value

\S4method{names}{CategoricalArrayVariable}(x)

\S4method{aliases}{VariableCatalog}(x)

\S4method{aliases}{VariableCatalog}(x) <- value

\S4method{notes}{VariableCatalog}(x)

\S4method{notes}{VariableCatalog}(x) <- value

\S4method{descriptions}{VariableCatalog}(x)

\S4method{descriptions}{VariableCatalog}(x) <- value

\S4method{types}{VariableCatalog}(x)

\S4method{ids}{VariableCatalog}(x)

\S4method{names}{VersionCatalog}(x)

\S4method{descriptions}{VersionCatalog}(x)

\S4method{timestamps}{VersionCatalog}(x)
}
\arguments{
\item{x}{a VariableCatalog, Subvariables, or similar object}

\item{value}{For the setters, an appropriate-length character vector to
assign}
}
\value{
Getters return the character object in the specified slot; setters
return \code{x} duly modified.
}
\description{
These methods let you get and set names and aliases for variables in a
Dataset's catalog, or within \code{\link{Subvariables}} in an array
variable. They work like the base R names methods.
}
\details{
Note that the \code{names} method on a Dataset returns the aliases of its
variables by default. This is controlled by
\code{getOption("crunch.namekey.dataset")}, which is "alias" by default.
Set \code{options(crunch.namekey.dataset="name")} if you wish to use
variable names. See the vignette on variables for more information.
}
\seealso{
\code{\link{Subvariables}} \code{\link{Categories}} \code{\link[base]{names}} \code{vignette("variables", package="crunch")}
}
