% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-category.R, R/category.R
\docType{methods}
\name{name,AbsCat-method}
\alias{name,AbsCat-method}
\alias{name<-,AbsCat-method}
\alias{name<-,NULL-method}
\alias{value,AbsCat-method}
\alias{value<-,AbsCat-method}
\alias{id,AbsCat-method}
\alias{is.selected,AbsCat-method}
\alias{describe-category}
\alias{value}
\alias{value<-}
\alias{id}
\alias{is.selected}
\alias{name,Category-method}
\alias{name<-,Category-method}
\alias{name<-,NULL-method}
\alias{value,Category-method}
\alias{value<-,Category-method}
\alias{id,Category-method}
\title{Category attributes}
\usage{
\S4method{name}{AbsCat}(x)

\S4method{name}{AbsCat}(x) <- value

\S4method{name}{`NULL`}(x) <- value

\S4method{value}{AbsCat}(x)

\S4method{value}{AbsCat}(x) <- value

\S4method{id}{AbsCat}(x)

\S4method{is.selected}{AbsCat}(x)

\S4method{name}{Category}(x)

\S4method{name}{Category}(x) <- value

\S4method{name}{`NULL`}(x) <- value

\S4method{value}{Category}(x)

\S4method{value}{Category}(x) <- value

\S4method{id}{Category}(x)
}
\arguments{
\item{x}{a Category}

\item{value}{For the setters, an appropriate value to set}
}
\value{
\code{name} returns character; \code{value} and \code{id} return
numeric; value but not id may be \code{NA}; \code{is.selected} returns
logical indicating whether this Category is a "selected" dichotomy. Setters
return \code{x} duly modified.
}
\description{
Functions to access and set category attributes.
}
\seealso{
\code{\link{Categories}} \code{\link{dichotomize}}
}
