% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/category.R
\docType{methods}
\name{name,Category-method}
\alias{describe-category}
\alias{id}
\alias{id,Category-method}
\alias{id,list-method}
\alias{is.selected}
\alias{is.selected,Category-method}
\alias{name,Category-method}
\alias{name<-,Category-method}
\alias{value}
\alias{value,Category-method}
\alias{value<-}
\alias{value<-,Category-method}
\title{Category attributes}
\usage{
\S4method{name}{Category}(x)

\S4method{name}{Category}(x) <- value

\S4method{value}{Category}(x)

\S4method{value}{Category}(x) <- value

\S4method{id}{Category}(x)

\S4method{id}{list}(x)

\S4method{is.selected}{Category}(x)
}
\arguments{
\item{x}{a Category}

\item{value}{For the setters, an appropriate value to set}
}
\value{
\code{name} returns character; \code{value} and \code{id} return
numeric; value but not id may be \code{NA}; \code{is.selected} returns
logical indicating whether this Category is a "selected" dichotomy. Setters
return \code{x} duly modified.
}
\description{
Category attributes
}
\seealso{
Categories dichotomize
}

