% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{dim-dataset}
\alias{dim,CrunchDataset-method}
\alias{dim-dataset}
\title{Dataset dimensions}
\usage{
\S4method{dim}{CrunchDataset}(x)
}
\arguments{
\item{x}{a Dataset}
}
\value{
integer vector of length 2, indicating the number of rows and
non-hidden variables in the dataset. Array subvariables are excluded from
the column count.
}
\description{
Dataset dimensions
}
\seealso{
\code{\link[base]{dim}}
}

