% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-comparisons.R
\name{dimension-comparison}
\alias{dimension-comparison}
\alias{compareCols}
\alias{compareRows}
\alias{compareDims}
\title{Column and row comparison}
\usage{
compareCols(cube, ...)

compareRows(cube, ...)

compareDims(cube, dim = c("cols", "rows"), baseline, x,
  value = c("statistic", "p.value"))
}
\arguments{
\item{cube}{a CrunchCube to calculate the comparison on}

\item{...}{arguments passed from \code{compareRows()} or \code{compareCols()} to
\code{compareDims()} (i.e. \code{baseline} and \code{x})}

\item{dim}{which dimension is being compared (\code{cols} or \code{rows}, only valid
for \code{compareDims()})}

\item{baseline}{a character, the column/row to use as a baseline to compare \code{x}
against}

\item{x}{a character, the column/row to compare against the baseline}

\item{value}{Value of the test to return. Defaults to \code{statistic}; use \code{p.value}
for the p-value of the test of the pair of columns/rows.}
}
\value{
the value for the column or row given in \code{x}, default is Chi-squared
}
\description{
Comparing a column or row with a baseline column or row. This calculates the
test statistic \eqn{\chi^2} of independence for each pair of columns/rows.
}
