% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\docType{methods}
\name{displaySettings}
\alias{displaySettings}
\alias{displaySettings<-}
\alias{displaySettings,CrunchSlide-method}
\alias{displaySettings<-,CrunchSlide,ANY-method}
\alias{displaySettings,AnalysisCatalog-method}
\alias{displaySettings<-,AnalysisCatalog,list-method}
\alias{displaySettings,Analysis-method}
\alias{displaySettings<-,Analysis,ANY-method}
\alias{cubes,CrunchDeck-method}
\title{Get or set a slide's display settings}
\usage{
displaySettings(x)

displaySettings(x) <- value

\S4method{displaySettings}{CrunchSlide}(x)

\S4method{displaySettings}{CrunchSlide,ANY}(x) <- value

\S4method{displaySettings}{AnalysisCatalog}(x)

\S4method{displaySettings}{AnalysisCatalog,list}(x) <- value

\S4method{displaySettings}{Analysis}(x)

\S4method{displaySettings}{Analysis,ANY}(x) <- value

\S4method{cubes}{CrunchDeck}(x)
}
\arguments{
\item{x}{a CrunchSlide, Analysis, or AnalysisCatalog}

\item{value}{a named list, for valid settings see docs.crunch.io}
}
\description{
A slide's display settings can be modified by assigning a named list
}
