% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\docType{methods}
\name{duplicated}
\alias{duplicated}
\alias{duplicated,CrunchVariable-method}
\alias{duplicated,CrunchExpr-method}
\title{"duplicated" method for Crunch objects}
\usage{
\S4method{duplicated}{CrunchVariable}(x, incomparables = FALSE, ...)

\S4method{duplicated}{CrunchExpr}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{CrunchVariable or CrunchExpr}

\item{incomparables}{Ignored}

\item{...}{Ignored}
}
\value{
A CrunchLogicalExpr that evaluates \code{TRUE} for all repeated
entries after the first occurrence of a value.
}
\description{
"duplicated" method for Crunch objects
}
\seealso{
\code{\link[base:duplicated]{base::duplicated}}
}
