% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{envOrOption}
\alias{envOrOption}
\title{Grab either env variable or option}
\usage{
envOrOption(opt)
}
\arguments{
\item{opt}{the option to get}
}
\value{
the value of the option
}
\description{
.Rprofile options are like "test.api", while env vars are "R_TEST_API".
This function will use the environment variable if it is found, otherwise
it looks for the R-based option value.
}
\keyword{internal}
