% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{escapeRegex}
\alias{escapeRegex}
\title{Escape a regular expression}
\usage{
escapeRegex(string)
}
\arguments{
\item{string}{A regular expression to escape}
}
\value{
\code{string}, escaped.
}
\description{
This function takes a string and escapes all of the special characters in the
string. For example, the \code{.} in \code{VB.NET} will be escaped with a slash (though
regular R printing will make it look like there are two slashes).
}
\examples{
\dontrun{
escapeRegex("Tom&Jerry")
escapeRegex(".Net")
}
}
\keyword{internal}
