% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-dataset.R
\docType{methods}
\name{exportDataset}
\alias{exportDataset}
\alias{write.csv,CrunchDataset-method}
\title{Export a dataset to a file}
\usage{
exportDataset(dataset, file, format = c("csv", "spss"), ...)

\S4method{write.csv}{CrunchDataset}(...)
}
\arguments{
\item{dataset}{CrunchDataset}

\item{file}{character local filename to write to}

\item{format}{character export format: currently supported values are "csv"
and "spss".}

\item{...}{additional arguments, currently ignored}
}
\value{
Invisibly, \code{file}.
}
\description{
Export a dataset to a file
}

