% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R, R/make-array.R
\name{expressions-internal}
\alias{expressions-internal}
\alias{\%in\%}
\alias{==}
\alias{!=}
\alias{!}
\alias{crunchDifftime}
\alias{datetimeFromCols}
\alias{\%ornm\%}
\alias{!,CrunchExpr-method}
\alias{\%in\%,CategoricalVariable,character-method}
\alias{\%in\%,CategoricalVariable,factor-method}
\alias{\%in\%,TextVariable,character-method}
\alias{\%in\%,NumericVariable,numeric-method}
\alias{\%in\%,DatetimeVariable,Date-method}
\alias{\%in\%,DatetimeVariable,POSIXt-method}
\alias{\%in\%,DatetimeVariable,character-method}
\alias{\%in\%,CategoricalVariable,numeric-method}
\alias{==,CategoricalVariable,numeric-method}
\alias{==,CategoricalVariable,character-method}
\alias{==,CategoricalVariable,factor-method}
\alias{!=,CategoricalVariable,numeric-method}
\alias{!=,CategoricalVariable,character-method}
\alias{!=,CategoricalVariable,factor-method}
\alias{is.na,CrunchVarOrExpr-method}
\alias{is.valid}
\alias{bin}
\alias{selectCategories}
\alias{crunchBetween}
\alias{rowAll}
\alias{rowAny}
\alias{rowAnyNA}
\alias{rowAllNA}
\alias{complete.cases,CrunchVarOrExpr-method}
\alias{is.selected,CrunchVarOrExpr-method}
\alias{asSelected}
\alias{selectedDepth}
\alias{arraySelections}
\alias{nchar,CrunchVarOrExpr-method}
\alias{trim}
\alias{alterCategoriesExpr}
\alias{alterArrayExpr}
\alias{arraySubsetExpr}
\alias{makeFrame}
\title{Crunch expressions internal}
\usage{
x \%in\% table

crunchDifftime(e1, e2, resolution = NULL)

datetimeFromCols(year, month, day, hour = NULL, minute = NULL, second = NULL)

e1 \%ornm\% e2

\S4method{!}{CrunchExpr}(x)

\S4method{\%in\%}{CategoricalVariable,character}(x, table)

\S4method{\%in\%}{CategoricalVariable,factor}(x, table)

\S4method{\%in\%}{TextVariable,character}(x, table)

\S4method{\%in\%}{NumericVariable,numeric}(x, table)

\S4method{\%in\%}{DatetimeVariable,Date}(x, table)

\S4method{\%in\%}{DatetimeVariable,POSIXt}(x, table)

\S4method{\%in\%}{DatetimeVariable,character}(x, table)

\S4method{\%in\%}{CategoricalVariable,numeric}(x, table)

\S4method{==}{CategoricalVariable,numeric}(e1, e2)

\S4method{==}{CategoricalVariable,character}(e1, e2)

\S4method{==}{CategoricalVariable,factor}(e1, e2)

\S4method{!=}{CategoricalVariable,numeric}(e1, e2)

\S4method{!=}{CategoricalVariable,character}(e1, e2)

\S4method{!=}{CategoricalVariable,factor}(e1, e2)

\S4method{is.na}{CrunchVarOrExpr}(x)

is.valid(x)

bin(x)

selectCategories(x, selections, collapse = TRUE)

crunchBetween(x, lower, upper, inclusive = c(TRUE, FALSE))

rowAll(x, na.rm = FALSE)

rowAny(x, na.rm = FALSE)

rowAnyNA(x)

rowAllNA(x)

\S4method{complete.cases}{CrunchVarOrExpr}(x, ...)

\S4method{is.selected}{CrunchVarOrExpr}(x)

asSelected(x)

selectedDepth(x)

arraySelections(x)

\S4method{nchar}{CrunchVarOrExpr}(x, type = "chars", allowNA = FALSE, keepNA = NA)

trim(x, min, max)

alterCategoriesExpr(
  x,
  categories = NULL,
  category_order = NULL,
  subvariables = NULL
)

alterArrayExpr(
  x,
  add = NULL,
  order = NULL,
  order_id = c("alias", "name", "id"),
  remove = NULL,
  remove_id = c("alias", "name", "id"),
  subreferences = NULL,
  subreferences_id = c("alias", "name", "id")
)

arraySubsetExpr(x, subvars, subvar_id = c("alias", "name", "id"))

makeFrame(x)
}
\arguments{
\item{x, e1, e2}{inputs}

\item{table, na.rm, selections, upper, lower, inclusive, regex, ignore_case, selections, collapse, tiers, }{cases,data,resolution,year,month,day,hours,minutes,seconds,min,max,categories,category_order,
subvariables,remove,remove_id,subreferences,subreferences_id
Other parameters used in some functions, see details of \code{\link{expressions}} for more details.}
}
\value{
Most functions return a CrunchExpr or CrunchLogicalExpr.
\code{as.vector} returns an R vector.
}
\description{
See \code{\link{expressions}} for more details.
}
\keyword{internal}
