% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expressions.R
\docType{methods}
\name{as.vector,CrunchExpr-method}
\alias{!,CrunchExpr-method}
\alias{\%in\%,CategoricalVariable,character-method}
\alias{\%in\%,CategoricalVariable,factor-method}
\alias{as.vector,CrunchExpr-method}
\alias{expressions}
\alias{is.na,CrunchVariable-method}
\title{Construct Crunch Expressions}
\usage{
\S4method{as.vector}{CrunchExpr}(x, mode = "any")

\S4method{!}{CrunchExpr}(x)

\S4method{\%in\%}{CategoricalVariable,character}(x, table)

\S4method{\%in\%}{CategoricalVariable,factor}(x, table)

\S4method{is.na}{CrunchVariable}(x)
}
\arguments{
\item{x}{an input}

\item{mode}{For \code{as.vector}. Ignored.}

\item{table}{For \code{\%in\%}. See \code{\link{base::match}}}
}
\value{
Most functions return a CrunchExpr or CrunchLogicalExpr.
\code{as.vector} returns an R vector.
}
\description{
Crunch Expressions, i.e. \code{CrunchExpr} and \code{CrunchLogicalExpr},
encapuslate derivations of Crunch variables, which are only evaluated when
passed to a function like \code{as.vector}. They allow you to compose
functional expressions of variables and evaluate them against the server
only when appropriate.
}

