% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{extralogical}
\alias{extralogical}
\alias{between,CrunchVariable-method}
\alias{between,CrunchExpr-method}
\title{"extra" logical method for CrunchLogicalExpr}
\usage{
\S4method{between}{CrunchVariable}(x, lower, upper, inclusive = c(TRUE, FALSE))

\S4method{between}{CrunchExpr}(x, lower, upper, inclusive = c(TRUE, FALSE))
}
\arguments{
\item{x}{a CrunchVariable or CrunchExpr}
}
\value{
a CrunchLogicalExpr
}
\description{
"extra" logical method for CrunchLogicalExpr
}
