% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\docType{methods}
\name{filter-catalog}
\alias{filter-catalog}
\alias{filters}
\alias{filters,CrunchDataset-method}
\alias{filters<-}
\alias{filters<-,CrunchDataset-method}
\title{Filter entities for a dataset}
\usage{
\S4method{filters}{CrunchDataset}(x)

\S4method{filters}{CrunchDataset}(x) <- value
}
\arguments{
\item{x}{a CrunchDataset}

\item{value}{for the assignment method, a FilterCatalog}
}
\value{
an object of class FilterCatalog containing references to Filter
entities usable in the web application. (Setter returns the Dataset.)
}
\description{
Filter entities for a dataset
}

