% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\docType{methods}
\name{filter-catalog}
\alias{filter-catalog}
\alias{filters}
\alias{filters,CrunchDataset-method}
\title{Filter entities for a dataset}
\usage{
\S4method{filters}{CrunchDataset}(x)
}
\arguments{
\item{x}{a CrunchDataset}
}
\value{
an object of class FilterCatalog containing references to Filter
entities usable in the web application.
}
\description{
Filter entities for a dataset
}

