% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{findVariables}
\alias{findVariables}
\title{Search a Dataset or list of Variables}
\usage{
findVariables(dataset, refs = NULL, pattern = "", key = namekey(dataset),
  ...)
}
\arguments{
\item{dataset}{the Dataset or list of Crunch objects to search}

\item{refs}{vector of specific values in \code{key} to look for, as an
alternative to pattern matching. Default is \code{NULL}, which means to
use \code{grep} instead.}

\item{pattern}{regular expression, passed to \code{grep}. If "", returns
all.}

\item{key}{the field in the Crunch objects in which to grep}

\item{...}{additional arguments passed to \code{grep}. If \code{value=TRUE},
returns the values of \code{key} where matches are found, not the variables
themselves}
}
\value{
indices of the Variables that match the pattern, or the matching
key values if value=TRUE is passed to \code{grep}
}
\description{
A version of \code{\link{grep}} for Crunch objects
}
\keyword{internal}

