% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fork-and-merge.R
\name{forkDataset}
\alias{forkDataset}
\title{Create a fork of a dataset}
\usage{
forkDataset(dataset, forkname)
}
\arguments{
\item{dataset}{The \code{CrunchDataset} to fork}

\item{forkname}{character name to give the fork. If omitted, one will be
provided for you}
}
\value{
The new fork, a \code{CrunchDataset}.
}
\description{
As with many other version control systems, in Crunch you can fork a
dataset's revision history, effectively making a copy on which you can work
independently of the original dataset. You can then merge those change back
to the original dataset or keep working independently.
}

