% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{generateWeightEntry}
\alias{generateWeightEntry}
\title{Generate entry for \link{makeWeight}}
\usage{
generateWeightEntry(expr)
}
\arguments{
\item{expr}{An expression}
}
\value{
A list of the variable id and the target weights.
}
\description{
Utility function to catch formula errors passed to \link{makeWeight}, and to
generate the appropriate entry.
}
\examples{
\dontrun{
validateWeightExpression(ds$var ~ c(10, 30, 60))
}
}
\keyword{internal}
