% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getAccountUserCatalog}
\alias{getAccountUserCatalog}
\title{Find all users on your account}
\usage{
getAccountUserCatalog(x = shojiURL(getAccount(), "catalogs", "users"))
}
\arguments{
\item{x}{URL of the user catalog. Default is the right thing; you shouldn't
specify one}
}
\value{
a \code{UserCatalog}
}
\description{
Find all users on your account
}

