% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube.R
\name{getCube}
\alias{getCube}
\title{Crosstab and otherwise aggregate variables in a dataset}
\usage{
getCube(formula, data, useNA = c("no", "ifany", "always"))
}
\arguments{
\item{formula}{an object of class 'formula' object with the
cross-classifying variables, separated by '+', on the right hand side.
Compare to \code{\link[stats]{xtabs}}.}

\item{data}{an object of class \code{CrunchDataset}}

\item{useNA}{whether to include missing values in tabular results. See
\code{\link[base]{table}}.}
}
\value{
an object of class \code{CrunchCube}
}
\description{
Create a contingency table or other aggregation from cross-classifying
variables in a CrunchDataset.
}

