% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{getTeams}
\alias{getTeams}
\title{Retrive all teams you're a member of}
\usage{
getTeams()
}
\value{
A \code{TeamCatalog}. Extract an individual team by name. Create
a team by assigning in with a new name.
}
\description{
Retrive all teams you're a member of
}
\seealso{
\link{teams}
}

