% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{getTeams}
\alias{getTeams}
\title{Retrieve your teams}
\usage{
getTeams()
}
\value{
A \code{TeamCatalog}. Extract an individual team by name. Create
a team by assigning in with a new name.
}
\description{
Teams contain a list of users. You can grant access to a group of
users by inviting the team. You can also share a set of
datasets with a user all at once by adding the user to a team that contains
those datasets.
}
\details{
\code{getTeams()} returns your \code{TeamCatalog}. You can extract an individual team
by name, or create a team by assigning into the function.
To create a team by  assignment, assign a list to
\code{teams("myteam") <- value_list}. The \code{value_list} can either empty
(to just create a team with that name), or can contain a "members" element
with the emails or URLs of users to add to the team. Users can be also be
added later with the \code{members<-}  method.
}
\seealso{
\code{\link{members}}
}
