% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variable-order.R
\name{grouped}
\alias{grouped}
\alias{ungrouped}
\title{Get un(grouped) VariableGroups}
\usage{
grouped(var.order)

ungrouped(var.order)
}
\arguments{
\item{var.order}{an object of class VariableOrder or VariableGroup}
}
\value{
For grouped(), a VariableOrder/Group, respectively, with "ungrouped"
omitted. For ungrouped(), a VariableGroup.
}
\description{
"ungrouped" is a magic VariableGroup that contains all variables not found
in groups at a given level of nesting.
}
\seealso{
VariableOrder
}

