% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-order.R
\name{grouped}
\alias{grouped}
\alias{ungrouped}
\title{Get un(grouped) OrderGroups}
\usage{
grouped(order.obj)

ungrouped(order.obj)
}
\arguments{
\item{order.obj}{an subclass of ShojiOrder or OrderGroup}
}
\value{
For grouped(), an Order/Group, respectively, with "ungrouped"
omitted. For ungrouped(), an OrderGroup subclass.
}
\description{
"ungrouped" is a magic OrderGroup that contains all entities not found
in groups at a given level of nesting.
}
\seealso{
\code{\link{VariableOrder}}
}
