% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hasFunction}
\alias{hasFunction}
\title{Check if a package has a function installed}
\usage{
hasFunction(fun, pkg)
}
\arguments{
\item{fun}{a string, name of the function to search for}

\item{pkg}{a string, the package to search for the function in}
}
\value{
Logical
}
\description{
This is useful for handling if the user has the correct version of a suggested
package installed.
}
