% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame.R
\name{head-tail}
\alias{head-tail}
\alias{head}
\alias{tail}
\alias{head,CrunchDataset-method}
\alias{head.CrunchDataFrame}
\alias{head,CrunchVariable-method}
\alias{tail,CrunchDataset-method}
\alias{tail.CrunchDataFrame}
\alias{tail,CrunchVariable-method}
\title{Head and tail methods for Crunch objects. See \code{\link[utils:head]{utils::head()}} for more details.}
\usage{
\S4method{head}{CrunchDataset}(x, n = 6L, ...)

head.CrunchDataFrame(x, n = 6L, ...)

\S4method{head}{CrunchVariable}(x, n = 6L, ...)

\S4method{tail}{CrunchDataset}(x, n = 6L, ...)

tail.CrunchDataFrame(x, n = 6L, ...)

\S4method{tail}{CrunchVariable}(x, n = 6L, ...)
}
\arguments{
\item{x}{a CrunchDataset, CrunchDataFrame, or CrunchVariable}

\item{n}{a single integer representing the length of the returning object.}

\item{...}{ignored}
}
\description{
Head and tail methods for Crunch objects. See \code{\link[utils:head]{utils::head()}} for more details.
}
