% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hide-variables.R
\name{hiddenVariables}
\alias{hiddenVariables}
\title{Show the names of hidden variables within the dataset}
\usage{
hiddenVariables(dataset, key = "name")
}
\arguments{
\item{dataset}{the Dataset}
}
\value{
a vector of the names of Variables marked as hidden.
}
\description{
Show the names of hidden variables within the dataset
}

