% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hiddenVariables}
\alias{hiddenVariables}
\title{Show the names of hidden variables within the dataset}
\usage{
hiddenVariables(dataset, key = "name")
}
\arguments{
\item{dataset}{the Dataset}
}
\value{
a vector of the names of Variables marked as hidden.
}
\description{
Show the names of hidden variables within the dataset
}

