% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide-variables.R
\name{hiddenVariables}
\alias{hiddenVariables}
\title{Show the names of a dataset's hidden variables}
\usage{
hiddenVariables(dataset, key = namekey(dataset))
}
\arguments{
\item{dataset}{the Dataset}

\item{key}{the Variable attribute to return. Default is "alias", following
\code{getOption("crunch.namekey.dataset")}.}
}
\value{
a vector of the names of Variables marked as hidden.
}
\description{
Show the names of a dataset's hidden variables
}
