% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide-variables.R, R/private-variables.R
\name{hiddenVariables}
\alias{hiddenVariables}
\alias{privateVariables}
\title{Show the names of a dataset's hidden or private variables}
\usage{
hiddenVariables(dataset, key = namekey(dataset))

privateVariables(dataset, key = namekey(dataset))
}
\arguments{
\item{dataset}{the Dataset}

\item{key}{the Variable attribute to return. Default is "alias", following
\code{getOption("crunch.namekey.dataset")}.}
}
\value{
a vector of the names of Variables marked as hidden/private.
}
\description{
Both hidden and private are hidden from most views in crunch by default.
Hidden variables can be accessed by an user, while private variables
(and all variables derived from them) are only accessible
by users granted "editor" access to the dataset and so can be used to secure
personally identifiable information from non-editors of a dataset.
}
