% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{hide,CrunchVariable-method}
\alias{hide,CrunchVariable-method}
\alias{unhide,CrunchVariable-method}
\title{Hide and Unhide Variables}
\usage{
\S4method{hide}{CrunchVariable}(x)

\S4method{unhide}{CrunchVariable}(x)
}
\arguments{
\item{x}{a Variable to hide or unhide}
}
\value{
(invisibly) the Variable, hidden or unhidden
}
\description{
Hide and Unhide Variables
}

