% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide-variables.R, R/private-variables.R
\name{hide}
\alias{hide}
\alias{unhide}
\alias{privatize}
\alias{deprivatize}
\alias{hide,CrunchVariable-method}
\alias{hide,VariableCatalog-method}
\alias{unhide,CrunchVariable-method}
\alias{unhide,VariableCatalog-method}
\alias{privatize,CrunchVariable-method}
\alias{privatize,VariableCatalog-method}
\alias{deprivatize,CrunchVariable-method}
\alias{deprivatize,VariableCatalog-method}
\alias{privatise}
\alias{deprivatise}
\alias{privatiseVariables}
\alias{deprivatiseVariables}
\title{Hide/Unhide or Privatize/Deprivatize Variables}
\usage{
\S4method{hide}{CrunchVariable}(x)

\S4method{hide}{VariableCatalog}(x)

\S4method{unhide}{CrunchVariable}(x)

\S4method{unhide}{VariableCatalog}(x)

\S4method{privatize}{CrunchVariable}(x)

\S4method{privatize}{VariableCatalog}(x)

\S4method{deprivatize}{CrunchVariable}(x)

\S4method{deprivatize}{VariableCatalog}(x)

privatise(x)

deprivatise(x)

privatiseVariables(dataset, variables)

deprivatiseVariables(dataset, variables)
}
\arguments{
\item{x}{a Variable or subset of a VariableCatalog to hide or unhide}
}
\value{
(invisibly) the Variable or VariableCatalog, hidden or unhidden
}
\description{
Both hidden and private are hidden from most views in crunch by default.
Hidden variables can be accessed by an user, while private variables
(and all variables derived from them) are only accessible
by users granted "editor" access to the dataset and so can be used to secure
personally identifiable information from non-editors of a dataset.
}
\seealso{
\code{\link{hideVariables}} \code{\link{privateVariables}}
}
