% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide-variables.R
\name{hideVariables}
\alias{hideVariables}
\alias{hiddenVariables<-}
\alias{unhideVariables}
\title{Hide and unhide variables within a dataset}
\usage{
hideVariables(dataset, variables)

hiddenVariables(x) <- value

unhideVariables(dataset, variables)
}
\arguments{
\item{dataset}{the Dataset to modify}

\item{variables}{names or indices of variables to (un)hide}

\item{x}{same as \code{dataset}, for \code{hiddenVariables<-}}

\item{value}{same as \code{variables}, for \code{hiddenVariables<-}}
}
\value{
(invisibly) \code{dataset} with the specified variables (un)hidden
}
\description{
Hide and unhide variables within a dataset
}
\seealso{
\code{\link{hide}}
}
