% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide-variables.R, R/private-variables.R
\name{hideVariables}
\alias{hideVariables}
\alias{hiddenVariables<-}
\alias{unhideVariables}
\alias{privatizeVariables}
\alias{privateVariables<-}
\alias{deprivatizeVariables}
\title{Hide and unhide variables within a dataset}
\usage{
hideVariables(dataset, variables)

hiddenVariables(x) <- value

unhideVariables(dataset, variables)

privatizeVariables(dataset, variables)

privateVariables(x) <- value

deprivatizeVariables(dataset, variables)
}
\arguments{
\item{dataset}{the Dataset to modify}

\item{variables}{names or indices of variables to (un)hide}

\item{x}{\code{dataset}, for \verb{hiddenVariables<-}}

\item{value}{\code{variables}, for \verb{hiddenVariables<-}}
}
\value{
(invisibly) \code{dataset} with the specified variables (un)hidden
}
\description{
Both hidden and private are hidden from most views in crunch by default.
Hidden variables can be accessed by an user, while private variables
(and all variables derived from them) are only accessible
by users granted "editor" access to the dataset and so can be used to secure
personally identifiable information from non-editors of a dataset.
}
\seealso{
\code{\link{hide}}
}
