% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api.R
\name{crGET}
\alias{crDELETE}
\alias{crGET}
\alias{crPATCH}
\alias{crPOST}
\alias{crPUT}
\title{HTTP methods for communicating with the Crunch API}
\usage{
crGET(...)

crPUT(...)

crPATCH(...)

crPOST(...)

crDELETE(...)
}
\arguments{
\item{...}{see \code{\link{crunchAPI}} for details. \code{url} is the first
named argument and is required; \code{body} is also required for PUT,
PATCH, and POST.}
}
\value{
Depends on the response status of the HTTP request and any custom
handlers.
}
\description{
HTTP methods for communicating with the Crunch API
}

