% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{http-methods}
\alias{http-methods}
\alias{crGET}
\alias{crPUT}
\alias{crPATCH}
\alias{crPOST}
\alias{crDELETE}
\title{HTTP methods for communicating with the Crunch API}
\usage{
crGET(...)

crPUT(...)

crPATCH(...)

crPOST(...)

crDELETE(...)
}
\arguments{
\item{...}{see \code{\link{crunchAPI}} for details. \code{url} is the first
named argument and is required; \code{body} is also required for PUT,
PATCH, and POST.}
}
\value{
Depends on the response status of the HTTP request and any custom
handlers.
}
\description{
These methods let you communicate with the Crunch API, for more background
see \href{https://crunch.io/r/crunch/articles/crunch-internals.html}{Crunch Internals}.
}
