% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/insertions.R
\docType{class}
\name{Insertions-class}
\alias{Insertions-class}
\alias{anchor}
\alias{anchor<-}
\alias{combinations}
\alias{combinations<-}
\alias{Insertions}
\alias{Insertion-class}
\alias{Insertion}
\alias{names,Insertions-method}
\alias{anchor,Insertion-method}
\alias{anchor<-,Insertion-method}
\alias{combinations,Insertion-method}
\alias{combinations<-,Insertion-method}
\title{Insert categories in transformations}
\usage{
Insertions(..., data = NULL)

Insertion(..., data = NULL)

\S4method{names}{Insertions}(x)

\S4method{anchor}{Insertion}(x)

\S4method{anchor}{Insertion}(x) <- value

\S4method{combinations}{Insertion}(x)

\S4method{combinations}{Insertion}(x) <- value
}
\arguments{
\item{...}{additional arguments to \code{[}, ignored}

\item{data}{For the constructor functions \code{Insertion} and
\code{Insertions}, you can either pass in attributes via \code{...} or you
can create the objects with a fully defined \code{list} representation of
the objects via the \code{data} argument. See the examples.}

\item{x}{For the attribute getters and setters, an object of class
Insertion or Insertions}

\item{value}{For \code{[<-}, the replacement Insertion to insert}
}
\description{
Insertions allow you to insert new categories into a categorical-like response on a variable's \href{Transforms}{transform}.
}
