% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/insertions.R
\docType{class}
\name{Insertions-class}
\alias{Insertions-class}
\alias{Insertions}
\alias{Insertion-class}
\alias{Insertion}
\alias{.Insertion}
\alias{Insertions}
\alias{anchor}
\alias{anchor<-}
\alias{anchors}
\alias{func}
\alias{func<-}
\alias{funcs}
\alias{args}
\alias{args<-}
\alias{anchor<-,Insertion-method}
\alias{subtotals<-,Insertion,ANY-method}
\alias{args,Insertion-method}
\alias{args,Subtotal-method}
\alias{args,Heading-method}
\alias{anchor,Insertion-method}
\alias{anchor,Subtotal-method}
\alias{anchor,Heading-method}
\alias{func,Insertion-method}
\alias{func,Subtotal-method}
\alias{func,Heading-method}
\alias{anchors,Insertions-method}
\alias{funcs,Insertions-method}
\title{Insert categories in transformations}
\usage{
Insertions(..., data = NULL)

Insertion(...)

.Insertion(..., data = NULL)

\S4method{anchor}{Insertion}(x) <- value

\S4method{subtotals}{Insertion,ANY}(x) <- value

\S4method{args}{Insertion}(x)

\S4method{args}{Subtotal}(x, var_categories)

\S4method{args}{Heading}(x, var_categories)

\S4method{anchor}{Insertion}(x)

\S4method{anchor}{Subtotal}(x, var_categories)

\S4method{anchor}{Heading}(x, var_categories)

\S4method{func}{Insertion}(x)

\S4method{func}{Subtotal}(x)

\S4method{func}{Heading}(x)

\S4method{anchors}{Insertions}(x)

\S4method{funcs}{Insertions}(x)
}
\arguments{
\item{...}{additional arguments to \code{[}, ignored}

\item{data}{For the constructor functions \code{Insertion} and
\code{Insertions}, you can either pass in attributes via \code{...} or you
can create the objects with a fully defined \code{list} representation of
the objects via the \code{data} argument. See the examples.}

\item{x}{For the attribute getters and setters, an object of class
Insertion or Insertions}

\item{value}{For \code{[<-}, the replacement Insertion to insert}

\item{var_categories}{categories (from \code{\link[=categories]{categories()}}) to used by \code{args} and
\code{anchor} methods when needed to translate between category names and category
ids.}
}
\description{
Insertions allow you to insert new categories into a categorical-like
response on a variable's \href{Transforms}{transform}.
}
