% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/abstract-categories.R,
%   R/abstract-category.R, R/insertions.R
\docType{class}
\name{Insertions-class}
\alias{Insertions-class}
\alias{anchor}
\alias{anchor<-}
\alias{anchors}
\alias{func}
\alias{func<-}
\alias{funcs}
\alias{args}
\alias{args<-}
\alias{Insertions}
\alias{Insertion-class}
\alias{Insertion}
\alias{anchors,AbsCats-method}
\alias{funcs,AbsCats-method}
\alias{args,AbsCat-method}
\alias{anchor,AbsCat-method}
\alias{func,AbsCat-method}
\alias{func,NULL-method}
\alias{anchor,NULL-method}
\alias{anchor<-,Insertion-method}
\alias{subtotals<-,Insertion,ANY-method}
\title{Insert categories in transformations}
\usage{
Insertions(..., data = NULL)

Insertion(..., data = NULL)

\S4method{anchors}{AbsCats}(x)

\S4method{funcs}{AbsCats}(x)

\S4method{args}{AbsCat}(x)

\S4method{anchor}{AbsCat}(x)

\S4method{func}{AbsCat}(x)

\S4method{func}{`NULL`}(x)

\S4method{anchor}{`NULL`}(x)

\S4method{anchor}{Insertion}(x) <- value

\S4method{subtotals}{Insertion,ANY}(x) <- value
}
\arguments{
\item{...}{additional arguments to \code{[}, ignored}

\item{data}{For the constructor functions \code{Insertion} and
\code{Insertions}, you can either pass in attributes via \code{...} or you
can create the objects with a fully defined \code{list} representation of
the objects via the \code{data} argument. See the examples.}

\item{x}{For the attribute getters and setters, an object of class
Insertion or Insertions}

\item{value}{For \code{[<-}, the replacement Insertion to insert}
}
\description{
Insertions allow you to insert new categories into a categorical-like
response on a variable's \href{Transforms}{transform}.
}
