% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R, R/category.R
\docType{methods}
\name{is-na-categories}
\alias{is-na-categories}
\alias{is.na,Categories-method}
\alias{is.na,Category-method}
\alias{is.na<-,Categories,character-method}
\alias{is.na<-,Categories,logical-method}
\alias{is.na<-,Category,logical-method}
\title{is.na for Categories}
\usage{
\S4method{is.na}{Categories}(x)

\S4method{is.na}{Categories,character}(x) <- value

\S4method{is.na}{Categories,logical}(x) <- value

\S4method{is.na}{Category}(x)

\S4method{is.na}{Category,logical}(x) <- value
}
\arguments{
\item{x}{Categories or a single Category}

\item{value}{To change the missingness of categories, supply either (1)
a logical vector of equal length of the categories (or length 1 for the
Category method), or (2) the names of the categories to mark as missing.
If supplying the latter, any categories already indicated as missing will
remain missing.}
}
\value{
Getters return logical, a named vector in the case of the Categories
method; setters return \code{x} duly modified.
}
\description{
is.na for Categories
}

