% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R, R/category.R
\docType{methods}
\name{is-selected-categories}
\alias{is-selected-categories}
\alias{is.selected<-}
\alias{is.selected,Categories-method}
\alias{is.selected<-,Categories-method}
\alias{is.selected,Category-method}
\alias{is.selected<-,Category-method}
\title{is.selected for Categories}
\usage{
\S4method{is.selected}{Categories}(x)

\S4method{is.selected}{Categories}(x) <- value

\S4method{is.selected}{Category}(x)

\S4method{is.selected}{Category}(x) <- value
}
\arguments{
\item{x}{Categories or a single Category}

\item{value}{A logical vector indicating whether the category should be selected.
For a single category the value should be either \code{TRUE} or \code{FALSE} to change the
selection status for a \code{Categories} object, supply a logical vector which is the
same length as the number of categories.}
}
\value{
Getters return a logical vector indicating selection status. Setters return
the \code{Categories} or \code{Category} object, duly modified.
}
\description{
Crunch Multiple Response variables identify one or more categories as "selected".
These methods allow you to get or set which categories should indicate a selection.
}
