% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listDatasets}
\alias{listDatasets}
\title{Show the names of all Crunch datasets}
\usage{
listDatasets(kind = c("active", "all", "archived"), refresh = FALSE)
}
\arguments{
\item{kind}{character specifying whether to look in active, archived, or all
datasets.}

\item{refresh}{logical: should the function check the Crunch API for new
datasets? Default is FALSE.}
}
\value{
Character vector of dataset names, each of which would be a valid
input for \code{\link{loadDataset}}
}
\description{
Show the names of all Crunch datasets
}

