\name{loadDataset}
\alias{loadDataset}
\title{Load a Crunch Dataset}
\usage{
  loadDataset(dataset.name)
}
\arguments{
  \item{dataset.name}{character, the name of a Crunch
  dataset you have access to.}

  \item{dataset.list,}{the local session store from which
  to retrieve the dataset's resource URLs. You shouldn't
  change this.}
}
\value{
  An object of class \code{CrunchDataset}
}
\description{
  Load a Crunch Dataset
}

