% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-datasets.R
\name{loadDataset}
\alias{loadDataset}
\title{Load a Crunch Dataset}
\usage{
loadDataset(dataset.name, kind = c("active", "all", "archived"),
  useAlias = default.useAlias())
}
\arguments{
\item{dataset.name}{character, the name of a Crunch dataset you have access
to.}

\item{kind}{character specifying whether to look in active, archived, or all
datasets.}

\item{useAlias}{logical whether variable alias or name should be used as R
variable names when the dataset is returned. Default is TRUE, meaning alias.
They're more computer friendly.}
}
\value{
An object of class \code{CrunchDataset}
}
\description{
Load a Crunch Dataset
}

