% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-datasets.R
\name{loadDataset}
\alias{loadDataset}
\title{Load a Crunch Dataset}
\usage{
loadDataset(dataset.name, kind = c("active", "all", "archived"))
}
\arguments{
\item{dataset.name}{character, the name of a Crunch dataset you have access
to.}

\item{kind}{character specifying whether to look in active, archived, or all
datasets.}
}
\value{
An object of class \code{CrunchDataset}
}
\description{
Load a Crunch Dataset
}

