\name{login}
\alias{login}
\title{Authenticate with the Crunch API}
\usage{
  login(email = getOption("crunch.email"),
    password = getOption("crunch.pw"), ...)
}
\arguments{
  \item{email}{the user's email address}

  \item{...}{additional parameters}
}
\description{
  Authenticate with the Crunch API
}

