% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical-array.R, R/multiple-response.R
\name{makeArray}
\alias{makeArray}
\alias{makeMR}
\title{Make a Categorical Array or Multiple Response variable}
\usage{
makeArray(subvariables, dataset = NULL, name, ...)

makeMR(subvariables, dataset = NULL, name, selections, ...)
}
\arguments{
\item{subvariables}{a list of Variable objects to bind together, or a
Dataset object containing only the Variables to bind (as in from subsetting
a Dataset), or values (e.g. names) of variables corresponding to \code{key}.
If omitted, must supply \code{dataset} and \code{pattern}. If specifying
values, must include \code{dataset}.}

\item{dataset}{Argument no longer supported}

\item{name}{character, the name that the new Categorical Array variable
should have. Required.}

\item{...}{Optional additional attributes to set on the new variable.}

\item{selections}{character, for \code{makeMR}, the names of the
categories to mark as the dichotomous selections. Required for
\code{makeMR}; ignored in \code{makeArray}.}
}
\value{
A VariableDefinition that when added to a Dataset will create the
categorical-array or multiple-response variable.
}
\description{
Make a Categorical Array or Multiple Response variable
}

