% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical-array.R, R/multiple-response.R
\name{makeArray}
\alias{makeArray}
\alias{makeMR}
\title{Make a Categorical Array or Multiple Response variable}
\usage{
makeArray(subvariables, dataset = NULL, pattern = NULL,
  key = namekey(dataset), name, ...)

makeMR(subvariables, dataset = NULL, pattern = NULL,
  key = namekey(dataset), name, selections, ...)
}
\arguments{
\item{subvariables}{a list of Variable objects to bind together, or a
Dataset object containing only the Variables to bind (as in from subsetting
a Dataset), or values (e.g. names) of variables corresponding to \code{key}.
If omitted, must supply \code{dataset} and \code{pattern}. If specifying
values, must include \code{dataset}.}

\item{dataset}{the Crunch Dataset to which the variables in
\code{subvariables} belong, or in which to search for variables based
on \code{pattern}. If omitted, \code{subvariables} must exist and all
Variables in the list must belong to the same Dataset}

\item{pattern}{An optional regular expression to search for variables to
bind within \code{dataset}.
Note that this argument is deprecated. If you wish to grep, you can
\code{grep(pattern, aliases(variables(dataset)))} or similar outside this
function.}

\item{key}{character, the name of the Variable field in which to search
with \code{pattern}. Default is 'alias'.}

\item{name}{character, the name that the new Categorical Array variable
should have. Required.}

\item{...}{Optional additional attributes to set on the new variable.}

\item{selections}{character, for \code{makeMR}, the names of the
categories to mark as the dichotomous selections. Required for
\code{makeMR}; ignored in \code{makeArray}.}
}
\value{
A VariableDefinition that when added to a Dataset will create the
categorical-array or multiple-response variable.
}
\description{
Make a Categorical Array or Multiple Response variable
}

