% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{makeArrayGadget}
\alias{makeArrayGadget}
\title{Launch array builder gadget}
\usage{
makeArrayGadget(env = globalenv())
}
\arguments{
\item{env}{the environment to run the gadget}
}
\value{
a valid call to \code{makeArray()} or \code{makeMR()}
}
\description{
Categorical Array and Multiple Response variables can be difficult to
construct without being able to investigate the available variables, and
their categories. This shiny gadget lets you select subvariables from the
dataset list, and ensures that those variables have consistent categories. To
use the gadget you must have at least one CrunchDataset loaded into the global
environment.
}
