% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-variables.R
\name{makeCaseVariable}
\alias{makeCaseVariable}
\title{Make a case variable}
\usage{
makeCaseVariable(cases, ...)
}
\arguments{
\item{cases}{A list of objects of class \code{\link{Case}}}

\item{...}{other properties to pass about the case variable (e.g. name)}
}
\description{
The \code{makeCaseVariable} function derives a variable using values from other
variables. These are evaluated in the order they are supplied in the list
as the \code{cases} argument (they proceed in an IF, ELSE IF, ELSE IF, ..., ELSE
fashion); the first one that matches selects the corresponding value from
the \code{\link{Case}} object.
}
