% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtotals-and-headings.R,
%   R/summary-insertions.R
\name{makeInsertion}
\alias{makeInsertion}
\alias{makeInsertion,Subtotal-method}
\alias{makeInsertion,Heading-method}
\alias{makeInsertion,Insertion-method}
\alias{makeInsertion,SummaryStat-method}
\title{Convert a child class of Insertion into a proper Insertion}
\usage{
\S4method{makeInsertion}{Subtotal}(x, var_categories)

\S4method{makeInsertion}{Heading}(x, var_categories)

\S4method{makeInsertion}{Insertion}(x, var_categories)

\S4method{makeInsertion}{SummaryStat}(x, var_categories)
}
\arguments{
\item{x}{an object the is a child of \link[=Insertions]{Insertion} (e.g.
\code{Subtotal}, \code{Heading})}

\item{var_categories}{categories (from \code{categories(variable)}) to used by
\code{makeInsertions} to make \code{Insertions} from a child object (e.g. \code{Subtotal} or
\code{Heading})}
}
\value{
an Insertion object
}
\description{
The Crunch API expects that \link[=SubtotalsHeadings]{Subtotals, Headings}, and
other insertions all have the same shape. Before sending insertions to the
server, we need to call \code{makeInsertion} on any insertions to make sure they
are proper \code{Insertion}s. This process sometimes requires a variable's
categories object (e.g. in order to convert from category names to category
ids)
}
\keyword{internal}
