% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{makeInteractions}
\alias{makeInteractions}
\title{Make case statements that define the interaction of two or more other variables}
\usage{
makeInteractions(..., sep)
}
\arguments{
\item{...}{A sequence of categorical variables to make an interaction of}

\item{sep}{a character to separate the values of the individual variables}
}
\value{
A list of cases that can be passed to \code{\link{makeCaseVariable}} to make
the interaction variable.
}
\description{
Make case statements that define the interaction of two or more other variables
}
\keyword{internal}
