% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{makeWeight}
\alias{makeWeight}
\title{Generate a weight vector}
\usage{
makeWeight(..., name)
}
\arguments{
\item{...}{A series of expressions of the form \code{variable ~ target_weights}. The variable must
be a categorical crunch variable, and the target weights must be a numeric vector whose
length is equal to the number of categories contained in the variable, and whose sum is equal to 100.}

\item{name}{The name of the resulting variable}
}
\value{
A crunch Variable Definiton of the weight variable
}
\description{
This function allows you to generate a weight vector by supplying a set of
categorical variables and the target distribution for each of the variable's categories.
For instance if you wanted to create a weight variable which equally weighted four categories stored
in \code{ds$var} you would call \code{makeWeight(ds$var ~ c(25, 25, 25, 25), name = "weight")}.
}
\examples{
\dontrun{
mtcars$cyl <- as.factor(mtcars$cyl)
mtcars$gear <- as.factor(mtcars$gear)
ds <- newDataset(mtcars, "mtcars")
ds$weight <- makeWeight(ds$cyl ~ c(30, 30, 40, 0), ds$gear ~ c(20, 20, 60, 0), name = "weight" )
as.vector(ds$weight)
}
}
