% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{makeWeight}
\alias{makeWeight}
\title{Generate a weight vecotr}
\usage{
makeWeight(..., name)
}
\arguments{
\item{...}{A series of expressions of the form \code{variable ~ target_weights}. The variable must
be a categorical crunch variable, and the target weights must be a numeric vector whose
length is equal to the number of categories contained in the variable, and whose sum is equal to 100.}

\item{name}{The name of the resulting variable}
}
\value{
A crunch Variable Definiton of the weight variable
}
\description{
This function allows you to generate a weight vector by supplying a set of
categorical variables and the target distribution for each variable's category.
For instance if you wanted to create a weight variable which equally weighted a four categories of the
categorical variable \code{ds$var} you would call \code{makeWeight(ds$var ~ c(25, 25, 25, 25), name = weight)}.
}
\examples{
\dontrun{
cars <- mtcars
cars$cyl <- as.factor(cars$cyl)
cars$gear <- as.factor(vars$gear)
ds <- newDataset(cars, "cars_fact")
makeWeight(ds$cyl ~ c(30, 30, 40, 0), ds$gear ~ c(20, 20, 60, 0), name = "weight" )
as.vector(ds$weight)
}
}
