% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame.R
\name{merge.CrunchDataFrame}
\alias{merge.CrunchDataFrame}
\title{Merge a CrunchDataFrame}
\usage{
\method{merge}{CrunchDataFrame}(x, y, by.x, by.y)
}
\arguments{
\item{x}{a CrunchDataFrame}

\item{y}{a standard data.frame}

\item{by.x}{name of the variable to match}

\item{by.y}{name of the variable to match}
}
\description{
Merge a CrunchDataFrame
}
