% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fork-and-merge.R
\name{mergeFork}
\alias{mergeFork}
\title{Merge changes to a dataset from a fork}
\usage{
mergeFork(dataset, fork, autorollback = TRUE)
}
\arguments{
\item{dataset}{The \code{CrunchDataset} to merge to}

\item{fork}{The \code{CrunchDataset}, perhaps forked from \code{dataset},
that is to be merged in.}

\item{autorollback}{logical If the merge fails, should \code{dataset} be
restored to its state prior to the merge, or should it be left in its
partially merged state for debugging and manual fixing? Default is
\code{TRUE}, i.e. the former.}
}
\value{
\code{dataset} with changes from \code{fork} merged to it.
}
\description{
Merge changes to a dataset from a fork
}

