% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\docType{methods}
\name{modifyWeightVariables,CrunchDataset-method}
\alias{modifyWeightVariables,CrunchDataset-method}
\alias{modifyWeightVariables,VariableCatalog-method}
\alias{modifyWeightVariables}
\title{modifyWeightVariables}
\usage{
\S4method{modifyWeightVariables}{CrunchDataset}(x, vars, type = "append")

\S4method{modifyWeightVariables}{VariableCatalog}(x, vars, type = "append")

modifyWeightVariables(x, vars, type = "append")
}
\arguments{
\item{x}{a dataset or variable catalog}

\item{vars}{Variables to add or remove  this can be a numeric Crunch variable,
list of numeric Crunch variables or a character vector with the aliases of numeric Crunch variables.}

\item{type}{a character string determining how the weightVariables
will be modified:
\itemize{
\item \code{"append"} : add \code{vars} to the current weight variables
\item \code{"remove"} : remove \code{vars} from the current list of weight variables
\item \code{"replace"}: replace the current weight variables with \code{vars}
}}
}
\description{
Change which variables can be set as a dataset's weight.
}
