% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{na-omit-categories}
\alias{na-omit-categories}
\alias{na.omit,Categories-method}
\title{Omit missing categories}
\usage{
\S4method{na.omit}{Categories}(object, ...)
}
\arguments{
\item{object}{Categories}

\item{...}{additional arguments, ignored}
}
\value{
\code{object} with any categories that have missing: TRUE excluded
}
\description{
Omit missing categories
}
