% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-dataset.R
\name{newDataset}
\alias{newDataset}
\alias{newDatasetByCSV}
\title{Upload a data.frame to Crunch to make a new dataset}
\usage{
newDataset(x, name = deparseAndFlatten(substitute(x), max_length = 40), ...)

newDatasetByCSV(...)
}
\arguments{
\item{x}{a data.frame}

\item{name}{the name to give the new Crunch dataset. By default the function uses the
name of the R object.}

\item{...}{additional arguments passed to \code{ \link{createDataset}}}
}
\value{
If successful, an object of class CrunchDataset.
}
\description{
This function uploads an R dataframe to Crunch. It uses the CSV+JSON import format,
which is generally faster and more effective than \link{newDatasetByColumn}. If you
want to upload an SPSS file, it is better to use \link{newDatasetFromFile} to upload
the file directly to the server rather than first reading it into R. Uploading
SPSS files directly to Crunch will preserve metadata which is stripped by the R import.
}
\examples{
\dontrun{
ds <- newDataset(mtcars, "cars")
}
}
