% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-dataset.R
\name{newDatasetByCSV}
\alias{newDatasetByCSV}
\title{Upload a data.frame to Crunch to make a new dataset}
\usage{
newDatasetByCSV(x, name = deparse(substitute(x))[1], ...)
}
\arguments{
\item{x}{a data.frame or other rectangular R object}

\item{name}{character, the name to give the new Crunch dataset. Default is
the name of the R object passed in \code{x}}

\item{...}{additional arguments passed to \code{ \link{createDataset}}}
}
\value{
If successful, an object of class CrunchDataset.
}
\description{
This function uses the CSV+JSON import format, which is faster and more
effective for certain dataset sizes and shapes than
\code{\link{newDatasetByColumn}}.
}
\seealso{
\code{\link{newDataset}} \code{\link{newDatasetByColumn}}
}

