% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-dataset.R
\name{newDatasetFromFile}
\alias{newDatasetFromFile}
\title{Upload a file to Crunch to make a new dataset}
\usage{
newDatasetFromFile(file, name = basename(file), ...)
}
\arguments{
\item{file}{character, the path to a file to upload. This should either be
a \code{.csv} or \code{.sav} (SPSS) file}

\item{name}{character, the name to give the new Crunch dataset. By default the name of the dataset
will be the filename}

\item{...}{additional arguments passed to \link{createDataset}}
}
\value{
On success, an object of class \link{CrunchDataset}
}
\description{
This function allows you to upload a \code{.csv`` or}.sav\code{file directly to Crunch without first reading it into R. This is useful both because it preserves SPSS metadata which is stripped by importing}.sav` files into R.
}
