% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-dataset.R
\name{newDatasetFromFile}
\alias{newDatasetFromFile}
\title{Upload a file to Crunch to make a new dataset}
\usage{
newDatasetFromFile(file, name = basename(file), ...)
}
\arguments{
\item{file}{character, the path to a local file to upload, or a URL.
This should either be a .csv or .sav (SPSS) file.}

\item{name}{character, the name to give the new Crunch dataset. Default is
the file name}

\item{...}{additional arguments passed to \code{ \link{createDataset}}}
}
\value{
On success, an object of class \code{CrunchDataset}.
}
\description{
Use this import method if you have an SPSS data file. Reading such a file
into R as a data.frame will result in lost metadata. You can just send it
directly to Crunch and let the server process it.
}
