% Generated by roxygen2 (4.0.1): do not edit by hand
\name{newDatasetFromFile}
\alias{newDatasetFromFile}
\title{Upload a file to Crunch to make a new dataset}
\usage{
newDatasetFromFile(file, name = basename(file),
  useAlias = default.useAlias(), ...)
}
\arguments{
\item{file}{character, the path to a file to upload}

\item{name}{character, the name to give the new Crunch dataset. Default is
the file name}

\item{useAlias}{logical whether variable alias or name should be used as R
variable names when the dataset is returned. Default is TRUE, meaning alias.
They're more computer friendly.}

\item{...}{additional arguments, currently not implemented}
}
\value{
If successful, an object of class crunchdf.
}
\description{
Upload a file to Crunch to make a new dataset
}

