% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{newSlide}
\alias{newSlide}
\title{Append a new slide to a Crunch Deck}
\usage{
newSlide(deck, query, display_settings = list(), title = "",
  subtitle = "", ...)
}
\arguments{
\item{deck}{A Crunch Deck}

\item{query}{A formula definition of a query to be used by the slide. This is
similar to CrunchCube query}

\item{display_settings}{(optional) A list of display settings, if om}

\item{title}{The slide's title}

\item{subtitle}{The slide's subtitle}

\item{...}{Further options to be passed on to the API}
}
\value{
CrunchSlide object
}
\description{
Append a new slide to a Crunch Deck
}
\examples{
\dontrun{
newSlide(
    main_deck,
    ~ cyl + wt,
    title = "Cyl and Weight",
    subtitle = "2017 Data"
)
}
}
