% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\docType{methods}
\name{ordering}
\alias{ordering}
\alias{ordering<-}
\alias{ordering,CrunchDataset-method}
\alias{ordering<-,CrunchDataset-method}
\alias{ordering,VariableCatalog-method}
\alias{ordering<-,VariableCatalog-method}
\alias{ordering,DatasetCatalog-method}
\alias{ordering,CrunchProject-method}
\alias{ordering,ProjectFolder-method}
\alias{ordering<-,DatasetCatalog-method}
\alias{ordering<-,CrunchProject-method}
\title{Get and set VariableOrder}
\usage{
\S4method{ordering}{CrunchDataset}(x)

\S4method{ordering}{CrunchDataset}(x) <- value

\S4method{ordering}{VariableCatalog}(x)

\S4method{ordering}{VariableCatalog}(x) <- value

\S4method{ordering}{DatasetCatalog}(x)

\S4method{ordering}{CrunchProject}(x)

\S4method{ordering}{ProjectFolder}(x)

\S4method{ordering}{DatasetCatalog}(x) <- value

\S4method{ordering}{CrunchProject}(x) <- value
}
\arguments{
\item{x}{a VariableCatalog or CrunchDataset}

\item{value}{a valid VariableOrder object}
}
\value{
\code{ordering} returns a VariableOrder object, while
\code{ordering<-} sets the VariableOrder
}
\description{
The \code{ordering} methods allow you to get and set a \code{\link{VariableOrder}} on a
\code{\link{CrunchDataset}} or on the \code{\link{VariableCatalog}} that the dataset contains.
}
