% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ordering.R
\docType{methods}
\name{ordering}
\alias{ordering}
\alias{ordering,CrunchDataset-method}
\alias{ordering,VariableCatalog-method}
\alias{ordering<-}
\alias{ordering<-,CrunchDataset-method}
\alias{ordering<-,VariableCatalog-method}
\title{Get and set VariableOrder}
\usage{
\S4method{ordering}{CrunchDataset}(x)

\S4method{ordering}{CrunchDataset}(x) <- value

\S4method{ordering}{VariableCatalog}(x)

\S4method{ordering}{VariableCatalog}(x) <- value
}
\arguments{
\item{x}{a VariableCatalog or CrunchDataset}

\item{value}{a valid VariableOrder object}
}
\value{
\code{ordering} returns a VariableOrder object, while
\code{ordering<-} sets the VariableOrder in \code{value} on \code{x}
}
\description{
The \code{ordering} methods allow you to get and set a
\code{\link{VariableOrder}} on
a \code{\link{CrunchDataset}} or on the \code{\link{VariableCatalog}} that
the dataset contains.
}

