% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-comparisons.R
\name{pairwiseMatrix}
\alias{pairwiseMatrix}
\title{Matrix of Chi-Squared Statistics for all rows or columns}
\usage{
pairwiseMatrix(cube, dim = c("cols", "rows"), value = c("statistic",
  "p.value"))
}
\arguments{
\item{cube}{}

\item{dim}{which dimension along which to compare (\code{cols} or \code{rows})}
}
\value{
A symmetric square matrix of all column-comparison or row-comparison
\eqn{\Chi^2} statistics. Typical element \eqn{i,j} is the test statistic equivalent
to \code{chisq.test} subsetting to just columns i and j (for dim="cols").
}
\description{
Genearte a matrix of pairwise comparisons of rows or columns, each against
the others.
}
\examples{
\dontrun{
some_cube <- crunch_example_data(cat_by_cat)

# By default, return the 'statistic' of chisq.test, evaluated
# for each pair of columns pairwise:
pairwiseMatrix(some_cube)

# or each pair of rows:
pairwiseMatrix(some_cube, dim="rows")

# Or, specify another "value" from `chisq.test`:
pairwiseMatrix(some_cube, value="p.value")
}
}
