% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-stream.R
\name{pendingStream}
\alias{pendingStream}
\title{Get the pending streams for a dataset}
\usage{
pendingStream(ds)
}
\arguments{
\item{ds}{a CrunchDataset}
}
\value{
number of pending messages in the stream for the dataset
}
\description{
retrieves the number of pending messages. Use \code{\link[=appendStreamedRows]{appendStreamedRows()}} to
append all pending streamed rows to the dataset.
}
