% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{population}
\alias{population}
\alias{popSize}
\alias{popMagnitude}
\alias{setPopulation}
\alias{popSize,CrunchDataset-method}
\alias{popMagnitude,CrunchDataset-method}
\alias{setPopulation,CrunchDataset-method}
\title{Get and set the market size for Crunch datasets}
\usage{
\S4method{popSize}{CrunchDataset}(x)

\S4method{popMagnitude}{CrunchDataset}(x)

\S4method{setPopulation}{CrunchDataset}(x, size, magnitude)
}
\arguments{
\item{x}{a Dataset}

\item{size}{the population size}

\item{magnitude}{the order of magnitude with which to display the population
size. Must be either 3, 6, or 9 for thousands, millions, and billions respectively.}
}
\value{
\code{popSize} and \code{popMagnitude} return the respective population size and
magnitude. \code{setPopultion} returns the modified dataset.
}
\description{
Get and set the market size for Crunch datasets
}
