% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{population}
\alias{population}
\alias{population<-}
\alias{popSize,CrunchDataset-method}
\alias{popSize<-,CrunchDataset-method}
\alias{popMagnitude,CrunchDataset-method}
\alias{popMagnitude<-,CrunchDataset-method}
\title{Get and set the market size for Crucn}
\usage{
\S4method{popSize}{CrunchDataset}(x)

\S4method{popSize}{CrunchDataset}(x) <- value

\S4method{popMagnitude}{CrunchDataset}(x)

\S4method{popMagnitude}{CrunchDataset}(x) <- value
}
\arguments{
\item{x}{a Dataset}

\item{value}{For the setters, a length-1 character vector to assign}
}
\value{
Getters return the character object in the specified slot; setters
return \code{x} duly modified.
}
\description{
Get and set the market size for Crucn
}
