% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchbox.R
\name{preCrunchBoxCheck}
\alias{preCrunchBoxCheck}
\title{Check if a dataset will make a good CrunchBox}
\usage{
preCrunchBoxCheck(dataset)
}
\arguments{
\item{dataset}{CrunchDataset, potentially subsetted on variables}
}
\value{
Invisibly, the dataset. Called for side-effect of printing things.
}
\description{
CrunchBoxes allows you to share data with the world in a simple, easy to embed format.
However, not all datasets naturally translate to the CrunchBox format. This
function checks your dataset to see if it
}
\seealso{
\code{\link{crunchBox}}
}
