\name{prepareBindInputs}
\alias{prepareBindInputs}
\title{Given inputs to makeArray/makeMR, parse and validate}
\usage{
  prepareBindInputs(list_of_variables, dataset = NULL,
    pattern = NULL, key = "alias", ...)
}
\arguments{
  \item{...}{Stuff from calling function that will be
  ignored.}
}
\description{
  Given inputs to makeArray/makeMR, parse and validate
}

