% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects}
\alias{projects}
\title{Get the project catalog}
\usage{
projects(x = getAPIRoot())
}
\arguments{
\item{x}{a \code{ShojiObject} that has a project catalog associated. If omitted,
the default value for \code{x} means that you will load the user's primary
project catalog. (Currently, there are no other project catalogs to load.)}
}
\value{
An object of class \code{ProjectCatalog}.
}
\description{
Get the project catalog
}
\examples{
\dontrun{
myprojects <- projects()
proj <- myprojects[["Project name"]]
}
}

