% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/shoji.R, R/variable.R
\docType{methods}
\name{refresh}
\alias{refresh}
\alias{refresh,CrunchDataset-method}
\alias{refresh,ShojiObject-method}
\alias{refresh,CrunchVariable-method}
\title{Get a fresh copy from the server}
\usage{
\S4method{refresh}{CrunchDataset}(x)

\S4method{refresh}{ShojiObject}(x)

\S4method{refresh}{CrunchVariable}(x)
}
\arguments{
\item{x}{pretty much any Crunch object}
}
\value{
a new version of \code{x}
}
\description{
Crunch objects usually keep themselves in sync with the server when you
manipulate them, but sometimes they can drift. Maybe someone else has
modified the dataset you're working on, or maybe
you have modified a variable outside of the context of its dataset.
refresh() allows you to get back in sync.
}
