% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise.R
\name{rowCount}
\alias{rowCount}
\title{Create variables based on row-wise functions for crunch Multiple Response Variables}
\usage{
rowCount(x, name, ...)
}
\arguments{
\item{x}{A crunch variable or expression}

\item{name}{a character to use as the name of the case variable to create}

\item{...}{description, alias, and other metadata passed to \code{\link[=VarDef]{VarDef()}}}
}
\value{
A Variable Definition
}
\description{
Quickly generate new variables that are based on row-wise summaries of Multiple Response
Variables.
}
\seealso{
\code{\link{expressions}} for the more flexible expressions that power
these functions and \code{\link[=rowDistinct]{rowDistinct()}} for other row-wise functions
}
