% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{saveVersion}
\alias{saveVersion}
\title{Create a new saved version}
\usage{
saveVersion(dataset, description = paste("Version", length(versions(dataset))
  + 1))
}
\arguments{
\item{dataset}{a \code{CrunchDataset}}

\item{description}{character name to give the saved version, as in a
commit message. You are encouraged, though not strictly required, to give
versions unique descriptions.}
}
\value{
invisibly, the URL of the newly created version
}
\description{
Create a new saved version
}
\seealso{
\code{\link{versions}} \code{\link{restoreVersion}}
}

