% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation.R
\name{script-catalog}
\alias{script-catalog}
\alias{scripts}
\alias{scripts,CrunchDataset-method}
\title{Crunch Automation scripts entities for a dataset}
\usage{
\S4method{scripts}{CrunchDataset}(x)
}
\arguments{
\item{x}{a CrunchDataset}
}
\value{
an object of class "ScriptCatalog" containing references to
Script entities.
}
\description{
Crunch Automation scripts entities for a dataset
}
