% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji.R, R/variable.R
\docType{methods}
\name{self}
\alias{self}
\alias{self,CrunchVariable-method}
\alias{self,ShojiObject-method}
\title{Get the URL of this object}
\usage{
\S4method{self}{ShojiObject}(x)

\S4method{self}{CrunchVariable}(x)
}
\arguments{
\item{x}{a Crunch object}
}
\value{
the URL for \code{x}
}
\description{
Get the URL of this object
}

