% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shoji.R
\docType{methods}
\name{self}
\alias{self}
\alias{self,ShojiObject-method}
\title{Get the URL of this object}
\usage{
\S4method{self}{ShojiObject}(x)
}
\arguments{
\item{x}{a Crunch object}
}
\value{
the URL for \code{x}
}
\description{
Get the URL of this object
}

