% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\docType{methods}
\name{setNames}
\alias{setNames}
\alias{setNames,VariableFolder-method}
\title{Change the name of the objects contained in the current folder}
\usage{
setNames(object, nm)

\S4method{setNames}{VariableFolder}(object, nm)
}
\arguments{
\item{object}{A \code{VariableFolder}}

\item{nm}{A character vector of new names of the same length as the number
of objects in the folder}
}
\value{
\code{object}, with the names of its children duly changed
}
\description{
If you want to rename all of the objects (variables, folders, etc.), you
can use \code{setNames()} to do so. It doesn't move variables or change anything
other than the names of the objects in the current folder.
}
\examples{
\dontrun{
ds <- ds \%>\%
    cd("Demographics") \%>\%
    setNames(c("Gender (4 category)", "Birth year", "Race (5 category)"))
}

}
\seealso{
\code{\link[=cd]{cd()}} and \code{\link[=mv]{mv()}}
}
