% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{share}
\alias{share}
\title{Share a dataset}
\usage{
share(dataset, users, edit = FALSE, notify = TRUE, message = NULL)
}
\arguments{
\item{dataset}{a CrunchDataset}

\item{users}{character: email address(es) or URLs of the users or teams with
whom to share the dataset. If there is no Crunch user associated with an
email, an invitation will be sent.}

\item{edit}{logical: should the specified user(s) be given edit privileges
on the dataset? Default is \code{FALSE}. \code{edit} can be a single value
or, if inviting multiple users, a vector of logical values of equal length
of the number of emails given.}

\item{notify}{logical: should users who are getting new privileges on this
dataset be sent an email informing them of this fact? Default is
\code{TRUE}.}

\item{message}{character: what message should users who are getting notified
about new permissions on this dataset receive?}
}
\value{
Invisibly, the dataset.
}
\description{
Share a dataset
}
\seealso{
\code{\link{unshare}}
}
