% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoji-catalog.R
\name{shoji-index}
\alias{shoji-index}
\alias{index}
\alias{index<-}
\alias{index,ShojiCatalog-method}
\alias{index<-,ShojiCatalog-method}
\title{Get the body of a Catalog}
\usage{
\S4method{index}{ShojiCatalog}(x)

\S4method{index}{ShojiCatalog}(x) <- value
}
\arguments{
\item{x}{a Catalog (VariableCatalog, Subvariables, or similar object)}

\item{value}{For the setters, an appropriate-length list to
assign}
}
\value{
Getters return the list object in the "index" slot; setters
return \code{x} duly modified.
}
\description{
The core of Catalog data is in its "index". These methods get and set that
slot.
}
