% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ShojiObject-class}
\alias{ShojiObject}
\alias{ShojiObject-class}
\title{Mix-in class for multiple inheritance of variables and datasets.}
\description{
Exists for common methods in interacting with Crunch API only. Has no
Extract methods declared so as not to conflict with the
vector/list/data.frame methods jointly inherited in CrunchVariable and
CrunchDataset.
}

