% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R, R/tab-book.R
\docType{methods}
\name{show-crunch}
\alias{show-crunch}
\alias{show,ShojiObject-method}
\alias{show,CrunchVariable-method}
\alias{show,Category-method}
\alias{show,Categories-method}
\alias{show,CrunchExpr-method}
\alias{show,CrunchLogicalExpr-method}
\alias{show,CrunchCube-method}
\alias{show,OrderGroup-method}
\alias{show,MultitableResult-method}
\title{Show methods for Crunch objects}
\usage{
\S4method{show}{ShojiObject}(object)

\S4method{show}{CrunchVariable}(object)

\S4method{show}{Category}(object)

\S4method{show}{Categories}(object)

\S4method{show}{CrunchExpr}(object)

\S4method{show}{CrunchLogicalExpr}(object)

\S4method{show}{CrunchCube}(object)

\S4method{show}{OrderGroup}(object)

\S4method{show}{MultitableResult}(object)
}
\arguments{
\item{object}{the object}
}
\value{
invisibly
}
\description{
Show methods for Crunch objects
}
\seealso{
\code{\link[methods]{show}}
}
