% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R, R/pretty-print.R, R/show.R,
%   R/slides.R, R/tab-book.R
\docType{methods}
\name{show,DeckCatalog-method}
\alias{show,DeckCatalog-method}
\alias{show,CrunchDeck-method}
\alias{show,ShojiFolder-method}
\alias{show-crunch}
\alias{show,ShojiObject-method}
\alias{show,CrunchVariable-method}
\alias{show,Category-method}
\alias{show,Categories-method}
\alias{show,Insertion-method}
\alias{show,Insertions-method}
\alias{show,CrunchExpr-method}
\alias{show,CrunchLogicalExpr-method}
\alias{show,CrunchCube-method}
\alias{show,OrderGroup-method}
\alias{show,CrunchGeography-method}
\alias{show,CrunchSlide-method}
\alias{show,MultitableResult-method}
\title{Show methods for Crunch objects}
\usage{
\S4method{show}{DeckCatalog}(object)

\S4method{show}{CrunchDeck}(object)

\S4method{show}{ShojiFolder}(object)

\S4method{show}{ShojiObject}(object)

\S4method{show}{CrunchVariable}(object)

\S4method{show}{Category}(object)

\S4method{show}{Categories}(object)

\S4method{show}{Insertion}(object)

\S4method{show}{Insertions}(object)

\S4method{show}{CrunchExpr}(object)

\S4method{show}{CrunchLogicalExpr}(object)

\S4method{show}{CrunchCube}(object)

\S4method{show}{OrderGroup}(object)

\S4method{show}{CrunchGeography}(object)

\S4method{show}{CrunchSlide}(object)

\S4method{show}{MultitableResult}(object)
}
\arguments{
\item{object}{the object}
}
\value{
invisibly
}
\description{
Show methods for Crunch objects
}
\seealso{
\code{\link[methods:show]{methods::show}}
}
