% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/category.R, R/show.R, R/variable-order.R
\docType{methods}
\name{show,Category-method}
\alias{show,CategoricalArrayVariable-method}
\alias{show,Categories-method}
\alias{show,Category-method}
\alias{show,CrunchCube-method}
\alias{show,CrunchDataset-method}
\alias{show,CrunchVariable-method}
\alias{show,Subvariables-method}
\alias{show,VariableOrder-method}
\title{Show methods for Crunch objects}
\usage{
\S4method{show}{Category}(object)

\S4method{show}{CrunchVariable}(object)

\S4method{show}{CategoricalArrayVariable}(object)

\S4method{show}{Categories}(object)

\S4method{show}{CrunchDataset}(object)

\S4method{show}{Subvariables}(object)

\S4method{show}{CrunchCube}(object)

\S4method{show}{VariableOrder}(object)
}
\arguments{
\item{object}{the object}
}
\value{
invisibly
}
\description{
Show methods for Crunch objects
}
\seealso{
\code{\link[methods]{show}}
}

