% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-transforms.R
\docType{methods}
\name{showTransforms,CrunchCube-method}
\alias{showTransforms,CrunchCube-method}
\title{Show the variable transformations on a CrunchCube}
\usage{
\S4method{showTransforms}{CrunchCube}(x)
}
\arguments{
\item{x}{a CrunchCube}
}
\value{
an array with any transformations applied
}
\description{
Show the variable transformations on a CrunchCube
}
