% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\docType{methods}
\name{slides}
\alias{slides}
\alias{slides<-}
\alias{slides,CrunchDeck-method}
\alias{slides<-,CrunchDeck-method}
\title{Access the slides of a CrunchDeck}
\usage{
slides(x)

slides(x) <- value

\S4method{slides}{CrunchDeck}(x)

\S4method{slides}{CrunchDeck}(x) <- value
}
\arguments{
\item{x}{a CrunchDeck}

\item{value}{a \code{SlideCatalog} or \code{CrunchSlide} to add}
}
\value{
a \code{SlideCatalog}
}
\description{
Return a \code{SlideCatalog} from a \code{CrunchDeck}. All slide catalog methods should be
available for \code{CrunchDecks}, but this function is used internally to model the
API.
}
