% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO-survey.R
\docType{data}
\name{SO_schema}
\alias{SO_schema}
\title{Schema for the 2017 Stack Overflow developer survey}
\format{
A data frame with 23 rows and 2 variables.
\describe{
\item{Column}{The column name of the survey data frame}
\item{Question}{Question asked of respondents}
}
}
\usage{
SO_schema
}
\description{
Survey questions and variable names for the 2017 Stack Overflow Developers
Survey #'
}
\keyword{datasets}
