% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-stream.R
\docType{methods}
\name{streaming}
\alias{streaming}
\alias{getPendingMessages}
\alias{streamRows}
\alias{appendStreamedRows}
\alias{getPendingMessages,CrunchDataset-method}
\alias{streamRows,CrunchDataset-method}
\alias{appendStreamedRows,CrunchDataset-method}
\title{Tools for working with streaming data}
\usage{
\S4method{getPendingMessages}{CrunchDataset}(x)

\S4method{streamRows}{CrunchDataset}(x, data)

\S4method{appendStreamedRows}{CrunchDataset}(x, messages)
}
\arguments{
\item{x}{CrunchDataset}

\item{data}{a dataframe with data to send as a stream, The given data values
must be in the Crunch I/O format (for example, category ids instead of
names or numeric_values)}

\item{messages}{the number of messages that are pending to be appended to
the dataset (default: NULL, which will append all pending messages)}
}
\value{
nothing?
}
\description{
Crunch allows you to stream data to a dataset. Streaming data is useful for
datasets which have frequent updates (see
\href{http://docs.crunch.io/#streaming-rows}{the Crunch documentation} for more)
information.
}
\examples{
\dontrun{
# need examples!
}
}
