% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-stream.R
\name{streaming}
\alias{streaming}
\alias{streaming<-}
\title{Set the streaming property of a dataset}
\usage{
streaming(x)

streaming(x) <- value
}
\arguments{
\item{x}{a CrunchDataset}

\item{value}{for setting only (values can be: \code{"no"}, \code{"streaming"}, or
\code{"finished"})}
}
\value{
the streaming status
}
\description{
Only datasets that have their streaming property set to "streaming" can
have rows streamed to them. Before attempting to streaming rows (with
\link{streamRows} for example), the dataset has to be set up to stream rows. Use
\code{streaming(ds)} to get the streaming status, and \code{streaming(ds) <- "streaming"} to set the streaming status.
}
