% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-stream.R
\docType{methods}
\name{streaming}
\alias{streaming}
\alias{pendingMessages}
\alias{pendingMessages,CrunchDataset-method}
\title{Tools for working with streaming data}
\usage{
\S4method{pendingMessages}{CrunchDataset}(x)
}
\arguments{
\item{x}{a CrunchDataset}
}
\value{
nothing?
}
\description{
Crunch allows you to stream data to a dataset. Streaming data is useful for
datasets which have frequent updates (see
\href{http://docs.crunch.io/#streaming-rows}{the Crunch documentation} for more)
information.
}
\examples{
\dontrun{
# need examples!
}
}
