% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/subtotals-and-headings.R
\docType{class}
\name{Subtotal-class}
\alias{Subtotal-class}
\alias{subtotals}
\alias{subtotals<-}
\alias{makeInsertion}
\alias{Subtotal}
\alias{Heading-class}
\alias{Heading}
\alias{subtotals,CrunchVariable-method}
\alias{subtotals,VariableTuple-method}
\alias{subtotals<-,CrunchVariable,ANY-method}
\alias{subtotals<-,CrunchVariable,NULL-method}
\alias{makeInsertion,Subtotal-method}
\alias{makeInsertion,Heading-method}
\alias{makeInsertion,ANY-method}
\title{Subtotals and headings}
\usage{
Subtotal(..., data = NULL)

Heading(..., data = NULL)

\S4method{subtotals}{CrunchVariable}(x)

\S4method{subtotals}{VariableTuple}(x)

\S4method{subtotals}{CrunchVariable,ANY}(x) <- value

\S4method{subtotals}{CrunchVariable,`NULL`}(x) <- value

\S4method{makeInsertion}{Subtotal}(x, var)

\S4method{makeInsertion}{Heading}(x, var)

\S4method{makeInsertion}{ANY}(x, var)
}
\arguments{
\item{...}{additional arguments to \code{[}, ignored}

\item{data}{For the constructor functions \code{Subtotal} and
\code{Subtotal}, you can either pass in attributes via \code{...} or you
can create the objects with a fully defined \code{list} representation of
the objects via the \code{data} argument. See the examples.}

\item{x}{either a variable or CrunchCube object to add or get subtotal
transforms for}

\item{value}{For \code{[<-}, the replacement Subtotal to insert}

\item{var}{the variable to use to make \code{Insertions} from a \code{Subtotal} object}
}
\description{
Subtotals and headings allow you to add subtotals and headings to variables
or CrunchCubes. These are especially useful for making aggregates across
multiple categories (sometimes referred to as \emph{nets}, \emph{top box}, or
\emph{top 2 box}).
}
\details{
To see the subtotals or headings set for a variable, use \code{subtotals(variable)}

Subtotals and headings can be added either by passing a list of \code{Subtotal}s
or \code{Heading}s, or they can be added one at a time by passing \code{Subtotal} or
\code{Heading} to \code{subtotals(variable)} alone.

Adding subtotals or headings is additive; meaning that subtotals or headings
that are already set on the variable are not removed when new subtotals or
headings are added. To remove all subtotals and headings, set
\code{subtotals(variable)} to \code{NULL}.

To get an array of just the subtotal rows from a CrunchCube, use the function
\code{subtotalArray(CrunchCube)}.
}
\examples{
\dontrun{
# given a variable ds$opinion, with categories: Strongly Agree, Somewhat
# Agree, Neither Agree nor Disagree, Somewhat Disagree, and Strongly Disagree,
# to make two subtotals for Agree and Disagree:
subtotals(ds$opinion) <- list(
    Subtotal(name = "Agree", categories = c("Strongly Agree", "Somewhat Agree"),
             after = "Somewhat Agree"),
    Subtotal(name = "Disagree", categories = c("Strongly Disagree", "Somewhat Disagree"),
             after = "Strongly Disagree")
)

# headings can also be added:
subtotals(ds$opinion) <- Heading(name = "All opinions", after = 0)

# to see the subtotals and headings associated with a variable
subtotals(ds$opinion)

# to remove all subtotals and headings
subtotals(ds$opinion) <- NULL
}

}
