% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Subvariables-class}
\alias{Subvariables}
\alias{Subvariables-class}
\alias{names,Subvariables-method}
\alias{names<-,Subvariables-method}
\alias{subvariables,CategoricalArrayVariable-method}
\alias{subvariables<-,CategoricalArrayVariable,ANY-method}
\title{Subvariables in Array Variables}
\usage{
\S4method{subvariables}{CategoricalArrayVariable}(x)

\S4method{subvariables}{CategoricalArrayVariable,ANY}(x) <- value

\S4method{names}{Subvariables}(x)

\S4method{names}{Subvariables}(x) <- value
}
\arguments{
\item{x}{an \R object.}

\item{value}{a character vector of up to the same length as \code{x}, or
    \code{NULL}.}
}
\description{
Multiple-response and categorical-array variables contain a set of
subvariables within them.
}

