% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-book.R
\name{tabBook}
\alias{tabBook}
\title{Compute a Tab Book}
\usage{
tabBook(
  multitable,
  dataset,
  weight = crunch::weight(dataset),
  output_format = c("json", "xlsx"),
  file = NULL,
  filter = NULL,
  use_legacy_endpoint = envOrOption("use.legacy.tabbook.endpoint", FALSE),
  append_default_wt = TRUE,
  ...
)
}
\arguments{
\item{multitable}{a \code{Multitable} object}

\item{dataset}{CrunchDataset, which may be subset with a filter expression
on the rows, and a selection of variables to use on the columns.}

\item{weight}{a CrunchVariable that has been designated as a potential
weight variable for \code{dataset}, or \code{NULL} for unweighted results.
Default is the currently applied \code{\link{weight}}. Additionally, weights can be
set on a per variable basis for json export only. To do so, specify the weight
as either a list (which will be passed to \code{tabBookWeightSpec()}, or a data.frame
that mimics the structure. See \code{\link[=tabBookWeightSpec]{tabBookWeightSpec()}} for more details.}

\item{output_format}{character export format: currently supported values are "json"
(default) and "xlsx".}

\item{file}{character local filename to write to. A default filename will be
generated from the \code{multitable}'s name if one is not supplied and the
"xlsx" format is requested. Not required for "json" format export.}

\item{filter}{a Crunch \code{filter} object or a vector of names
of \code{\link{filters}} defined in the dataset.}

\item{use_legacy_endpoint}{Logical, indicating whether to use a 'legacy'
endpoint for compatibility (this endpoint will be removed in the future).
Defaults to \code{FALSE}, but can be set in the function, or with the environment
variable \code{R_USE_LEGACY_TABBOOK_ENDPOINT} or R option
\code{use.legacy.tabbook.endpoint}.}

\item{append_default_wt}{passed to \code{\link[=tabBookWeightSpec]{tabBookWeightSpec()}}}

\item{...}{Additional "options" passed to the tab book POST request.
More details can be found
\href{https://docs.crunch.io/endpoint-reference/endpoint-multitable.html#options}{in the crunch API documentation}
or \href{https://docs.crunch.io/endpoint-reference/endpoint-tabbook.html#options}{for the legacy endpoint}}
}
\value{
If "json" format is requested, the function returns an object of
class \code{TabBookResult}, containing a list of \code{MultitableResult}
objects, which themselves contain \code{CrunchCube}s. For single weight tabbook,
the variables are always sorted in the order the server stores them in,
but complex weights are sorted in the order of the data.frame given. If
"xlsx" is requested, the function invisibly returns the filename (\code{file},
if specified, or the the autogenerated file name). If you request "json" and
wish to access the JSON data underlying the \code{TabBookResult}, pass in a path
for \code{file} and you will get a JSON file written there as well.
}
\description{
This function allows you to generate a tab book from a multitable and data.
As with other functions, you can select the rows and columns you want to
work with by subsetting the \code{dataset} you pass into the function.
}
\details{
By specifying a "json" \code{format}, instead of generating an Excel
workbook, you'll get a TabBookResult object, containing nested CrunchCube
results. You can then further format these and construct custom tab reports.

Tabbook pages are organized in the order the variables are stored by the
server, unless complex weights are specified, in which case, the variables
are sorted in the order of the weight specification dataset (like the
one created by \code{tabBookWeightSpec()}).
}
\examples{
\dontrun{
m <- newMultitable(~ gender + age4 + marstat, data = ds)
tabBook(m, ds, format = "xlsx", file = "wealthy-tab-book.xlsx", filter = "wealthy")
book <- tabBook(m, ds) # Returns a TabBookResult
tables <- prop.table(book, 2)
}
}
