% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-book.R
\name{tabFramePrepare}
\alias{tabFramePrepare}
\title{Prepare a tabFrame}
\usage{
tabFramePrepare(dataset, weight)
}
\arguments{
\item{dataset}{CrunchDataset, which may be subset with a filter expression
on the rows, and a selection of variables to use on the columns.}

\item{weight}{a CrunchVariable that has been designated as a potential
weight variable for \code{dataset}, or \code{NULL} for unweighted results. Alternatively
a named list where the name is the alias of the weight and the contents of
the list component are a character vector of aliases to which that weight
should apply.
Default is the currently applied \code{\link{weight}}.}
}
\description{
A tabFrame is a data.frame that acts as instructions to
tabBook for repacking a complexly weighted set of tabBook results
with appropriate ordering.
}
