% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-summary.R
\name{table}
\alias{table}
\title{Table function for Crunch objects}
\usage{
table(..., exclude, useNA = c("no", "ifany", "always"), dnn,
  deparse.level)
}
\arguments{
\item{...}{CrunchVariables to tabulate}

\item{exclude}{see \code{\link[base:table]{base::table}}}

\item{useNA}{see \code{\link[base:table]{base::table}}}

\item{dnn}{see \code{\link[base:table]{base::table}}}

\item{deparse.level}{see \code{\link[base:table]{base::table}}}
}
\value{
a table object
}
\description{
Table function for Crunch objects
}
\seealso{
\code{\link[base:table]{base::table}}
}
