% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\docType{methods}
\name{team}
\alias{team}
\alias{team<-}
\alias{team,CrunchFilter-method}
\alias{team<-,CrunchFilter,CrunchTeam-method}
\alias{team<-,CrunchFilter,ANY-method}
\alias{team,Multitable-method}
\alias{team<-,Multitable,CrunchTeam-method}
\alias{team<-,Multitable,ANY-method}
\title{Share Crunch assets with a team}
\usage{
team(x)

team(x) <- value

\S4method{team}{CrunchFilter}(x)

\S4method{team}{CrunchFilter,CrunchTeam}(x) <- value

\S4method{team}{CrunchFilter,ANY}(x) <- value

\S4method{team}{Multitable}(x)

\S4method{team}{Multitable,CrunchTeam}(x) <- value

\S4method{team}{Multitable,ANY}(x) <- value
}
\arguments{
\item{x}{a \code{CrunchFilter} or \code{Multitable}}

\item{value}{a \code{CrunchTeam} or url for a Crunch team}
}
\value{
a \code{CrunchTeam} that the asset is shared with.
}
\description{
You can share filters and multitables with a team that you are on. This will
give all team members access to view and edit these filters. Use \code{getTeams()}
to see what teams you are on.
}
