% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise.R
\name{tieredVar}
\alias{tieredVar}
\title{Create a categorical variable using tiers of a Categorical Array}
\usage{
tieredVar(x, tiers, ..., tier_type = NULL)
}
\arguments{
\item{x}{A Categorical Array variable}

\item{tiers}{A vector of ids, category names, or values}

\item{...}{Metadata like name or description that is passed to \code{\link[=VarDef]{VarDef()}}}

\item{tier_type}{Specify how \code{tiers} are identifying the categories, if left \code{NULL},
the default, assumes that a numeric vector is ids and a character vector is names.}
}
\value{
A Variable Defintiion
}
\description{
Collapses a categorical array to the first value of tiers that is found.
}
