% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/R-to-variable.R, R/expressions.R
\docType{methods}
\name{toVariable}
\alias{toVariable}
\alias{toVariable,CrunchExpr-method}
\alias{toVariable,Date-method}
\alias{toVariable,VariableDefinition-method}
\alias{toVariable,character-method}
\alias{toVariable,factor-method}
\alias{toVariable,logical-method}
\alias{toVariable,numeric-method}
\title{Generic method for converting objects to Crunch representations}
\usage{
toVariable(x, ...)

\S4method{toVariable}{character}(x, ...)

\S4method{toVariable}{numeric}(x, ...)

\S4method{toVariable}{factor}(x, ...)

\S4method{toVariable}{Date}(x, ...)

\S4method{toVariable}{VariableDefinition}(x, ...)

\S4method{toVariable}{logical}(x, ...)

\S4method{toVariable}{CrunchExpr}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional arguments}
}
\value{
a list object suitable for POSTing to the Crunch API. See the API
documentation for specifications.
}
\description{
If you have other object types you wish to convert to Crunch variables,
you can declare methods for \code{toVariable}
}

