% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/variable-transforms.R
\docType{class}
\name{Transforms-class}
\alias{Transforms-class}
\alias{transforms}
\alias{transforms<-}
\alias{Transforms}
\alias{transforms,CrunchVariable-method}
\alias{transforms,VariableTuple-method}
\alias{transforms,CrunchCube-method}
\alias{transforms<-,CrunchVariable-method}
\title{Transformations of variable and cube views}
\usage{
Transforms(...)

\S4method{transforms}{CrunchVariable}(x)

\S4method{transforms}{VariableTuple}(x)

\S4method{transforms}{CrunchCube}(x)

\S4method{transforms}{CrunchVariable}(x) <- value
}
\arguments{
\item{...}{For the constructor function \code{Transforms} you can pass
in attributes via \code{...}}

\item{x}{For the attribute getters and setters, an object of class
Transforms}

\item{value}{For \code{[<-}, the replacement Transforms to insert}
}
\description{
Transformations allow you to change how a variable or cube is displayed without changing the underlying data.
}
