% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/cube-transforms.R,
%   R/variable-transforms.R
\docType{class}
\name{Transforms-class}
\alias{Transforms-class}
\alias{transforms}
\alias{transforms<-}
\alias{Transforms}
\alias{transforms,CrunchCube-method}
\alias{transforms<-,CrunchCube,NULL-method}
\alias{transforms,CrunchVariable-method}
\alias{transforms,VariableTuple-method}
\alias{transforms<-,CrunchVariable,ANY-method}
\alias{transforms<-,CrunchVariable,NULL-method}
\title{Transformations of variable and cube views}
\usage{
Transforms(..., data = NULL)

\S4method{transforms}{CrunchCube}(x)

\S4method{transforms}{CrunchCube,`NULL`}(x) <- value

\S4method{transforms}{CrunchVariable}(x)

\S4method{transforms}{VariableTuple}(x)

\S4method{transforms}{CrunchVariable,ANY}(x) <- value

\S4method{transforms}{CrunchVariable,`NULL`}(x) <- value
}
\arguments{
\item{...}{For the constructor function \code{Transforms} you can pass
in attributes via \code{...}}

\item{data}{For the constructor function \code{Transforms} you can either pass in
attributes via \code{...} or you can create the objects with a fully defined
\code{list} representation of the objects via the \code{data} argument. See the examples.}

\item{x}{For the attribute getters and setters, an object of class
Transforms}

\item{value}{For \code{[<-}, the replacement Transforms to insert}
}
\description{
Transformations allow you to change how a variable or cube is displayed
without changing the underlying data.
}
