% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R, R/shoji-entity.R, R/tuple.R
\docType{methods}
\name{entity,CrunchProject-method}
\alias{entity,CrunchProject-method}
\alias{$,ShojiEntity-method}
\alias{$<-,ShojiEntity-method}
\alias{[[,ShojiEntity,ANY-method}
\alias{[[<-,ShojiEntity,ANY,ANY,ANY-method}
\alias{tuple-methods}
\alias{entity}
\alias{refresh,ShojiTuple-method}
\alias{$,ShojiTuple-method}
\alias{$<-,ShojiTuple-method}
\alias{[[,ShojiTuple,ANY-method}
\alias{[[<-,ShojiTuple,ANY,ANY,ANY-method}
\alias{self,ShojiTuple-method}
\alias{entity,VariableTuple-method}
\alias{entity,CrunchVariable-method}
\alias{entity,DatasetTuple-method}
\alias{id,DatasetTuple-method}
\alias{delete,ShojiTuple-method}
\alias{delete,DatasetTuple-method}
\alias{name,ShojiTuple-method}
\alias{name<-,ShojiTuple-method}
\alias{type,ShojiTuple-method}
\title{Methods for ShojiTuples}
\usage{
\S4method{entity}{CrunchProject}(x)

\S4method{$}{ShojiEntity}(x, name)

\S4method{$}{ShojiEntity}(x, name) <- value

\S4method{[[}{ShojiEntity,ANY}(x, i)

\S4method{[[}{ShojiEntity,ANY,ANY,ANY}(x, i) <- value

\S4method{refresh}{ShojiTuple}(x)

\S4method{$}{ShojiTuple}(x, name)

\S4method{$}{ShojiTuple}(x, name) <- value

\S4method{[[}{ShojiTuple,ANY}(x, i)

\S4method{[[}{ShojiTuple,ANY,ANY,ANY}(x, i) <- value

\S4method{self}{ShojiTuple}(x)

\S4method{entity}{VariableTuple}(x)

\S4method{entity}{CrunchVariable}(x)

\S4method{entity}{DatasetTuple}(x)

\S4method{id}{DatasetTuple}(x)

\S4method{delete}{ShojiTuple}(x, ...)

\S4method{delete}{DatasetTuple}(x, ...)

\S4method{name}{ShojiTuple}(x)

\S4method{name}{ShojiTuple}(x) <- value

\S4method{type}{ShojiTuple}(x)
}
\arguments{
\item{x}{a Tuple}

\item{name}{a Tuple slot to get or set}

\item{value}{What to set in a given slot}

\item{i}{In [[, a Tuple slot to get}

\item{...}{additional arguments to [[, ignored}

\item{object}{Same as \code{x} but for the \code{alias} method, in order to
match the generic from another package.}
}
\description{
ShojiTuples are objects extracted from ShojiCatalogs. They are internally
used.
}
\keyword{internal}
