% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tuple.R
\docType{methods}
\name{tuple-methods}
\alias{$,IndexTuple-method}
\alias{$<-,IndexTuple-method}
\alias{[[,IndexTuple,ANY-method}
\alias{[[<-,IndexTuple,ANY,ANY,ANY-method}
\alias{delete,DatasetTuple-method}
\alias{delete,IndexTuple-method}
\alias{entity}
\alias{entity,DatasetTuple-method}
\alias{entity,VariableTuple-method}
\alias{name,IndexTuple-method}
\alias{tuple-methods}
\alias{type,IndexTuple-method}
\title{Methods for IndexTuples}
\usage{
\S4method{refresh}{IndexTuple}(x)

\S4method{$}{IndexTuple}(x, name)

\S4method{$}{IndexTuple}(x, name) <- value

\S4method{[[}{IndexTuple,ANY}(x, i)

\S4method{[[}{IndexTuple,ANY,ANY,ANY}(x, i) <- value

\S4method{entity}{VariableTuple}(x)

\S4method{entity}{DatasetTuple}(x)

\S4method{delete}{IndexTuple}(x)

\S4method{delete}{DatasetTuple}(x, confirm = interactive(), ...)

\S4method{name}{IndexTuple}(x)

\S4method{type}{IndexTuple}(x)
}
\arguments{
\item{x}{a Tuple}

\item{name}{a Tuple slot to get or set}

\item{value}{What to set in a given slot}

\item{i}{In [[, a Tuple slot to get}

\item{confirm}{For \code{delete}, whether confirmation is required. See
\code{\link{delete}}.}

\item{...}{additional arguments to [[, ignored}
}
\description{
IndexTuples are objects extracted from ShojiCatalogs. They are internally
used.
}

