% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-datasets.R
\name{updateDatasetList}
\alias{updateDatasetList}
\title{Refresh the local list of Crunch datasets}
\usage{
updateDatasetList()
}
\value{
Nothing. Called for its side effects of resetting the cache.
}
\description{
Refreshes the local environment list of Crunch datasets. This function is deprecated.
}
\keyword{internal}
