% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateDatasetList}
\alias{updateDatasetList}
\title{Refresh the local list of Crunch datasets}
\usage{
updateDatasetList()
}
\value{
Nothing. Called for its side effects of setting local environment variables.
}
\description{
Refresh the local list of Crunch datasets
}

