\name{updateDatasetList}
\alias{updateDatasetList}
\title{Refresh the local list of Crunch datasets}
\usage{
  updateDatasetList(x = getUserDatasetURLs())
}
\arguments{
  \item{x}{the dataset URLs. You shouldn't change the
  default.}
}
\value{
  Nothing. Called for its side effects of setting local
  environment variables.
}
\description{
  Refresh the local list of Crunch datasets
}

