% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{validateWeightExpression}
\alias{validateWeightExpression}
\title{Validate an expression passed to makeWeight}
\usage{
validateWeightExpression(expr)
}
\arguments{
\item{expr}{An expression}
}
\value{
NULL value, function is called for its side effects
}
\description{
Utility function to catch formula errors passed to \link{makeWeight}.
}
\examples{
\dontrun{
validateWeightExpression(ds$var ~ c(10, 30, 60))
}

}
\keyword{internal}
