% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{validateWeightExpression}
\alias{validateWeightExpression}
\title{Validate an expression passed to makeWeight}
\usage{
validateWeightExpression(expr)
}
\arguments{
\item{expr}{An expression which should be in the form \code{var ~ target_weights}}
}
\value{
NULL value
}
\description{
Convenience function to catch formula errors passed to \link{makeWeight}.
}
\examples{
\dontrun{
validateWeightExpression(ds$var ~ c(10, 30, 60))
}

}
