% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{var-categories}
\alias{var-categories}
\alias{categories}
\alias{categories<-}
\alias{categories,VariableTuple-method}
\alias{categories,CrunchVariable-method}
\alias{categories,CategoricalVariable-method}
\alias{categories,CategoricalArrayVariable-method}
\alias{categories,VariableEntity-method}
\alias{categories<-,CategoricalVariable,Categories-method}
\alias{categories<-,CategoricalArrayVariable,Categories-method}
\alias{categories<-,CategoricalVariable,numeric-method}
\alias{categories<-,CategoricalVariable,character-method}
\alias{categories<-,CategoricalVariable,ANY-method}
\alias{categories<-,CategoricalArrayVariable,numeric-method}
\alias{categories<-,CategoricalArrayVariable,character-method}
\alias{categories<-,CategoricalArrayVariable,ANY-method}
\alias{categories<-,CrunchVariable,ANY-method}
\title{Get and set Categories on Variables}
\usage{
categories(x)

categories(x) <- value

\S4method{categories}{VariableTuple}(x)

\S4method{categories}{CrunchVariable}(x)

\S4method{categories}{CategoricalVariable}(x)

\S4method{categories}{CategoricalArrayVariable}(x)

\S4method{categories}{VariableEntity}(x)

\S4method{categories}{CategoricalVariable,Categories}(x) <- value

\S4method{categories}{CategoricalArrayVariable,Categories}(x) <- value

\S4method{categories}{CategoricalVariable,numeric}(x) <- value

\S4method{categories}{CategoricalVariable,character}(x) <- value

\S4method{categories}{CategoricalVariable,ANY}(x) <- value

\S4method{categories}{CategoricalArrayVariable,numeric}(x) <- value

\S4method{categories}{CategoricalArrayVariable,character}(x) <- value

\S4method{categories}{CategoricalArrayVariable,ANY}(x) <- value

\S4method{categories}{CrunchVariable,ANY}(x) <- value
}
\arguments{
\item{x}{a Variable}

\item{value}{for the setters, an object of class Categories to set.}
}
\value{
Getters return Categories; setters return \code{x} duly modified.
}
\description{
Get and set Categories on Variables
}
