% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variable.R
\docType{methods}
\name{variable-extract}
\alias{[,CrunchVariable,logical-method}
\alias{[,CrunchVariable,numeric-method}
\alias{variable-extract}
\title{"Subset" a Variable}
\usage{
\S4method{[}{CrunchVariable,CrunchExpr}(x, i, j, ..., drop = TRUE)

\S4method{[}{CrunchVariable,numeric}(x, i, j, ..., drop = TRUE)

\S4method{[}{CrunchVariable,logical}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a Variable}

\item{i}{a CrunchExpr, logical, or numeric}

\item{j}{Invalid}

\item{...}{additional arguments, ignored}

\item{drop}{Invalid}
}
\value{
a CrunchExpr containing references to the variable \code{x} and the
filter logic contained in \code{i}
}
\description{
These methods subset variables by creating Expressions, which can be
composed and evaluated as needed.
}

