% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R, R/variable-to-R.R
\docType{methods}
\name{as.vector,CrunchExpr-method}
\alias{as.vector,CrunchExpr-method}
\alias{variable-to-R}
\alias{as.vector,CrunchVariable-method}
\title{Convert Variables to local R objects}
\usage{
\S4method{as.vector}{CrunchExpr}(x, mode = "any")

\S4method{as.vector}{CrunchVariable}(x, mode = "any")
}
\arguments{
\item{x}{a CrunchVariable subclass}

\item{mode}{for Categorical variables, one of either "factor" (default,
which returns the values as factor); "numeric" (which returns the numeric
values); or "id" (which returns the category ids). If "id", values
corresponding to missing categories will return as the underlying integer
codes; i.e., the R representation will not have any \code{NA}s. Otherwise,
missing categories will all be returned \code{NA}. For non-Categorical
variables, the \code{mode} argument is ignored.}
}
\value{
an R vector of the type corresponding to the Variable. E.g.
CategoricalVariable yields type factor by default, NumericVariable yields
numeric, etc.
}
\description{
\code{as.vector} transfers data from Crunch to a local R session. Note:
\code{as.vector} returns the vector in the row order of the dataset. If filters
are set that specify an order that is different from the row order of the
dataset, the results will ignore that order. If you need the vector ordered
in that way, use syntax like \code{as.vector(ds$var)[c(10, 5, 2)]} instead.
}
