% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R, R/variable-to-R.R
\docType{methods}
\name{as.vector,CrunchExpr-method}
\alias{as.vector,CrunchExpr-method}
\alias{as.vector,CrunchVariable-method}
\alias{variable-to-R}
\title{Convert Variables to local R objects}
\usage{
\S4method{as.vector}{CrunchExpr}(x, mode)

\S4method{as.vector}{CrunchVariable}(x, mode)
}
\arguments{
\item{x}{a CrunchVariable subclass}

\item{mode}{for Categorical variables, one of either "factor" (default,
which returns the values as factor); "numeric" (which returns the numeric
values); or "id" (which returns the category ids). If "id", values
corresponding to missing categories will return as the underlying integer
codes; i.e., the R representation will not have any \code{NA}s. Otherwise,
missing categories will all be returned \code{NA}. For non-Categorical
variables, the \code{mode} argument is ignored.}
}
\value{
an R vector of the type corresponding to the Variable. E.g.
CategoricalVariable yields type factor by default, NumericVariable yields
numeric, etc.
}
\description{
Convert Variables to local R objects
}

