% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variable-update.R
\docType{methods}
\name{[<-,CrunchVariable,ANY,missing,ANY-method}
\alias{[<-,CategoricalArrayVariable,ANY,missing,character-method}
\alias{[<-,CategoricalArrayVariable,ANY,missing,factor-method}
\alias{[<-,CategoricalArrayVariable,ANY,missing,numeric-method}
\alias{[<-,CategoricalVariable,ANY,missing,character-method}
\alias{[<-,CategoricalVariable,ANY,missing,factor-method}
\alias{[<-,CategoricalVariable,ANY,missing,numeric-method}
\alias{[<-,CrunchVariable,ANY,missing,ANY-method}
\alias{[<-,CrunchVariable,ANY,missing,CrunchExpr-method}
\alias{[<-,CrunchVariable,CrunchExpr,missing,CrunchExpr-method}
\alias{[<-,DatetimeVariable,ANY,missing,Date-method}
\alias{[<-,DatetimeVariable,ANY,missing,POSIXt-method}
\alias{[<-,NumericVariable,ANY,missing,numeric-method}
\alias{[<-,TextVariable,ANY,missing,character-method}
\title{Updating variables with expressions or values}
\usage{
\S4method{[}{CrunchVariable,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{TextVariable,ANY,missing,character}(x, i, j) <- value

\S4method{[}{NumericVariable,ANY,missing,numeric}(x, i, j) <- value

\S4method{[}{DatetimeVariable,ANY,missing,Date}(x, i, j) <- value

\S4method{[}{DatetimeVariable,ANY,missing,POSIXt}(x, i, j) <- value

\S4method{[}{CrunchVariable,ANY,missing,CrunchExpr}(x, i, j) <- value

\S4method{[}{CrunchVariable,CrunchExpr,missing,CrunchExpr}(x, i, j) <- value

\S4method{[}{CategoricalVariable,ANY,missing,numeric}(x, i, j) <- value

\S4method{[}{CategoricalVariable,ANY,missing,character}(x, i, j) <- value

\S4method{[}{CategoricalVariable,ANY,missing,factor}(x, i, j) <- value

\S4method{[}{CategoricalArrayVariable,ANY,missing,numeric}(x, i, j) <- value

\S4method{[}{CategoricalArrayVariable,ANY,missing,character}(x, i, j) <- value

\S4method{[}{CategoricalArrayVariable,ANY,missing,factor}(x, i, j) <- value
}
\arguments{
\item{x}{a Variable}

\item{i}{a CrunchLogicalExpr or R index, optionally}

\item{j}{Invalid}

\item{value}{an R vector or a CrunchExpr with which to update}
}
\value{
\code{x} duly modified
}
\description{
Updating variables with expressions or values
}

