% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame.R
\name{as.data.frame.VariableCatalog}
\alias{as.data.frame.VariableCatalog}
\title{as.data.frame method for VariableCatalog}
\usage{
\method{as.data.frame}{VariableCatalog}(x, row.names = NULL,
  optional = FALSE, keys = c("alias", "name", "type"), ...)
}
\arguments{
\item{x}{A \code{VariableCatalog}, as produced by \code{variables(ds)}.}

\item{row.names}{part of \code{as.data.frame} signature. Ignored.}

\item{optional}{part of \code{as.data.frame} signature. Ignored.}

\item{keys}{A character vector of the variable catalog attributes which you
would like included in the data.frame. To include all attributes or see
which ones are available, set keys to "all". By default, the function will
return three fields: \code{c("alias", "name", "type")}.}

\item{...}{Additional arguments passed to \code{data.frame}}
}
\value{
A \code{data.frame} including metadata about each variable stored in the
variable catalog. The fields in the data.frame match the \code{keys} argument
provided to the function, and each row represents a variable.
}
\description{
This method gives you a view of a \code{VariableCatalog} as a \code{data.frame} in
order to facilitate further exploration.
}
\details{
Modifying the \code{data.frame} produced by this function will not update the
dataset on the Crunch server. Other methods exist for updating the metadata
in the variable catalog. See \code{vingette("variables", package = "crunch")}.
}
\examples{
\dontrun{
ds <- loadDataset("iris")
vars <- variables(ds)
var_df <- as.data.frame(vars, keys = "all")
}

}
