\name{VariableCatalog}
\alias{VariableCatalog}
\title{Collection of Variables within a Dataset}
\usage{
  VariableCatalog(...)
}
\description{
  A VariableCatalog contains references to all variables in
  a dataset, plus some descriptive metadata about each.
  VariableCatalogs also contain a
  \code{\link{VariableOrder}} that governs how variables
  within it are organized.
}

