% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{VariableCatalog-class}
\alias{VariableCatalog}
\alias{VariableCatalog-class}
\title{Collection of Variables within a Dataset}
\description{
A VariableCatalog contains references to all variables in a dataset, plus
some descriptive metadata about each. VariableCatalogs also contain a
\code{\link{VariableOrder}} that governs how variables within it are
organized.
}

