% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-metadata.R
\name{variableMetadata}
\alias{variableMetadata}
\title{Get all variable metadata for a dataset}
\usage{
variableMetadata(dataset)
}
\arguments{
\item{dataset}{CrunchDataset}
}
\value{
A VariableCatalog with all variable properties, including categories
and subvariables.
}
\description{
Crunch stores variable information in several catalogs containing information
abut the variable class, its missingness and subvariables. This function allows you
to access that information.
}
