% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-metadata.R
\name{variableMetadata}
\alias{variableMetadata}
\title{Get all variable metadata for a dataset}
\usage{
variableMetadata(dataset, parent = FALSE)
}
\arguments{
\item{dataset}{CrunchDataset}

\item{parent}{logical: Embed in subvariables a reference to their array
parent? Default is \code{FALSE}.}
}
\value{
A VariableCatalog that has things like categories embedded in each
categorical variable, and all subvariables are represented
}
\description{
Get all variable metadata for a dataset
}

