% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variable-order-slots.R
\docType{methods}
\name{VariableOrder-slots}
\alias{VariableOrder-slots}
\alias{entities}
\alias{entities,VariableOrder-method}
\alias{entities,list-method}
\alias{entities<-}
\alias{entities<-,VariableGroup-method}
\alias{entities<-,VariableOrder-method}
\alias{name,VariableGroup-method}
\alias{name<-,VariableGroup-method}
\alias{names,VariableOrder-method}
\alias{names<-,VariableOrder-method}
\title{Manipulate VariableGroup and VariableOrder}
\usage{
\S4method{entities}{VariableGroup}(x, simplify = FALSE)

\S4method{entities}{VariableOrder}(x, simplify = FALSE)

\S4method{entities}{list}(x, simplify = FALSE)

\S4method{entities}{VariableGroup}(x) <- value

\S4method{entities}{VariableOrder}(x) <- value

\S4method{name}{VariableGroup}(x)

\S4method{name}{VariableGroup}(x) <- value

\S4method{names}{VariableOrder}(x)

\S4method{names}{VariableOrder}(x) <- value
}
\arguments{
\item{x}{a VariableGroup or VariableOrder}

\item{simplify}{logical: should variable URLs inside of groups be flattened
or preserved in their nested lists? Default is \code{FALSE}.}

\item{value}{For name, a character (length-1 vector); for names, a character
vector of equal length to the number of VariableGroups beind modified; for
entities, either a character vector of variable URLs or a list containing a
combination of variable URLs and VariableGroups. Note that group names must
be unique, should be greater than 0 characters long, and "ungrouped" is a
reserved group name.}
}
\value{
Variable references, VariableGroups, or group names, as appropriate.
}
\description{
Manipulate VariableGroup and VariableOrder
}
\seealso{
VariableOrder

grouped
}

