% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/variable.R
\docType{methods}
\name{webApp}
\alias{webApp}
\alias{webApp,CrunchDataset-method}
\alias{webApp,CrunchVariable-method}
\title{View a Crunch Object in the Web Application}
\usage{
\S4method{webApp}{CrunchDataset}(x)

\S4method{webApp}{CrunchVariable}(x)
}
\arguments{
\item{x}{a Crunch Dataset or Variable}
}
\value{
Nothing; called for side effect of opening your web browser.
}
\description{
Convenience function that will use your system's "open" command to open
a Crunch object in our web application in your default browser.
}
