% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{webApp}
\alias{webApp}
\title{View a Crunch Object in the Web Application}
\usage{
webApp(x)
}
\arguments{
\item{x}{a Crunch Dataset or Variable}
}
\value{
Nothing; called for side effect of opening your web browser.
}
\description{
Convenience function that will use your system's "open" command to open
a Crunch object in our web application in your default browser.
}
