% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\docType{methods}
\name{weight}
\alias{weight}
\alias{weightVariables}
\alias{weight<-}
\alias{weightVariables,CrunchDataset-method}
\alias{weightVariables,VariableCatalog-method}
\title{Dataset weights}
\usage{
weight(x)

weight(x) <- value

\S4method{weightVariables}{CrunchDataset}(x)

\S4method{weightVariables}{VariableCatalog}(x)
}
\arguments{
\item{x}{a Dataset}

\item{value}{a Variable to set as weight, or NULL to remove the existing
weight}
}
\value{
For the \code{weight} getter, a Variable if there is a weight, otherwise the function
returns \code{NULL}. For the setter, the dataset's weights are modified accordingly.
\code{weightVariables} returns the aliases (or names, according to \code{options(crunch.namekey.dataset)}),
of the variables designated as weights.
}
\description{
\code{weight} lets you view and set your user's currently applied weight on the
server. \code{weightVariables} lets you view all of the variables that have been
designated as valid to use as weights.
}
