% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/decks.R, R/slides.R,
%   R/weight.R
\name{is.weight<-}
\alias{is.weight<-}
\alias{weight}
\alias{weight<-}
\alias{weight<-,CrunchDeck,ANY-method}
\alias{weight,CrunchSlide-method}
\alias{weight<-,CrunchSlide,ANY-method}
\alias{weight,Analysis-method}
\alias{weight<-,Analysis,CrunchVariable-method}
\alias{weight<-,Analysis,NULL-method}
\alias{weight,CrunchDataset-method}
\alias{weight<-,CrunchDataset,ANY-method}
\alias{is.weight}
\alias{is.weight<-,NumericVariable-method}
\title{Dataset weights}
\usage{
is.weight(x) <- value

weight(x)

weight(x) <- value

\S4method{weight}{CrunchDeck,ANY}(x) <- value

\S4method{weight}{CrunchSlide}(x)

\S4method{weight}{CrunchSlide,ANY}(x) <- value

\S4method{weight}{Analysis}(x)

\S4method{weight}{Analysis,CrunchVariable}(x) <- value

\S4method{weight}{Analysis,`NULL`}(x) <- value

\S4method{weight}{CrunchDataset}(x)

\S4method{weight}{CrunchDataset,ANY}(x) <- value

is.weight(x)

\S4method{is.weight}{NumericVariable}(x) <- value
}
\arguments{
\item{x}{a Dataset}

\item{value}{a Variable, VariableDefinition, or \code{NULL}. If a
VariableDefinition is passed, the variable will first be created and then set
as the datasets weight. Set to \code{NULL} to remove existing weights from the
dataset.}
}
\value{
For the \code{weight} getter, a Variable if there is a weight, else
NULL. For the setter, x, modified accordingly. \code{weightVariables} returns
the aliases (or names, according to \code{options(crunch.namekey.dataset)}),
of the variables designated as weights.
}
\description{
\code{weight} lets you view and set your user's currently applied weight on the
server. \code{weightVariables} lets you view all of the variables that have been
designated as valid to use as weights.
}
\seealso{
\code{\link[=weightVariables]{weightVariables()}} \code{\link[=makeWeight]{makeWeight()}}
}
