% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\docType{methods}
\name{weight}
\alias{weight}
\alias{weightVariables}
\alias{weight<-}
\alias{weightVariables,CrunchDataset-method}
\alias{weightVariables,VariableCatalog-method}
\title{Dataset weights}
\usage{
weight(x)

weight(x) <- value

\S4method{weightVariables}{CrunchDataset}(x)

\S4method{weightVariables}{VariableCatalog}(x)
}
\arguments{
\item{x}{a Dataset}

\item{value}{a Variable, VariableDefinition, or \code{NULL}. If a
VariableDefinition is passed, the variable will first be created and then set
as the datasets weight. Set to \code{NULL} to remove existing weights from the
dataset.}
}
\value{
For the \code{weight} getter, a Variable if there is a weight, else
NULL. For the setter, x, modified accordingly. \code{weightVariables} returns
the aliases (or names, according to \code{options(crunch.namekey.dataset)}),
of the variables designated as weights.
}
\description{
\code{weight} lets you view and set your user's currently applied weight on the
server. \code{weightVariables} lets you view all of the variables that have been
designated as valid to use as weights.
}
