% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\docType{methods}
\name{weightVariables}
\alias{weightVariables}
\alias{weightVariables,CrunchDataset-method}
\alias{weightVariables,VariableCatalog-method}
\alias{weightVariables<-,CrunchDataset-method}
\alias{weightVariables<-,VariableCatalog-method}
\alias{is.weightVariable}
\alias{is.weightVariable<-,NumericVariable-method}
\title{Change which variables can be set as weights}
\usage{
\S4method{weightVariables}{CrunchDataset}(x)

\S4method{weightVariables}{VariableCatalog}(x)

\S4method{weightVariables}{CrunchDataset}(x) <- value

\S4method{weightVariables}{VariableCatalog}(x) <- value

is.weightVariable(x)

\S4method{is.weightVariable}{NumericVariable}(x) <- value
}
\arguments{
\item{x}{a dataset or variable catalog}

\item{value}{For the setter variables to set as the \code{weightVariables} this can
be a numeric Crunch variable, list of numeric Crunch variables or
a character vector with the aliases of numeric Crunch variables.}
}
\value{
For the getter the variables which can be set as the dataset's weight. For
the setter the dataset or variable catalog duly modified.
}
\description{
Editors can change which variables can be set as the weighting variable
for a dataset. For instance if several weights have been calculated they
can let the user choose which of those variables to use a weight, but prevent
the user from choosing other variables as weight. This function allows you
to get and set the \code{weightVariables} of a dataset, which are the variables which
can be set as the dataset \code{\link[=weight]{weight()}}.
}
\seealso{
\code{\link[=weight]{weight()}} \code{\link[=makeWeight]{makeWeight()}}
}
