% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/weight.R
\name{weightVariables}
\alias{weightVariables}
\alias{weightVariables,CrunchDataset-method}
\alias{weightVariables,VariableCatalog-method}
\title{Get a dataset's weightVariables}
\usage{
weightVariables(x)

\S4method{weightVariables}{CrunchDataset}(x)

\S4method{weightVariables}{VariableCatalog}(x)
}
\arguments{
\item{x}{a CrunchDataset}
}
\value{
\code{weightVariables} returns a character vector of the aliases of the
variables that are eligible to be used as weights.
}
\description{
Get a dataset's weightVariables
}
\seealso{
\code{\link[=weight]{weight()}} \code{\link[=makeWeight]{makeWeight()}} \code{\link[=modifyWeightVariables]{modifyWeightVariables()}}
}
